### 2025-03-07 #7885 @editorialbot
[PRE REVIEW]: Dyad: a binary-star dynamics and statistics package for Python
**Submitting author:** @AmeryGration (<a href="http://orcid.org/0000-0003-1379-6696">Amery Gration</a>)
**Repository:** https://github.com/AmeryGration/dyad
**Branch with paper.md** (empty if default branch): joss_paper
**Version:** 0.0.0
