### 2025-03-07 #7885 @warrickball
Hi @AmeryGration, thanks for your submission! All the suitable JOSS editors are currently working at capacity so I'm going to "waitlist" this review until an editor with the relevant expertise is available to take it on. Thanks for your patience!

For other editors, the pipelines look like they're being thrown off by the `joss_paper` branch only containing the JOSS manuscript and not any of the main code. I haven't seen this before so will follow up but it isn't fundamentally a problem. I ran `cloc` manually and the results were:
```
Language                     files          blank        comment           code
HTML                            99          17100            694          46421
Python                          32           2898           5850           7359
JavaScript                      13            181            271           1126
CSS                              7            204             70            937
reStructuredText                65            334            397            291
DOS Batch                        1              8              1             26
Markdown                         2              7              0             22
Text                             3              0              0             12
SVG                              2              0              0              9
make                             1              4              7              9
SUM:                           225          20736           7290          56212
```
but I have comments for the author below.

In the meantime @AmeryGration, there are a few details we can iron out.

1. The paper looks like it's failing to build on account of a tab in YAML header. If you fix that you can try running the pipeline again here by posting a comment starting with `@editorialbot generate pdf`.
2. Don't worry about the automatic check failing to find the license. I can see the GPL v3 licence in the `COPYING` file.
3. It looks like you've bundled the built HTML docs in the repo, which I'd disadvise mostly to avoid repo churn and the potential to not update them. You may want to rather host the built docs on e.g. GitHub pages, but either way I'd suggest remove everything under `docs/build` from the repo and let users know they can rebuild them with Sphinx.
4. It looks like some temporary files have also made their way into the repo and should be removed. I found at least
    ```
    ./dyad/constants/#__init__.py#
    ./dyad/stats/data/log_period/#sample_pdf.py#
    ./dyad/stats/data/period/#sample_pdf.py#
    ```
6. The largest Python source file is [`dyad/stats/_distn_infrastructure.py`](https://github.com/AmeryGration/dyad/blob/df45404bb152f7016d34bf1865741048564f7d1f/dyad/stats/_distn_infrastructure.py), which looks like it overlaps a lot with [the same file in the `scipy.stats` module](https://github.com/scipy/scipy/blob/5cf433589639f28cbf5c3911061b07f0c6d6bd68/scipy/stats/_distribution_infrastructure.py). Can you explain a bit more what's going on here? I haven't looked in too much detail but if it's the infrastructure for constructing your own statistical distributions with the SciPy API, why not just use the SciPy implementation? I note the commit message says
    > Local versions of scipy.stats._distn_infrastructure.py and scipy.stats.distr_params created to allow automatic generation of docstrings

    but duplicating so much source doesn't feel like the best solution. (I'm happy to help work out a better one. I've subclassed SciPy stats classes before but I don't remember offhand what we did—if anything—about API docs, if that's the issue.)

    (For editors wondering about scopy, `cloc` reports 3625 lines of code in `_distribution_infrastructure.py`, so there still appears to be ~3800 other lines of code in the repo.)
