+++ dyad/stats/data/log_period/plot.py
@@ -13,7 +13,8 @@ f = np.loadtxt("./frequency_sample.dat")
 F = np.loadtxt("./cumulative_frequency_sample.dat")
 
 fig, ax = plt.subplots()
-ax.pcolormesh(xx, mm, f, cmap="Grays", rasterized=True)
+# ax.pcolormesh(xx, mm, f, rasterized=True)
+ax.pcolormesh(xx, mm, f, rasterized=True)
 ax.set_xlabel(r"$x$")
 ax.set_ylabel(r"$m_{1}$")
 plt.savefig("./Figures/pdf_log10_period.jpg")
@@ -21,7 +22,7 @@ plt.savefig("./Figures/pdf_log10_period.pdf")
 plt.show()
 
 fig, ax = plt.subplots()
-ax.pcolormesh(xx, mm, F, cmap="Grays", rasterized=True)
+ax.pcolormesh(xx, mm, F, rasterized=True)
 ax.set_xlabel(r"$x$")
 ax.set_ylabel(r"$m_{1}$")
 plt.savefig("./Figures/cdf_log10_period.jpg")
