+++ docs/build/html/_sources/community_guidelines/index.rst.txt
@@ -1,30 +1,0 @@
-.. _community_guidelines:
-
-******************
-Contribute to Dyad
-******************
-
-I very much welcome contributions to Dyad. If you would like to
-contribute feedback then please `email me`_.  If you would like to
-contribute code then please fork the repository and make a pull
-request. For guidance on this, see Github's `advice`_.
-
-*************************
-Report problems with Dyad
-*************************
-
-You can report problems with Dyad using the repository's `Issues page`_.
-
-************
-Seek support
-************
-
-You can seek support by `emailing me`_. I will be very happy to help.
-
-.. _email me: https://github.com/AmeryGration
-
-.. _advice: https://docs.github.com/en/get-started/exploring-projects-on-github/contributing-to-a-project
-
-.. _Issues page: https://github.com/AmeryGration/dyad/issues
-
-.. _emailing me: https://github.com/AmeryGration
