+++ docs/build/html/_sources/generated/dyad.constants.rst.txt
@@ -1,7 +1,0 @@
-dyad.constants package
-======================
-
-.. automodule:: dyad.constants
-   :members:
-   :undoc-members:
-   :show-inheritance:
