+++ docs/build/html/_sources/generated/generated/dyad.period_from_semimajor_axis.rst.txt
@@ -1,6 +1,0 @@
-﻿dyad.period\_from\_semimajor\_axis
-==================================
-
-.. currentmodule:: dyad
-
-.. autofunction:: period_from_semimajor_axis
