+++ docs/build/html/_sources/generated/generated/dyad.stats.eccentricity.moe2017.rst.txt
@@ -1,6 +1,0 @@
-﻿dyad.stats.eccentricity.moe2017
-===============================
-
-.. currentmodule:: dyad.stats.eccentricity
-
-.. autodata:: moe2017
