+++ docs/build/html/_sources/generated/generated/generated/dyad.stats.semimajor_axis.opik1924.rst.txt
@@ -1,6 +1,0 @@
-﻿dyad.stats.semimajor\_axis.opik1924
-===================================
-
-.. currentmodule:: dyad.stats.semimajor_axis
-
-.. autodata:: opik1924
