+++ docs/build/html/_sources/generated/generated/dyad.stats.mass.kroupa2002.rst.txt
@@ -1,6 +1,0 @@
-﻿dyad.stats.mass.kroupa2002
-==========================
-
-.. currentmodule:: dyad.stats.mass
-
-.. autodata:: kroupa2002
