+++ docs/build/html/_sources/api_reference/generated/generated/dyad.stats.log_period.moe2017.rst.txt
@@ -1,6 +1,0 @@
-dyad.stats.log\_period.moe2017
-==============================
-
-.. currentmodule:: dyad.stats.log_period
-
-.. autodata:: moe2017
