+++ paper.md
@@ -12,7 +12,7 @@ authors:
     affiliation: 1
 affiliations:
   - name: School of Mathematics and Phyics, University of Surrey, Guildford, GU2 7XH, United Kingdom
-	index: 1
+    index: 1
     ror: 00ks66431
 date: 3 March 2025
 bibliography: paper.bib
