+++ docs/build/html/_sources/tutorials/tutorial_2.rst.txt
@@ -1,0 +1,178 @@
+.. _tutorial_2:
+
+*****************************************
+Synthesizing a population of binary stars
+*****************************************
+
+Dyad has a subpackage, :mod:`dyad.stats`, that contains probability
+distributions for the masses, mass ratios, and orbital elements of a
+population of binary stars. These probability distribitions are
+implemented in the same way as Scipy's continous random variables (see
+:class:`scipy.stats.rv_continuous` and, for an example,
+:class:`scipy.stats.loguniform`). As a result, they come equipped with
+a large number of useful methods, in particular ``pdf``, which
+computes the probability density function (PDF), and ``rvs``, which
+generates random variates (i.e. which generates a sample).
+
+For example, consider the random variable for inclination, :math:`I`, where :math:`\sin(I) \sim U(0, \pi)`. This is implemented by Dyad using the class :class:`dyad.stats.inclination`. Instantiate this class as follows.
+
+.. sourcecode:: python
+
+   >>> import dyad.stats as stats
+   >>> rv = stats.inclination
+
+Now compute the PDF on the interval :math:`[0, \pi]`.
+
+.. sourcecode:: python
+
+   >>> import numpy as np
+   >>> x = np.linspace(0., np.pi)
+   >>> pdf = rv.pdf(x)
+
+And synthesize :math:`10~000` realizations.
+
+.. sourcecode:: python
+
+   >>> sample = rv.rvs(size=10_000)
+
+Now plot our results.
+
+.. sourcecode:: python
+
+   >>> import matplot.pyplot as plt
+   >>> fig, ax = plt.subplots()
+   >>> ax.hist(sample, bins=25, density=True, histtype="step")
+   >>> ax.set_xticks([0., 0.5*np.pi, np.pi], [r"$0$", r"$\pi/2$", r"$\pi$"])
+   >>> ax.set_xlabel(r"$i$")
+   >>> ax.set_ylabel(r"$f_{I}$")
+   >>> ax.plot(x, pdf)
+   >>> plt.show()
+
+.. _inclination:
+.. figure:: ../figures/pdf_inclination.jpg
+   :figwidth: 75%
+   :align: center
+
+   The probability density function for inclination.
+
+Some of Dyad's random variables are conditional on other random variables. We must use their shape parameters to fully specify them. For example, the true anomaly of a body moving on an elliptical orbit in a gravitational central potentia depends on that orbit's eccentricity, and so is a conditional random variable, :math:`\Theta|E = e`. Again, we may synthesize a sample and compute the PDF. Suppose that :math:`e = 0.5`.
+
+.. sourcecode:: python
+
+   >>> rv = stats.true_anomaly(e=0.5)
+   >>> x = np.linspace(0., 2.*np.pi)
+   >>> pdf = rv.pdf(x)
+   >>> sample = rv.rvs(size=10_000)
+
+Plot the results.
+
+.. sourcecode:: python
+
+   >>> fig, ax = plt.subplots()
+   >>> ax.hist(sample, bins=25, density=True, histtype="step")
+   >>> ax.plot(x, pdf)
+   >>> ax.set_xticks([0., np.pi, 2.*np.pi], [r"$0$", r"$\pi$", r"$2\pi$"])
+   >>> ax.set_xlabel(r"$\theta$")
+   >>> ax.set_ylabel(r"$f_{\Theta}$")
+   >>> plt.show()
+
+.. _inclincation:
+.. figure:: ../figures/pdf_true_anomaly.jpg
+   :figwidth: 75%
+   :align: center
+
[…]
