+++ docs/build/html/_sources/generated/dyad.constants.rst.txt
@@ -1,0 +1,7 @@
+dyad.constants package
+======================
+
+.. automodule:: dyad.constants
+   :members:
+   :undoc-members:
+   :show-inheritance:
