+++ docs/build/html/_sources/api_reference/generated/generated/dyad.stats.mass.kroupa2002.rst.txt
@@ -1,0 +1,6 @@
+dyad.stats.mass.kroupa2002
+==========================
+
+.. currentmodule:: dyad.stats.mass
+
+.. autodata:: kroupa2002
