+++ docs/build/html/_sources/generated/dyad.stats.rst.txt
@@ -1,0 +1,58 @@
+dyad.stats package
+==================
+
+.. automodule:: dyad.stats
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Submodules
+----------
+
+dyad.stats.eccentricity module
+------------------------------
+
+.. automodule:: dyad.stats.eccentricity
+   :members:
+   :undoc-members:
+   :show-inheritance:
+      
+dyad.stats.log\_period module
+-----------------------------
+
+.. automodule:: dyad.stats.log_period
+   :members:
+   :undoc-members:
+   :show-inheritance:
+      
+dyad.stats.mass module
+----------------------
+
+.. automodule:: dyad.stats.mass
+   :members:
+   :undoc-members:
+   :show-inheritance:
+      
+dyad.stats.mass\_ratio module
+-----------------------------
+
+.. automodule:: dyad.stats.mass_ratio
+   :members:
+   :undoc-members:
+   :show-inheritance:
+      
+dyad.stats.period module
+------------------------
+
+.. automodule:: dyad.stats.period
+   :members:
+   :undoc-members:
+   :show-inheritance:
+      
+dyad.stats.semimajor\_axis module
+---------------------------------
+
+.. automodule:: dyad.stats.semimajor_axis
+   :members:
+   :undoc-members:
+   :show-inheritance:
