+++ docs/build/html/_sources/generated/dyad.rst.txt
@@ -1,0 +1,16 @@
+dyad package
+============
+
+.. automodule:: dyad
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Subpackages
+-----------
+
+.. toctree::
+   :maxdepth: 1
+
+   dyad.constants
+   dyad.stats
