+++ docs/build/html/_sources/generated/generated/dyad.true_anomaly_from_eccentric_anomaly.rst.txt
@@ -1,0 +1,6 @@
+﻿dyad.true\_anomaly\_from\_eccentric\_anomaly
+============================================
+
+.. currentmodule:: dyad
+
+.. autofunction:: true_anomaly_from_eccentric_anomaly
