+++ docs/build/html/_sources/generated/generated/generated/dyad.stats.eccentricity.moe2017.rst.txt
@@ -1,0 +1,6 @@
+﻿dyad.stats.eccentricity.moe2017
+===============================
+
+.. currentmodule:: dyad.stats.eccentricity
+
+.. autodata:: moe2017

[…]
