+++ docs/build/html/_sources/api_reference/generated/dyad.stats.longitude_of_ascending_node.rst.txt
@@ -1,6 +1,0 @@
-﻿dyad.stats.longitude\_of\_ascending\_node
-=========================================
-
-.. currentmodule:: dyad.stats
-
-.. autodata:: longitude_of_ascending_node
