+++ docs/build/html/_sources/generated/generated/dyad.mean_anomaly_from_eccentric_anomaly.rst.txt
@@ -1,6 +1,0 @@
-﻿dyad.mean\_anomaly\_from\_eccentric\_anomaly
-============================================
-
-.. currentmodule:: dyad
-
-.. autofunction:: mean_anomaly_from_eccentric_anomaly
