+++ docs/build/html/_sources/generated/generated/generated/dyad.stats.eccentricity.thermal.rst.txt
@@ -1,6 +1,0 @@
-﻿dyad.stats.eccentricity.thermal
-===============================
-
-.. currentmodule:: dyad.stats.eccentricity
-
-.. autodata:: thermal
