+++ docs/build/html/_sources/generated/generated/generated/dyad.stats.mass_ratio.duquennoy1991.rst.txt
@@ -1,6 +1,0 @@
-﻿dyad.stats.mass\_ratio.duquennoy1991
-====================================
-
-.. currentmodule:: dyad.stats.mass_ratio
-
-.. autodata:: duquennoy1991
