+++ docs/build/html/_sources/generated/generated/dyad.stats.period.trunclognorm.rst.txt
@@ -1,6 +1,0 @@
-﻿dyad.stats.period.trunclognorm
-==============================
-
-.. currentmodule:: dyad.stats.period
-
-.. autodata:: trunclognorm
