+++ docs/build/html/_sources/index.rst.txt
@@ -1,64 +1,0 @@
-####
-Dyad
-####
-
-.. include:: ../../README.md
-
-Getting started
-===============
-
-Get started with Dyad using the following brief guides.
-
-.. toctree::
-   :maxdepth: 1
-
-   getting_started/installation
-   getting_started/what_is_dyad
-
-Tutorials
-=========
-
-Get to know Dyad with the following tutorials.
-
-.. toctree::
-   :maxdepth: 1
-
-   tutorials/tutorial_1
-   tutorials/tutorial_2
-
-API reference
-=============
-
-Get the details of Dyad's API using the reference documentation.
-
-.. toctree::
-   :maxdepth: 1
-
-   api_reference/index
-
-Community guidelines
-====================
-
-Learn how to contribute to, report problems with, and seek support for Dyad with the Community guidelines.
-
-.. toctree::
-   :maxdepth: 1
-
-   community_guidelines/index
-   
-License and attribution
-=======================
-
-Dyad was written and is maintained by Amery Gration. It is free
-software made available under the GNU General Public License. If you
-make use of it then please cite `my paper`_ in *Journal of open source
-software*.
-
-Indices and tables
-------------------
-
-* :ref:`genindex`
-* :ref:`modindex`
-* :ref:`search`
-
-.. _my paper: xxx
