+++ docs/build/html/_sources/generated/generated/dyad.Orbit.rst.txt
@@ -1,44 +1,0 @@
-﻿dyad.Orbit
-==========
-
-.. currentmodule:: dyad
-
-.. autoclass:: Orbit
-
-   
-   .. automethod:: __init__
-
-   
-   .. rubric:: Methods
-
-   .. autosummary::
-   
-      ~Orbit.__init__
-      ~Orbit.eccentric_anomaly
-      ~Orbit.mean_anomaly
-      ~Orbit.radius
-      ~Orbit.speed
-      ~Orbit.state
-   
-   
-
-   
-   
-   .. rubric:: Attributes
-
-   .. autosummary::
-   
-      ~Orbit.apoapsis
-      ~Orbit.area
-      ~Orbit.argument_of_pericentre
-      ~Orbit.eccentricity
-      ~Orbit.inclination
-      ~Orbit.longitude_of_ascending_node
-      ~Orbit.mass
-      ~Orbit.periapsis
-      ~Orbit.period
-      ~Orbit.semilatus_rectum
-      ~Orbit.semimajor_axis
-      ~Orbit.semiminor_axis
-   
-   
