### 2025-04-11 #7885 @AmeryGration
Thank you so much for examining the repo in so much detail and sorry to have been so long in replying. I'm happy to wait for a reviewer to become available of course. If I can be of any help in finding one, please let me know.

1. Tab removed and pipeline rerun. The resulting PDF is now showing.
2. Thank you.
3. Directory `docs/build/` removed from repo. I will move the documentation to Read the Docs this coming week.
4. Temporary files removed from repo. I've found some other superfluous files that I've also removed.
5. Yes, it's ugly, isn't it! As you have worked out, the issue was the documentation. In order to automatically generate the documentation for instances of random varables Scipy uses developer-specified data kept in `_distr_params.py`. I couldn't find a way to add the relevant data without recompiling Scipy itself. If you could suggest a better way of avoiding this I would be thrilled.
