### 2025-04-23 d21ad2d
Mass limits of moe2017 changed to 0.8 and 40
