+++ dyad/stats/data/log_period/sample_pdf.py
@@ -166,8 +166,8 @@ def pdf(x, primary_mass):
 #############################################################################
 # Create grid of sample points
 #############################################################################
-m_min = 0.08
-m_max = 60.
+m_min = 0.8
+m_max = 40.
 primary_mass_boundary = (m_min, 1.2, 3.5, 6., m_max)
 log10_period_boundary = (
     0.2, 1., 1.3, 2., 2.5, 3.4, 3.5, 4., 4.5, 5.5, 6., 6.5, 8.
