+++ dyad/stats/data/secondary_mass/plot.py
@@ -22,80 +22,80 @@ cdf_sample = np.loadtxt("cumulative_frequency_sample.dat").T
 ########################################################################
 # Plot PDF
 ########################################################################
-# fig, ax = plt.subplots()
-# ax.pcolormesh(
-#     primary_mass_sample, mass_ratio_sample,
-#     pdf_sample, cmap=cmap, rasterized=True
-# )
-# ax.contour(
-#     primary_mass_sample, mass_ratio_sample,
-#     np.log10(pdf_sample), colors="k", levels=10
-# )
-# ax.hlines(0.1, m_min, m_max, color="k", ls="dashed")
-# ax.hlines(0.3, m_min, m_max, color="k", ls="dashed")
-# ax.hlines(0.95, m_min, m_max, color="k", ls="dashed")
-# ax.vlines(primary_mass_boundary, 0., 1., ls="dashed")
-# ax.set_xlim(m_min, m_max)
-# ax.set_ylim(0., 1.)
-# ax.set_xscale("log")
-# ax.set_xlabel(r"$m_{1}$")
-# ax.set_ylabel(r"$q$")
-# fig.savefig("./Figures/f_QgivenM1_moe2017_pairing_2d.pdf", dpi=300)
-# fig.savefig("./Figures/f_QgivenM1_moe2017_pairing_2d.jpg", dpi=300)
-# fig.show()
+fig, ax = plt.subplots()
+ax.pcolormesh(
+    primary_mass_sample, mass_ratio_sample,
+    pdf_sample, cmap=cmap, rasterized=True
+)
+ax.contour(
+    primary_mass_sample, mass_ratio_sample,
+    np.log10(pdf_sample), colors="k", levels=10
+)
+ax.hlines(0.1, m_min, m_max, color="k", ls="dashed")
+ax.hlines(0.3, m_min, m_max, color="k", ls="dashed")
+ax.hlines(0.95, m_min, m_max, color="k", ls="dashed")
+ax.vlines(primary_mass_boundary, 0., 1., ls="dashed")
+ax.set_xlim(m_min, m_max)
+ax.set_ylim(0., 1.)
+ax.set_xscale("log")
+ax.set_xlabel(r"$m_{1}$")
+ax.set_ylabel(r"$q$")
+fig.savefig("./Figures/f_QgivenM1_moe2017_pairing_2d.pdf", dpi=300)
+fig.savefig("./Figures/f_QgivenM1_moe2017_pairing_2d.jpg", dpi=300)
+fig.show()
 
-# fig, ax = plt.subplots()
-# ax.pcolormesh(
-#     primary_mass_sample, mass_ratio_sample,
-#     np.log10(cdf_sample),
-#     cmap=cmap, rasterized=True
-# )
-# ax.contour(
-#     primary_mass_sample, mass_ratio_sample,
-#     np.log10(cdf_sample), colors="k",
-#     levels=10
-# )
-# ax.hlines(0.1, m_min, m_max, color="k", ls="dashed")
-# ax.hlines(0.3, m_min, m_max, color="k", ls="dashed")
-# ax.hlines(0.95, m_min, m_max, color="k", ls="dashed")
-# ax.vlines(primary_mass_boundary, 0., 1., ls="dashed")
-# ax.set_xlim(m_min, m_max)
-# ax.set_ylim(0., 1.)
-# ax.set_xscale("log")
-# ax.set_xlabel(r"$m_{1}$")
-# ax.set_ylabel(r"$q$")
-# fig.savefig("./Figures/F_QgivenM1_moe2017_pairing_2d.pdf", dpi=300)
[…]
