+++ .readthedocs.yaml
@@ -8,17 +8,15 @@ version: 2
 build:
   os: ubuntu-24.04
   tools:
-    python: "3.13"
+    python: "3.10.12"
 
 # Build documentation in the "docs/" directory with Sphinx
 sphinx:
-   configuration: ./docs/source/conf.py
+   configuration: docs/source/conf.py
 
-# Optionally, but recommended,
-# declare the Python requirements required to build your documentation
-# See https://docs.readthedocs.io/en/stable/guides/reproducible-builds.html
+# # Optionally, but recommended,
+# # declare the Python requirements required to build your documentation
+# # See https://docs.readthedocs.io/en/stable/guides/reproducible-builds.html
 # python:
 #    install:
-#    - requirements: docs/requirements.txt
-        
-
+#    - requirements: requirements.txt
