+++ .readthedocs.yaml
@@ -12,7 +12,7 @@ build:
 
 # Build documentation in the "docs/" directory with Sphinx
 sphinx:
-   configuration: docs/source/conf.py
+   configuration: ./docs/source/conf.py
 
 # # Optionally, but recommended,
 # # declare the Python requirements required to build your documentation
