+++ setup.py
@@ -49,8 +49,6 @@ setup(
     install_requires=[
         "numpy",
         "scipy",
-        "scipy.integrate",
-        "scipy.stats",
         "parameterized",
     ],
     include_package_data=True,
