+++ .readthedocs.yaml
@@ -14,9 +14,9 @@ build:
 sphinx:
    configuration: ./docs/source/conf.py
 
-# # Optionally, but recommended,
-# # declare the Python requirements required to build your documentation
-# # See https://docs.readthedocs.io/en/stable/guides/reproducible-builds.html
-# python:
-#    install:
-#    - requirements: requirements.txt
+# Optionally, but recommended,
+# declare the Python requirements required to build your documentation
+# See https://docs.readthedocs.io/en/stable/guides/reproducible-builds.html
+python:
+   install:
+   - method: pip
