+++ .readthedocs.yaml
@@ -20,3 +20,4 @@ sphinx:
 python:
    install:
    - method: pip
+     path: .
