+++ docs/source/conf.py
@@ -185,7 +185,8 @@ pygments_style = "sphinx"
 # # HTML output
 # ########################################################################
 
-html_theme = "pydata_sphinx_theme"
+# html_theme = "pydata_sphinx_theme"
+html_theme = "sphinx_rtd_theme"
 
 # # html_logo = "_static/logo.svg"
 # # html_favicon = "_static/favicon.ico"
