+++ docs/build/html/_sources/generated/generated/generated/dyad.stats.secondary_mass.random.rst.txt
@@ -1,0 +1,6 @@
+dyad.stats.secondary\_mass.random
+=================================
+
+.. currentmodule:: dyad.stats.secondary_mass
+
+.. autodata:: random
