+++ setup.py
@@ -54,7 +54,7 @@ setup(
         "intersphinx_registry",
         "sphinx_copybutton",
         "sphinx_design",
-        "pydata_sphinx_theme",
+        "sphinx-rtd-theme",
         "parameterized",
     ],
     include_package_data=True,
