### 2025-06-02 #7885 @AmeryGration
Hi @warrickball. Is there any news on a potential JOSS editor for the paper? Thanks!
