### 2025-06-03 e947f53
Logo added to documentation
