### 2025-06-03 982f299
Correction to formula for power-law index in stats.eccentricity.moe2017
