+++ docs/source/getting_started/installation.rst
@@ -4,15 +4,7 @@
 Installation
 ************
 
-You may install Dyad from the command line using PIP.
-
-.. sourcecode:: bash
-
-   $ pip install dyad
-
-It requires Python version 3.10 or greater.
-
-You may also install Dyad from the source distribution, available at
+You may install Dyad from the source distribution, available at
 the Dyad repository
 (`https://github.com/AmeryGration/dyad`__). Make a local copy of
 the repository and navigate to its top level (the directory
@@ -23,3 +15,5 @@ the repository and navigate to its top level (the directory
    $ python -m pip install .
    
 __ https://github.com/AmeryGration/dyad
+
+Dyad requires Python version 3.10 or greater.
