### 2025-07-07 7f09102
Develop the secondary-mass distribution
