+++ build/lib/dyad/_core.py
@@ -1,0 +1,1228 @@
+"""
+This module contains Dyad's core classes and functions. It is private
+but the objects it contains are available under the ``dyad``
+namespace.
+
+"""
+
+__all__ = [
+    "semimajor_axis_from_period",
+    "period_from_semimajor_axis",
+    "true_anomaly_from_mean_anomaly",
+    "true_anomaly_from_eccentric_anomaly",
+    "mean_anomaly_from_eccentric_anomaly",
+    "mean_anomaly_from_true_anomaly",
+    "eccentric_anomaly_from_true_anomaly",
+    "eccentric_anomaly_from_mean_anomaly",
+    "primary_semimajor_axis_from_semimajor_axis",
+    "secondary_semimajor_axis_from_semimajor_axis",
+    "primary_semimajor_axis_from_secondary_semimajor_axis",
+    "secondary_semimajor_axis_from_primary_semimajor_axis",
+    "Orbit",
+    "TwoBody",
+]
+
+import numpy as np
+import scipy as sp
+import dyad.constants as constants
+
+def _check_mass(m):
+    if not np.all(np.isreal(m)):
+        raise TypeError("m must be scalar.")
+    if np.any(m <= 0.):
+        raise ValueError("m must be positive.")
+    res = m
+
+    return res
+
+def _check_eccentricity(e):
+    # The the number 0.9999999999999999 < 1.
+    # But the number 0.99999999999999999 == 1.
+    if not np.all(np.isreal(e)):
+        raise TypeError("e must be scalar.")
+    if np.any((e < 0.) | (e >= 1.)):
+        raise ValueError("e must be nonnegative and less than one.")
+    res = e
+
+    return res
+
+def _check_semimajor_axis(a):
+    if not np.all(np.isreal(a)):
+        raise TypeError("a must be scalar.")
+    if np.any(a <= 0.):
+        raise ValueError("a must be positive.")
+    res = a
+
+    return res
+
+def _check_angle(x):
+    if not np.all(np.isreal(x)):
+        raise TypeError("Omega, i, and omega must be scalar.")
+    res = x
+
+    return res
+
+def _check_period(p):
+    if not np.all(np.isreal(p)):
+        raise TypeError("p must be scalar.")
+    if np.any(p <= 0.):
+        raise ValueError("p must be positive.")
+    res = p
+
+    return res
+
+def semimajor_axis_from_period(p, m_1, m_2):
+    """Return the semimajor axis given the period
+
+    Parameters
+    ----------
+
+    p : array-like
+
+        Period
+
+    m_1 : array-like
+
+        Mass of the more-massive body, :math:`m_{1}`.
+
+    m_2 : array-like
+
+        Mass of the less-massive body, :math:`m_{2}`.
+
+    Returns
+    -------
+
+    res : ndarray
+
+        Semimajor axis.
+
+    Examples
+    --------
+
+    Scalar parameters.
+
+    >>> semimajor_axis_from_period(365.25, 1., 3.00362e-6)
+    np.float64(0.9999884101100887)
+
+    Array-like parameters defining multiple orbits.
+
+    >>> p, m_1, m_2 = [365.25, 365.25], [1., 1.], [3.00362e-6, 3.00362e-6]
+    >>> semimajor_axis_from_period(p, m_1, m_2)
+    array([0.99998841, 0.99998841])
+
+    """
+    p = np.asarray(p)
+    m_1 = np.asarray(m_1)
+    m_2 = np.asarray(m_2)
+    p = _check_period(p)
+    m_1 = _check_mass(m_1)
+    m_2 = _check_mass(m_2)
[…]

[…]
