### 2025-07-15 932eb62
Conditional PDF for Q given M_1 restricted to valid Q
