+++ dyad/stats/secondary_mass.py
@@ -87,12 +87,12 @@ class moe2017_gen(_distn_infrastructure.rv_continuous):
     %(example)s
 
     """
-    def _shape_info(self):
-        return [_ShapeInfo("primary_mass", False, (0, np.inf), (False, False))]
+    # def _shape_info(self):
+    #     return [_ShapeInfo("primary_mass", False, (0, np.inf), (False, False))]
 
-    # def _argcheck(self, primary_mass):
-    #     return (0.8 <= primary_mass) & (primary_mass <= 60.)
-    #     # return (0. <= primary_mass) & (primary_mass < np.inf)
+    def _argcheck(self, primary_mass):
+        return (0.8 <= primary_mass) & (primary_mass <= 40.)
+        # return (0. <= primary_mass) & (primary_mass < np.inf)
     
     def _pdf(self, x, primary_mass):
         x = np.asarray(x)
@@ -170,7 +170,7 @@ _moe2017_values = np.tile(
 )
 _moe2017_ppf_interp = LinearNDInterpolator(_moe2017_points.T, _moe2017_values)
 
-# moe2017 = moe2017_gen(a=0.1, b=60., name="secondary_mass.moe2017")
+# moe2017 = moe2017_gen(a=0.8, b=40., name="secondary_mass.moe2017")
 moe2017 = moe2017_gen(name="secondary_mass.moe2017")
 
 
