+++ dyad/stats/data/secondary_mass/sample_pdf_test.py
@@ -54,31 +54,37 @@ from dyad.stats import period as period
 # Define the integrand
 ########################################################################
 def f(p, q, m):
-    """Return the joint probability density for given mass ratio and period"""
+    """Return joint probability density given mass ratio and period"""
     res = (
         mass_ratio.moe2017(np.log10(p), m).pdf(q)
         *period.moe2017(m).pdf(p)
     )
+    res = np.where((0.8 <= q*m) & (q*m <= 40.), res, 0.)
 
     return res
 
 ########################################################################
 # Create regular lattice of sample points
-
 ########################################################################
 primary_mass_boundary = (0.8, 1.2, 3.5, 6., 40.)
 mass_ratio_boundary = (0.1, 0.3, 0.95, 1.)
 log10_period_boundary = (
     0.2, 1., 1.3, 2., 2.5, 3.4, 3.5, 4., 4.5, 5.5, 6., 6.5, 8.
 )
-
 n = 50
+# primary_mass_sample = np.hstack([
+#     np.logspace(np.log10(0.8), np.log10(1.2), n),
+#     np.logspace(np.log10(1.2), np.log10(3.5), n)[1:],
+#     np.logspace(np.log10(3.5), np.log10(6.), n)[1:],
+#     np.logspace(np.log10(6.), np.log10(40.), n)[1:],
+# ])
 primary_mass_sample = np.hstack([
     np.linspace(0.8, 1.2, n),
     np.linspace(1.2, 3.5, n)[1:],
     np.linspace(3.5, 6., n)[1:],
     np.linspace(6., 40., n)[1:],
 ])
+n = 50
 mass_ratio_sample = np.hstack([
     np.linspace(0.1, 0.3, n),
     np.linspace(0.3, 0.95, n)[1:],
@@ -112,18 +118,20 @@ pdf_sample = trapezoid(f_sample, period_sample, axis=1).T
 cdf_sample = cumulative_trapezoid(
     pdf_sample, mass_ratio_sample, axis=1, initial=0.
 )
+
 # Check normalization
 I = trapezoid(pdf_sample, mass_ratio_sample)
 print(I)
 
-# ########################################################################
-# # Fine-tune the normalization of the PDF and CDF
-# ########################################################################
-# pdf_sample = pdf_sample/cdf_sample[:,-1:]
-# cdf_sample = cdf_sample/cdf_sample[:,-1:]
-# # Check normalization
-# I = trapezoid(pdf_sample, mass_ratio_sample)
-# print(I)
+########################################################################
+# Fine-tune the normalization of the PDF and CDF
+########################################################################
+pdf_sample = pdf_sample/cdf_sample[:,-1:]
+cdf_sample = cdf_sample/cdf_sample[:,-1:]
+
+# Check normalization
+I = trapezoid(pdf_sample, mass_ratio_sample)
+print(I)
 
 # ########################################################################
 # # Save data
@@ -134,147 +142,43 @@ print(I)
 # np.savetxt("./cumulative_frequency_sample.dat", cdf_sample)
 
[…]
