+++ dyad/stats/data/secondary_mass/secondary_pdf.py
@@ -1,0 +1,169 @@
+#!/usr/bin/env python
+
+import numpy as np
+import matplotlib as mpl
+import matplotlib.pyplot as plt
+
+from scipy.integrate import trapezoid
+from scipy.integrate import cumulative_trapezoid
+from scipy.interpolate import RegularGridInterpolator
+from dyad.stats import mass_ratio
+from dyad.stats import period as period
+
+mpl.style.use("sm")
+
+########################################################################
+# Interpolate the pairing function: PDF and CDF
+########################################################################
+pdf_interp = RegularGridInterpolator(
+    (mass_ratio_sample, primary_mass_sample),
+    pdf_sample.T,
+    bounds_error=False,
+    fill_value=0.
+)
+# cdf_interp = RegularGridInterpolator(
+#     (mass_ratio_sample, primary_mass_sample),
+#     cdf_sample.T,
+#     bounds_error=False,
+#     fill_value=0.
+# )
+
+########################################################################
+#
+########################################################################
+# m = np.linspace(0.8, 40., 2**9)
+m_1 = np.logspace(np.log10(0.8), np.log10(40.), 2**9)
+m_2 = np.logspace(np.log10(0.8), np.log10(40.), 2**9)
+m1m1, m2m2 = np.meshgrid(m_1, m_2)
+z = pdf_interp((m2m2/m1m1, m1m1))/m1m1
+Z = cumulative_trapezoid(z, m_2, axis=0, initial=0.)
+z /= Z[-1]
+Z /= Z[-1]
+
+########################################################################
+# Plot PDF
+########################################################################
+fig, ax = plt.subplots()
+ax.pcolormesh(m1m1, m2m2, np.log10(z), rasterized=True)#, cmap="Greys")
+ax.contour(m1m1, m2m2, np.log10(z), colors="k")
+ax.plot(m_1, 0.1*m_1, color="k", ls="solid")
+ax.plot(m_1, 0.3*m_1, color="k", ls="dashed")
+ax.plot(m_1, 0.95*m_1, color="k", ls="dashed")
+# ax.plot(m_1, m_2, color="k", ls="solid")
+ax.vlines(primary_mass_boundary, 0.8, 40., ls="dashed")
+ax.set_xlim(0.8, 40.)
+ax.set_ylim(0.8, 40.)
+ax.set_xscale("log")
+ax.set_yscale("log")
+ax.set_xlabel(r"$m_{1}$")
+ax.set_ylabel(r"$m_{2}$")
+ax.set_title(r"$f_{M_{2}|M_{1}}$")
+fig.savefig("f_M2givenM2_moe2017_pairing.pdf", dpi=300)
+fig.savefig("f_M2givenM2_moe2017_pairing.jpg", dpi=300)
+plt.show()
+
+########################################################################
+# Plot PDF
+########################################################################
+fig, ax_1 = plt.subplots()
+ax_2 = ax_1.twinx()
+ax_1.pcolormesh(m1m1, m2m2, Z, rasterized=True)#, cmap="Greys")
+ax_1.contour(m1m1, m2m2, Z, colors="k")
+ax_1.plot(m_1, 0.1*m_1, color="k", ls="solid")
+ax_1.plot(m_1, 0.3*m_1, color="k", ls="dashed")
+ax_1.plot(m_1, 0.95*m_1, color="k", ls="dashed")
+ax_1.plot(m_1, m_1, color="k", ls="solid")
+ax_1.vlines(primary_mass_boundary, 0.8, 40., ls="dashed")
[…]
