+++ dyad/stats/data/secondary_mass/sample_pdf_test.py
@@ -152,24 +152,30 @@ mpl.style.use("sm")
 
 fig, ax, cbar = plot.plot_with_colorbar()
 im = ax.pcolormesh(mass_ratio_sample, primary_mass_sample,
-                   np.log10(pdf_sample), rasterized=True)
+                   pdf_sample, norm="log", rasterized=True)
+ax.contour(mass_ratio_sample, primary_mass_sample, np.log10(pdf_sample),
+           colors="k", levels=25)
 # ax.vlines(mass_ratio_boundary, 0.8, 40., ls="dashed")
 # ax.hlines(primary_mass_boundary, 0.1, 1., ls="dashed")
 ax.plot(mass_ratio_sample, 0.8/mass_ratio_sample)
-ax.text(0.3, 1.25, r"$qm_{1} < m_{\min}$", ha="center")
-ax.text(0.7, 10., r"$qm_{1} > m_{\min}$", ha="center")
-ax.set_xlim(0.1, 1.)
-ax.set_ylim(0.8, 40.)
+# ax.text(0.3, 1.25, r"$qm_{1} < m_{\min}$", ha="center")
+# ax.text(0.7, 10., r"$qm_{1} > m_{\min}$", ha="center")
+ax.set_xlim(0., 1.)
+# ax.set_ylim(0.8, 40.)
+ax.set_ylim(0.1, 100.)
 ax.set_xlabel(r"$q$")
 ax.set_ylabel(r"$M_{1}/\mathrm{M}_{\odot}$")
 ax.set_yscale("log")
 fig.colorbar(im, cax=cbar, label=r"$f_{Q|M_{1}}$")
+fig.savefig("f_QgivenM2_moe2017_pairing.jpg", dpi=300)
 fig.savefig("f_QgivenM2_moe2017_pairing.pdf", dpi=300)
 plt.show()
 
 fig, ax, cbar = plot.plot_with_colorbar()
 im = ax.pcolormesh(mass_ratio_sample, primary_mass_sample,
                    np.log10(cdf_sample), rasterized=True)
+ax.contour(mass_ratio_sample, primary_mass_sample, np.log10(cdf_sample),
+           colors="k", levels=25)
 # ax.vlines(mass_ratio_boundary, 0.8, 40., ls="dashed")
 # ax.hlines(primary_mass_boundary, 0.1, 1., ls="dashed")
 ax.plot(mass_ratio_sample, 0.8/mass_ratio_sample)
@@ -180,5 +186,5 @@ ax.set_xlabel(r"$q$")
 ax.set_ylabel(r"$M_{1}/\mathrm{M}_{\odot}$")
 ax.set_yscale("log")
 fig.colorbar(im, cax=cbar, label=r"$F_{Q|M_{1}}$")
-fig.savefig("F_QgivenM2_moe2017_pairing.pdf", dpi=300)
+# fig.savefig("F_QgivenM2_moe2017_pairing.pdf", dpi=300)
 plt.show()
