### 2025-08-07 f826662
Primary mass random variable for case of random pairing added
