+++ dyad/stats/primary_mass/__init__.py
@@ -1,0 +1,23 @@
+"""
+=============================================
+Primary mass (:mod:`dyad.stats.primary_mass`)
+=============================================
+
+.. currentmodule:: dyad.stats.primary_mass
+
+This module contains probability distributions for the primary
+masses of a population of binary stars. In its documentation the
+random variable is denoted :math:`M_{1}` and a realization of that
+random variable is denoted :math:`m_{1}`.
+
+Probability distributions
+=========================
+
+.. autosummary::
+   :toctree: generated/
+
+   kroupa2002
+
+"""
+
+from . import kroupa2002
