### 2025-08-08 5ae5e42
Upper and lower bounds of stats.mass.kroupa2002 changed
