### 2025-08-13 #7885 @warrickball
I'm sorry to say that not _quite_ yet. Your submission is now almost first in line and it now depends partly on editor expertise as well as availability. There are also some editors who are or have been away on account of vacations and conferences. I hope (but can't guarantee) that this might get picked up by early September. 🤞
