### 2025-08-14 #7885 @AmeryGration
That's great. Thanks again, @warrickball. Fingers crossed for early September.
