### 2025-08-19 11c5d99
Mass domain returned to [0.8, 40] for all moe2017 random variables
