+++ dyad/stats/period.py
@@ -238,10 +238,10 @@ class moe2017_gen(_distn_infrastructure.rv_continuous):
 
     """
     def _shape_info(self):
-        return [_ShapeInfo("p", False, (0., np.inf), (True, False))]
+        return [_ShapeInfo("p", False, (0.8, 40.), (True, False))]
 
     def _argcheck(self, primary_mass):
-        return (0. <= primary_mass) & (primary_mass < np.inf)
+        return (0.8 <= primary_mass) & (primary_mass < 40.)
 
     def _pdf(self, x, primary_mass):
         x = np.asarray(x)
