### 2025-08-19 27a1889
Mass domain returned to [0.8, 40] for all moe2017 functions in stats/data
