+++ dyad/stats/data/secondary_mass/sample_pdf.py
@@ -66,7 +66,7 @@ def f(p, q, m):
 #############################################################################
 # Create regular lattice of sample points
 #############################################################################
-primary_mass_boundary = (0.8, 1.2, 3.5, 6., 60.)
+primary_mass_boundary = (0.8, 1.2, 3.5, 6., 40.)
 mass_ratio_boundary = (0.1, 0.3, 0.95, 1.)
 log10_period_boundary = (
     0.2, 1., 1.3, 2., 2.5, 3.4, 3.5, 4., 4.5, 5.5, 6., 6.5, 8.
@@ -77,7 +77,7 @@ primary_mass_sample = np.hstack([
     np.linspace(0.8, 1.2, n),
     np.linspace(1.2, 3.5, n)[1:],
     np.linspace(3.5, 6., n)[1:],
-    np.linspace(6., 60., n)[1:],
+    np.linspace(6., 40., n)[1:],
 ])
 mass_ratio_sample = np.hstack([
     np.linspace(0.1, 0.3, n),
@@ -118,112 +118,10 @@ cdf_sample = cumulative_trapezoid(
 pdf_sample = pdf_sample/cdf_sample[:,-1:]
 cdf_sample = cdf_sample/cdf_sample[:,-1:]
 
-# #############################################################################
-# # Save data
-# #############################################################################
-# np.savetxt("./primary_mass_sample.dat", primary_mass_sample)
-# np.savetxt("./mass_ratio_sample.dat", mass_ratio_sample)
-# np.savetxt("./frequency_sample.dat", pdf_sample)
-# np.savetxt("./cumulative_frequency_sample.dat", cdf_sample)
-
-import matplotlib as mpl
-import matplotlib.pyplot as plt
-
-mpl.style.use("sm")
-
-# fig, ax = plt.subplots()
-# ax.pcolormesh(mass_ratio_sample, primary_mass_sample, np.log10(pdf_sample))
-# ax.vlines(mass_ratio_boundary, 0.8, 60.)
-# ax.hlines(primary_mass_boundary, 0.1, 1.)
-# ax.set_xlim(0.1, 1.)
-# ax.set_ylim(0.8, 60.)
-# ax.set_yscale("log")
-# plt.show()
-
-# fig, ax = plt.subplots()
-# ax.pcolormesh(mass_ratio_sample, primary_mass_sample, np.log10(cdf_sample))
-# ax.vlines(mass_ratio_boundary, 0.8, 60.)
-# ax.hlines(primary_mass_boundary, 0.1, 1.)
-# ax.set_yscale("log")
-# plt.show()
-
-########################################################################
-# Interpolate the pairing function: PDF and CDF
-########################################################################
-pdf_interp = RegularGridInterpolator(
-    (mass_ratio_sample, primary_mass_sample),
-    pdf_sample.T,
-    bounds_error=False,
-    fill_value=0.
-)
-cdf_interp = RegularGridInterpolator(
-    (mass_ratio_sample, primary_mass_sample),
-    cdf_sample.T,
-    bounds_error=False,
-    fill_value=0.
-)
-
-m = np.logspace(np.log10(0.8), np.log10(60), 2**9)
-m1m1, m2m2 = np.meshgrid(m, m)
-z = pdf_interp((m2m2/m1m1, m1m1))
-Z = cdf_interp((m2m2/m1m1, m1m1))
-
[…]
