+++ dyad/stats/data/secondary_mass/sample_extended_pdf.py
@@ -1,145 +1,0 @@
-#!/usr/bin/env python
-
-r"""Sample of the secondary-mass distribution of Moe and Stefano (2017)
-
-We require the PDF for secondary mass (conditional on primary
-mass). This is not given by Moe and Stefano (2017) but can be computed
-as follows:
-
-First, 
-
-.. math::
-
-   f_{M_{\secondary}|M_{\primary} = m_{\primary}}(m_{\secondary}|m_{\primary}) = \dfrac{1}{m_{\primary}}f_{Q}(m_{\secondary}/m_{\primary}).
-
-where
-
-.. math::
-
-   f_{Q|M_{\primary} = m_{\primary}}(q|m_{\primary}) = \int_{p_{\min}}^{p_{\max}}f_{(Q, P)|M_{\primary} = m_{\primary}}(q, p|m_{\primary})\diff{}p
-
-and, by the chain rule for probability,
-
-.. math::
-
-   f_{(Q, P)|M_{\primary} = m_{\primary}}(q, p|m_{\primary}) = f_{Q|(P, M_{\primary}) = (p, m_{\primary})}(q|p, m_{\primary})f_{P|M_{\primary} = m_{\primary}}(p|m_{\primary}).
-
-The two factors of this last formula are in fact given by Moe and
-Stefano.
-
-Neither the probability density function (PDF) nor cumulative
-distribution function (CDF) for the primary mass random variable
-implied by Moe and Stefano (2017) have closed-form expression. Dyad
-evaluates these functions by interpolating between values pre-computed
-on a regular lattice of arguments. The CDF is computed by integrating
-the PDF using the trapezium rule with nodes placed on the points of
-this lattice. Moe and Stefano give the observed frequency of
-log-period rather than its PDF. To compute the latter we must
-normalize the former. This script performs the required sampling,
-integration, and normalization, saving the results to file. Those
-files are read by `dyad/stats/secondary_mass.py`, which implements the
-random variable `dyad.stats.secondary_mass.moe2017` by performing the
-interpolation.
-
-"""
-import numpy as np
-
-from scipy.integrate import trapezoid
-from scipy.integrate import cumulative_trapezoid
-from scipy.interpolate import RegularGridInterpolator
-from dyad.stats import mass_ratio
-from dyad.stats import period as period
-
-#############################################################################
-# Define the integrand
-#############################################################################
-def f(p, q, m):
-    """Return joint probability density for mass ratio and period"""
-    res = (
-        mass_ratio.moe2017(np.log10(p), m).pdf(q)
-        *period.moe2017(m).pdf(p)
-    )
-    res = np.where((0.08 <= q*m) & (q*m <= 150.), res, 0.)
-
-    return res
-
-#############################################################################
-# Create regular lattice of sample points
-#############################################################################
-primary_mass_boundary = (0.08, 0.8, 1.2, 3.5, 6., 40., 150.)
-mass_ratio_boundary = (0., 0.1, 0.3, 0.95, 1.)
-log10_period_boundary = (
-    0.2, 1., 1.3, 2., 2.5, 3.4, 3.5, 4., 4.5, 5.5, 6., 6.5, 8.
-)
-
-n = 50
-# primary_mass_sample = np.hstack([
-#     np.linspace(0.08, 0.8, n),
-#     np.linspace(0.0, 1.2, n)[1:],
-#     np.linspace(1.2, 3.5, n)[1:],
-#     np.linspace(3.5, 6., n)[1:],
-#     np.linspace(6., 40., n)[1:],
-#     np.linspace(40., 150., n)[1:],
-# ])
[…]

[…]
