### 2025-08-19 e6d3fe7
Bug with import of log_period corrected
