+++ dyad/stats/__init__.py
@@ -74,7 +74,7 @@ from . import period
 from . import log_period
 from . import semimajor_axis
 from . import primary_mass
-from . import secondary_mass
+# from . import secondary_mass
 
 
 class _true_anomaly_gen(_distn_infrastructure.rv_continuous):
