+++ dyad/stats/data/secondary_mass/plot.py
@@ -5,13 +5,15 @@ import matplotlib as mpl
 import matplotlib.pyplot as plt
 
 from scipy.interpolate import RegularGridInterpolator
+from scipy.interpolate import LinearNDInterpolator
 
 primary_mass_sample = np.loadtxt("./primary_mass_sample.dat")
 mass_ratio_sample = np.loadtxt("./mass_ratio_sample.dat")
 pdf_sample = np.loadtxt("./frequency_sample.dat")
 cdf_sample = np.loadtxt("./cumulative_frequency_sample.dat")
 
-primary_mass_boundary = (0.8, 1.2, 3.5, 6., 40.)
+# primary_mass_boundary = (0.8, 1.2, 3.5, 6., 40.)
+primary_mass_boundary = (0.08, 1.2, 3.5, 6., 150.)
 mass_ratio_boundary = (0.1, 0.3, 0.95, 1.)
 
 mpl.style.use("sm")
@@ -27,7 +29,7 @@ ax.contour(
 ax.vlines(mass_ratio_boundary, 0.8, 40.)
 ax.hlines(primary_mass_boundary, 0.1, 1.)
 ax.set_xlim(0.1, 1.)
-ax.set_ylim(0.8, 40.)
+# ax.set_ylim(0.8, 40.)
 ax.set_yscale("log")
 ax.set_xlabel(r"$q$")
 ax.set_ylabel(r"$m_{1}/\mathrm{M}_{\odot}$")
@@ -55,20 +57,21 @@ plt.show()
 ########################################################################
 # Interpolate the pairing function: PDF and CDF
 ########################################################################
-pdf_interp = RegularGridInterpolator(
+pdf_interp = LinearNDInterpolator(
     (mass_ratio_sample, primary_mass_sample),
     pdf_sample.T,
-    bounds_error=False,
-    fill_value=0.
+    # bounds_error=False,
+    # fill_value=0.
 )
-cdf_interp = RegularGridInterpolator(
+cdf_interp = LinearNDInterpolator(
     (mass_ratio_sample, primary_mass_sample),
     cdf_sample.T,
-    bounds_error=False,
-    fill_value=0.
+    # bounds_error=False,
+    # fill_value=0.
 )
 
-m = np.logspace(np.log10(0.8), np.log10(40.), 2**9)
+# m = np.logspace(np.log10(0.8), np.log10(40.), 2**9)
+m = np.logspace(np.log10(0.08), np.log10(150.), 2**9)
 m1m1, m2m2 = np.meshgrid(m, m)
 z = pdf_interp((m2m2/m1m1, m1m1))
 Z = cdf_interp((m2m2/m1m1, m1m1))
@@ -97,8 +100,8 @@ ax.plot(m, 0.1*m, color="k", ls="solid")
 ax.plot(m, 0.3*m, color="k", ls="dashed")
 ax.plot(m, m, color="k", ls="solid")
 ax.vlines(primary_mass_boundary, 0.8, 40., ls="dashed")
-ax.set_xlim(0.8, 40.)
-ax.set_ylim(0.8, 40.)
+# ax.set_xlim(0.8, 40.)
+# ax.set_ylim(0.8, 40.)
 ax.set_xscale("log")
 ax.set_yscale("log")
 ax.set_xlabel(r"$m_{1}$")
@@ -107,7 +110,8 @@ ax.set_ylabel(r"$m_{2}$")
 # fig.savefig("./Figures/F_M2givenM2_moe2017_pairing.jpg", dpi=300)
 plt.show()
 
[…]
