### 2025-08-19 6d5bcaa
Mass domain changed to [0.08, 150] for all moe2017 random variables and in stats/data
