+++ dyad/stats/data/secondary_mass/plot.py
@@ -12,7 +12,6 @@ mass_ratio_sample = np.loadtxt("./mass_ratio_sample.dat")
 pdf_sample = np.loadtxt("./frequency_sample.dat")
 cdf_sample = np.loadtxt("./cumulative_frequency_sample.dat")
 
-# primary_mass_boundary = (0.8, 1.2, 3.5, 6., 40.)
 primary_mass_boundary = (0.08, 1.2, 3.5, 6., 150.)
 mass_ratio_boundary = (0.1, 0.3, 0.95, 1.)
 
@@ -26,10 +25,10 @@ ax.pcolormesh(
 ax.contour(
     mass_ratio_sample, primary_mass_sample, pdf_sample, colors="k", norm="log"
 )
-ax.vlines(mass_ratio_boundary, 0.8, 40.)
+ax.vlines(mass_ratio_boundary, 0.08, 150.)
 ax.hlines(primary_mass_boundary, 0.1, 1.)
 ax.set_xlim(0.1, 1.)
-# ax.set_ylim(0.8, 40.)
+ax.set_ylim(0.08, 150.)
 ax.set_yscale("log")
 ax.set_xlabel(r"$q$")
 ax.set_ylabel(r"$m_{1}/\mathrm{M}_{\odot}$")
@@ -45,7 +44,7 @@ ax.pcolormesh(
 ax.contour(
     mass_ratio_sample, primary_mass_sample, cdf_sample, colors="k", norm="log"
 )
-ax.vlines(mass_ratio_boundary, 0.8, 40.)
+ax.vlines(mass_ratio_boundary, 0.08, 150.)
 ax.hlines(primary_mass_boundary, 0.1, 1.)
 ax.set_yscale("log")
 ax.set_xlabel(r"$q$")
@@ -54,85 +53,3 @@ plt.savefig("./Figures/F_QgivenM1_moe2017.pdf", dpi=300)
 plt.savefig("./Figures/F_QgivenM1_moe2017.jpg", dpi=300)
 plt.show()
 
-########################################################################
-# Interpolate the pairing function: PDF and CDF
-########################################################################
-pdf_interp = LinearNDInterpolator(
-    (mass_ratio_sample, primary_mass_sample),
-    pdf_sample.T,
-    # bounds_error=False,
-    # fill_value=0.
-)
-cdf_interp = LinearNDInterpolator(
-    (mass_ratio_sample, primary_mass_sample),
-    cdf_sample.T,
-    # bounds_error=False,
-    # fill_value=0.
-)
-
-# m = np.logspace(np.log10(0.8), np.log10(40.), 2**9)
-m = np.logspace(np.log10(0.08), np.log10(150.), 2**9)
-m1m1, m2m2 = np.meshgrid(m, m)
-z = pdf_interp((m2m2/m1m1, m1m1))
-Z = cdf_interp((m2m2/m1m1, m1m1))
-
-fig, ax = plt.subplots()
-ax.pcolormesh(m, m, np.log10(z), cmap="Greys")
-ax.contour(m, m, np.log10(z), colors="k")
-ax.plot(m, 0.1*m, color="k", ls="solid")
-ax.plot(m, 0.3*m, color="k", ls="dashed")
-ax.plot(m, m, color="k", ls="solid")
-ax.vlines(primary_mass_boundary, 0.8, 40., ls="dashed")
-ax.set_xlim(0.8, 40.)
-ax.set_ylim(0.8, 40.)
-ax.set_xscale("log")
[…]
