+++ dyad/stats/data/secondary_mass/sample_pdf.py
@@ -104,14 +104,19 @@ period_sample = 10.**log10_period_sample
 # Sample the PDF and CDF
 #############################################################################
 primary_mass_sample_tmp = np.copy(primary_mass_sample)
-primary_mass_sample_tmp[primary_mass_sample_tmp<0.8] = 0.8 + 1.e-9
-primary_mass_sample_tmp[primary_mass_sample_tmp>=40.] = 40. - 1.e-9
+primary_mass_sample_tmp[primary_mass_sample_tmp<0.8] = 0.8
+primary_mass_sample_tmp[primary_mass_sample_tmp>=40.] = 40.
 
 pp, qq, m1m1 = np.meshgrid(
     period_sample, mass_ratio_sample, primary_mass_sample_tmp
 )
 f_sample = f(pp, qq, m1m1)
+
 pdf_sample = trapezoid(f_sample, period_sample, axis=1).T
+qq, m1m1 = np.meshgrid(
+    mass_ratio_sample, primary_mass_sample
+)
+pdf_sample[m1m1 < 0.08/qq] = 0.
 cdf_sample = cumulative_trapezoid(
     pdf_sample, mass_ratio_sample, axis=1, initial=0.
 )
