+++ dyad/stats/data/secondary_mass/plot.py
@@ -10,20 +10,22 @@ from scipy.interpolate import LinearNDInterpolator
 primary_mass_sample = np.loadtxt("./primary_mass_sample.dat")
 mass_ratio_sample = np.loadtxt("./mass_ratio_sample.dat")
 pdf_sample = np.loadtxt("./frequency_sample.dat")
+pdf_sample[pdf_sample == np.max(pdf_sample)] = 0.
 cdf_sample = np.loadtxt("./cumulative_frequency_sample.dat")
 
-primary_mass_boundary = (0.08, 1.2, 3.5, 6., 150.)
+primary_mass_boundary = (0.08, 0.8, 1.2, 3.5, 6., 40.)
 mass_ratio_boundary = (0.1, 0.3, 0.95, 1.)
 
 mpl.style.use("sm")
 
 fig, ax = plt.subplots()
 ax.pcolormesh(
-    mass_ratio_sample, primary_mass_sample, np.log10(pdf_sample),
+    mass_ratio_sample, primary_mass_sample, pdf_sample, norm="log",
     rasterized=True
 )
 ax.contour(
-    mass_ratio_sample, primary_mass_sample, pdf_sample, colors="k", norm="log"
+    mass_ratio_sample, primary_mass_sample, np.log10(pdf_sample), levels=10,
+    colors="k"
 )
 ax.plot(mass_ratio_sample, 0.08/mass_ratio_sample)
 ax.vlines(mass_ratio_boundary, 0.08, 150.)
@@ -39,11 +41,12 @@ plt.show()
 
 fig, ax = plt.subplots()
 ax.pcolormesh(
-    mass_ratio_sample, primary_mass_sample, np.log10(cdf_sample),
-    rasterized=True
+    mass_ratio_sample, primary_mass_sample, cdf_sample, norm="log",
+    rasterized=True, 
 )
 ax.contour(
-    mass_ratio_sample, primary_mass_sample, cdf_sample, colors="k", norm="log"
+    mass_ratio_sample, primary_mass_sample, np.log10(cdf_sample), levels=10,
+    colors="k",
 )
 ax.plot(mass_ratio_sample, 0.08/mass_ratio_sample)
 ax.vlines(mass_ratio_boundary, 0.08, 150.)
