### 2025-08-19 8f6f859
Mass domain changed to [0.08, 150] for secondary_mass random variable
