### 2025-08-22 65b2881
Mass bounds in moe2017 rvs returned to 0.8 and 40. The rvs primary_mass and secondary_mass removed
