+++ dyad/stats/data/log_period/sample_pdf.py
@@ -165,18 +165,18 @@ def pdf(x, primary_mass):
 #############################################################################
 # Create grid of sample points
 #############################################################################
-primary_mass_boundary = (0.08, 1.2, 3.5, 6., 150.)
+primary_mass_boundary = (0.8, 1.2, 3.5, 6., 40.)
 log10_period_boundary = (
     0.2, 1., 1.3, 2., 2.5, 3.4, 3.5, 4., 4.5, 5.5, 6., 6.5, 8.
 )
 
-n = 150
+n = 50
 primary_mass_sample = np.hstack(
     [
-        np.logspace(np.log10(0.08) + 1.e-9, np.log10(1.2), n),
+        np.logspace(np.log10(0.8) + 1.e-9, np.log10(1.2), n),
         np.logspace(np.log10(1.2), np.log10(3.5), n)[1:],
         np.logspace(np.log10(3.5), np.log10(6.), n)[1:],
-        np.logspace(np.log10(6.), np.log10(150.) - 1.e-9, n)[1:],
+        np.logspace(np.log10(6.), np.log10(40.) - 1.e-9, n)[1:],
     ]
 )
 log10_period_sample = np.hstack(
