+++ dyad/stats/__init__.py
@@ -70,11 +70,9 @@ from . import _distn_infrastructure
 from . import eccentricity
 from . import mass
 from . import mass_ratio
-from . import period
 from . import log_period
+from . import period
 from . import semimajor_axis
-from . import primary_mass
-from . import secondary_mass
 
 
 class _true_anomaly_gen(_distn_infrastructure.rv_continuous):

[…]
