### 2025-08-30 #7885 @warrickball
Sorry again for the long wait @AmeryGration. I've got some capacity coming up so will take this on as editor and start	looking for (at least 2) reviewers. This usually takes 2-3 weeks. If there's anyone you can recommend from your own community or our [reviewer database](https://reviewers.joss.theoj.org/), do let me know, though bear in mind our [conflict of interest policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy).
