### 2025-08-30 #7885 @warrickball
Also, I'm having a closer look at the paper itself and it isn't clear to me personally what the precise purpose of Dyad is, beyond simply implementing distributions to sample from. I noticed this sentence:

> Specifically, Dyad allows the user to synthesize primary- and secondary-star masses in a manner consistent with given distributions of stellar masses and mass ratios

If this is a major feature, could you perhaps add a figure and snippet of text to demonstrate this? E.g. a figure with two distributions, first with whatever is the method that is *in*consistent and then with Dyad, showing the improvement in the latter. While we like short papers at JOSS, you still have space to demonstrate why Dyad is useful.
