### 2025-09-03 #7885 @AmeryGration
Dear @warrickball. That's great. Thank you so much. I'm afraid that I'm on annual leave for a fortnight and won't be back in the office until 15 September. Could I take a look at this once I'm back? I'm sorry that the paper doesn't make Dyad's purpose clear enough. I must be too close to the material. It has two purposes. One is to provide Python classes that compute various physical quantities related to orbital motion The ~Orbit~ class does this for a body moving in a central potential. The ~TwoBody~ class does this for the components of a two-body system. The other purpose is indeed to provide distributions to sample from. Using the properties generated using these distributions you can then instantiate an Orbit or TwoBody class. The main use case would be the initialization of stellar evolution population synthesis codes or dynamical population synthesis codes. I can certainly add a figure and a snippet of code.
