### 2025-09-03 #7885 @warrickball
> One is to provide Python classes that compute various physical quantities related to orbital motion The Orbit class does this for a body moving in a central potential. The TwoBody class does this for the components of a two-body system.

This on its own sounds a bit small for a JOSS submission.

> The other purpose is indeed to provide distributions to sample from. Using the properties generated using these distributions you can then instantiate an Orbit or TwoBody class.

Again, the risk here is that sampling from distributions isn't obviously a big enough application for a JOSS submission. But! My impression from reading the paper is that the situation isn't so simple and that sampling correctly is (a) not trivial and (b) not implemented in existing open source tools (or those tools have notable pitfalls).

If there are specific downstream packages you have in mind, it's worth mentioning them if you haven't already. I've only glanced at some references and no specific packages jumped out at me.
