### 2025-09-03 #7885 @warrickball
I notice the automatic reference test above didn't run at first. I'll see if I can help figure out the error above but I noticed that there are generally few DOIs. Astronomy submissions often use the BibTeX entries from NASA ADS (e.g. [Arroyo-Polonio et al. 2023](https://ui.adsabs.harvard.edu/abs/2023A%26A...677A..95A/exportcitation)), though you need to replace the journal macros (e.g. `\aap`) with the full journal name (*Astronomy & Astrophysics*).
