### 2025-09-15 #7885 @AmeryGration
Hi @warrickball. Thank you for waiting while I finished my break. Regarding the scope of Dyad, you're right in saying that in several cases the sampling isn't trivial and isn't implemented by any existing tools. This is true of the distributions named *.moe2017. These are implementations of the distributions published by Moe & Di Stefano (2017). It has taken me considerable time to convert these into explicit form and to implement them.
