### 2025-09-18 2d8f7d6
Primary- and secondary-mass random variables added
