+++ dyad/stats/period.py
@@ -219,9 +219,9 @@ class moe2017_gen(_distn_infrastructure.rv_continuous):
        f_{P|M_{1}}(p|m_{1}) =
        \left|\dfrac{1}{\ln(10)p}\right|f_{X|M_{1}}(\log_{10}(p)|m_{1}).
 
-    for :math:`p \in [10^{0.2}, 10^{8}]` and where :math:`f_{X|M_{1}}` is the
-    probability density function for log-period given by Moe and
-    Stefano (2017).
+    for :math:`p \in [10^{0.2}, 10^{8}]` and where :math:`f_{X|M_{1}}`
+    is the probability density function for log-period given by Moe
+    and Stefano (2017).
 
     See also
     --------
@@ -244,22 +244,22 @@ class moe2017_gen(_distn_infrastructure.rv_continuous):
         return (0.8 <= primary_mass) & (primary_mass < 40.)
 
     def _pdf(self, x, primary_mass):
-        x = np.asarray(x)
-        primary_mass = np.asarray(primary_mass)
+        # x = np.asarray(x)
+        # primary_mass = np.asarray(primary_mass)
         res = _moe2017_pdf_interp((x, primary_mass))
         
         return res
 
     def _cdf(self, x, primary_mass):
-        x = np.asarray(x)
-        primary_mass = np.asarray(primary_mass)
+        # x = np.asarray(x)
+        # primary_mass = np.asarray(primary_mass)
         res = _moe2017_cdf_interp((x, primary_mass))
         
         return res
 
     def _ppf(self, q, primary_mass):
-        q = np.asarray(q)
-        primary_mass = np.asarray(primary_mass)
+        # q = np.asarray(q)
+        # primary_mass = np.asarray(primary_mass)
         res = _moe2017_ppf_interp((q, primary_mass))
         
         return res
