### 2025-09-18 #9011 @warrickball
@AmeryGration — I think the bibliography check is failing on the "forthcoming" `gration2025a` and `gration2025b` entries in the BibTeX file. They're defined twice each, and the first definitions have blank `month = ,` fields which are what I think the parser is failing on. If you could please fix that (removing the first definitions of each entry should work), let me know and I'll rerun the bibliography check (which mainly just checks our DOIs coverage).
