+++ docs/source/generated/generated/dyad.stats.primary_mass.rst
@@ -1,0 +1,6 @@
+﻿dyad.stats.primary\_mass
+========================
+
+.. automodule:: dyad.stats.primary_mass
+
+   

[…]
