### 2025-09-22 5f8092a
Secondary-mass random variable returned to codebase. Doctests implemented for tutorials.
