+++ dyad/stats/secondary_mass.py
@@ -1,0 +1,186 @@
+"""
+=================================================
+Secondary mass (:mod:`dyad.stats.secondary_mass`)
+=================================================
+
+.. currentmodule:: dyad.stats.secondary_mass
+
+This module contains probability distributions for the secondary
+masses of a population of binary stars. In its documentation the
+random variable is denoted :math:`M_{2}` and a realization of that
+random variable is denoted :math:`m_{2}`. The random variable for
+secondary mass is conditional on primary mass. The PDF for secondary
+mass is known as the `pairing function'
+
+Probability distributions
+=========================
+
+.. autosummary::
+   :toctree: generated/
+
+   moe2017
+
+"""
+
+__all__ = [
+    "moe2017",
+]
+
+import os
+import numpy as np
+import scipy as sp
+
+from importlib.resources import files
+from scipy.interpolate import RegularGridInterpolator
+from scipy.interpolate import LinearNDInterpolator
+from . import _distn_infrastructure
+
+
+class moe2017_gen(_distn_infrastructure.rv_continuous):
+    r"""The secondary-star mass random variable of Moe and Stefano (2017)
+
+    %(before_notes)s
+
+    Notes
+    -----
+    The probability density function for `moe2017` is:
+
+    .. math::
+       f_{M_{2}|M_{1} = m_{1}}(m_{2}|m_{1}) = \dfrac{1}{m_{1}}f_{Q}(m_{2}/m_{1}).
+
+    where
+
+    .. math::
+       f_{Q|M_{1} = m_{1}}(q|m_{1}) = \int_{p_{\min}}^{p_{\max}}f_{(Q, P)|M_{1} = m_{1}}(q, p|m_{1})\mathrm{d}\,p
+
+    and, by the chain rule for probability,
+
+    .. math::
+       f_{(Q, P)|M_{1} = m_{1}}(q, p|m_{1}) = f_{Q|(P, M_{1}) = (p, m_{1})}(q|p, m_{1})f_{P|M_{1} = m_{1}}(p|m_{1}).
+
+    for primary-star mass :math:`m_{1} \in (m_{1}, 60)`,
+    secondary-star mass :math:`m_{2} \in (0.1m_{1}, 60)`, mass ratio
+    :math:`q \in [0.1, 1]`, and period :math:`p \in [10^{0.2},
+    10^{8}]`. The functions :math:`f_{Q|(P, M_{1}) = (p, m_{1})}` and
+    :math:`f_{P|M_{1} = m_{1}}` are the probability density functions
+    for random variables :class:`dyad.stats.mass_ratio.moe2017` and
+    :class:`dyad.stats.period.moe2017`.
+
+    ``moe2017`` takes ``primary_mass`` as a shape parameter for
+    :math:`m_{1}`, the primary mass.
+
+    %(after_notes)s
+
+    See also
+    --------
+
+    References
+    ----------
+    Moe, Maxwell, and Rosanne Di Stefano. 2017. \'Mind your Ps and Qs:
+    the interrelation between period (P) and mass-ratio (Q)
+    distributions of binary stars.\' *The Astrophysical Journal
+    Supplement Series* 230 (2): 15.
+
+    %(example)s
+
+    """
+    def _shape_info(self):
+        return [
+            _ShapeInfo("primary_mass", False, (0.08, 150.), (False, False))
+        ]
+
+    def _argcheck(self, primary_mass):
+        return (0.08 <= primary_mass) & (primary_mass <= 150.)
+        # return (0. <= primary_mass) & (primary_mass < np.inf)
+    
+    def _pdf(self, x, primary_mass):
+        # x = np.asarray(x)
+        # primary_mass = np.asarray(primary_mass)
[…]
