+++ dyad/stats/__init__.py
@@ -72,6 +72,7 @@ from . import mass
 from . import mass_ratio
 from . import log_period
 from . import period
+from . import secondary_mass
 from . import semimajor_axis
 
 
