+++ docs/source/generated/generated/dyad.stats.secondary_mass.rst
@@ -9,5 +9,4 @@
    .. autosummary::
    
       moe2017_gen
-      random_gen
    

[…]
