+++ README.md
@@ -1,4 +1,6 @@
-# Dyad
+####
+Dyad
+####
 
 Dyad is a two-body dynamics and binary-star statistics package for astrophysicists.
 
