+++ README.md
@@ -1,7 +1,3 @@
-####
-Dyad
-####
-
 Dyad is a two-body dynamics and binary-star statistics package for astrophysicists.
 
 It provides:
