+++ docs/source/index.rst
@@ -1,3 +1,7 @@
+####
+Dyad
+####
+
 .. include:: ../../README.md
 
 Getting started
