### 2025-09-22 8f2c44c
Bibliography updated. DOIs added
