### 2025-09-22 214a7ee
Bibliography updated. DOIs added
