### 2025-09-22 3e2d7f2
Bibliography updated. DOIs added
