+++ paper.bib
@@ -54,7 +54,7 @@
 @article{arroyo-polonio2023,
   author = {{Arroyo-Polonio}, J. M. and Battaglia, G. and Thomas, G. F. and Irwin, M. J. and McConnachie, A. W. and Tolstoy, E.},
   year = {2023},
-  month = sep
+  month = sep,
   volume = {677},
   pages = {A95}
   journal = {Astronomy \& Astrophysics},
