### 2025-09-22 cf91f02
Bibliography updated. DOIs added
