+++ paper.bib
@@ -56,7 +56,7 @@
   year = {2023},
   month = sep,
   volume = {677},
-  pages = {A95}
+  pages = {A95},
   journal = {Astronomy \& Astrophysics},
   doi = {10.1051/0004-6361/202346843},
 }
