### 2025-09-22 dcff3a7
Bibliography updated. DOIs added
