### 2025-09-22 eb723e3
Bibliography updated. DOIs added
