+++ paper.bib
@@ -62,7 +62,7 @@
 }
 
 @article{gration2025a,
-  author = {Gration, A., Hendricks, D., Das, P., Heber, D., Izzard, R.},
+  author = {Gration, A. and Hendricks, D. and Das, P. and Heber, D. and Izzard, R.},
   year = {2025},
   month = sep,
   journal = {arXiv e-prints},
@@ -70,7 +70,7 @@
 }
 
 @article{gration2025b,
-  author = {Gration, A., Izzard, R. G., and Das, P.},
+  author = {Gration, A. and Izzard, R. G. and Das, P.},
   year = {In prep},
   month = ,
   journal = {},
@@ -80,7 +80,7 @@
 }
 
 @article{hendriks2023,
-  author = {Hendriks, D. D., and Izzard, R. G.},
+  author = {Hendriks, D. D. and Izzard, R. G.},
   journal = {Journal of Open Source Software},
   year = {2023},
   publisher = {The Open Journal},
