### 2025-09-22 db87b6b
Bibliography updated. DOIs added
