+++ paper.bib
@@ -72,11 +72,6 @@
 @article{gration2025b,
   author = {Gration, A. and Izzard, R. G. and Das, P.},
   year = {In prep},
-  month = ,
-  journal = {},
-  volume = {},
-  pages = {},
-  doi = {},
 }
 
 @article{hendriks2023,
