### 2025-09-22 404e7f1
Bibliography updated. DOIs added
