+++ paper.bib
@@ -71,7 +71,7 @@
 
 @article{gration2025b,
   author = {Gration, A. and Izzard, R. G. and Das, P.},
-  year = {In prep},
+  pubstate = {inpreparation},
 }
 
 @article{hendriks2023,
