### 2025-09-22 f759a6e
Bibliography updated. DOIs added
