+++ paper.md
@@ -1,3 +1,4 @@
+---
 title: 'Dyad: a binary-star dynamics and statistics package for Python'
 tags:
   - Python
