### 2025-09-23 d7ed7ff
Bibliography updated
