+++ paper.bib
@@ -134,3 +134,12 @@
   address = {Cambridge},
   doi = {10.1017/9781108553070.006},
 }
+
+@article{duquennoy1991,
+  author = {Duquennoy, A. and Mayor, M.},
+  year = {1991},
+  month = aug,
+  journal = {Astronomy and Astrophysics},
+  volume = {248},
+  pages = {485--524},
+}
