### 2025-09-23 089019b
Bibliography updated
