+++ paper.bib
@@ -142,4 +142,5 @@
   journal = {Astronomy and Astrophysics},
   volume = {248},
   pages = {485--524},
+  doi = {10.3847/1538-4365/aa6fb6}
 }
