### 2025-09-30 #9011 @stegmaja
@AmeryGration

Regarding the "Substantial scholarly effort"/"Statement of need"/"Functionality":

The functionality of `dyad.Orbit` and `dyad.TwoBody` looks quite limited at the moment and already covered by many other codes (e.g., [twobody](https://twobody.readthedocs.io/en/latest/). Looking at the [list of functions](https://dyad-a-two-body-dynamics-and-binary-star-statistics-package.readthedocs.io/en/latest/generated/dyad.html#module-dyad) of the classes, these need to be implemented for almost any science application to binaries. However, given their simple forms I am unsure whether users will pick up `dyad` for this task. [To compute the kinematic properties of binary systems](https://raw.githubusercontent.com/openjournals/joss-papers/joss.09011/joss.09011/10.21105.joss.09011.pdf) and to ensure that the codes makes [addressing research challenges significantly better (e.g., faster, easier, simpler)](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) I think the current functionality needs to be significantly extended. For instance, it would be very useful if, instead of (m,a,e,Omega,omega,i) I could supply either the Cartesian coordinates (x_1,x_2,v_1,v_2) of two particles, their relative vectors (r,v), or vectorial elements (e,h,n) and the code would return all of the alternative representations. I do see some of the Cartesian representation in the [stats examples](https://dyad-a-two-body-dynamics-and-binary-star-statistics-package.readthedocs.io/en/latest/tutorials/tutorial_2.html#a-complete-population); so please clarify if something similar is already included. A bonus would be if this could be done for any number N>=2 of particles in hierarchical systems (e.g., triples).

Additionally, it is not entirely clear `dyad.Orbit` and `dyad.TwoBody` are two separate functions since mathematically the two-body problem [reduces to that of one body in a central potential](https://en.wikipedia.org/wiki/Two-body_problem).
