### 2025-10-02 #9011 @AmeryGration
> Hi! Happy to be on board with the review process and I apologize for the delay in responsiveness. I just have a follow-up question on the cited text below, with the premise that I am still doing explorative steps within the code.
>
> > Thank you for looking at my code so carefully, [@stegmaja](https://github.com/stegmaja). You are quite right that the functionality of `dyad.Orbit` and `dyad.TwoBody` is limited, that it is easily implemented, and has in fact been implemented elsewhere. The real scholarly effort has been not in implementing these classes but in implementing the library of random variables provided by the `dyad.stats` module. This library provides a way of initializing `dyad.Orbit` and `dyad.TwoBody` with randomly chosen values for stellar mass, period, and orbital elements using a chosen distribution. It was a particular effort to implement the distributions given by Moe and Di Stefano in their 2017 paper on zero-age main-sequence binary systems and I believe that Dyad is the only code to implement them in the way that paper suggests (i.e. as formulae for the parameters' PDFs). You'll find these under `dyad.stats.{eccentricity,log_period,period,mass_ratio,secondary_mass}.moe2017`. This library contains Dyad's real functionality and I intend to add to it over time. The `dyad.Orbit` and `dyad.TwoBody` classes are there for the sake of completeness. Users may well wish to use the random variables of Dyad in conjunction with another dynamics code such as `twobody` or indeed some other population synthesis code. And you are right of course that `dyad.TwoBody` is redundant. I provide it as a convenience class that wraps two instantiations of `dyad.Orbit`.
>
> In a more general perspective, I am wondering how this relates to codes such as BOSSA, which also is a population synthesis sampling code (https://academic.oup.com/mnras/article/535/3/2019/7825861). For instance to my knowledge the developer(s) also implements the IMF for the parameters' PDFs, and uses Moe & di Stefano (2017) not only for IMF consideration, but also weights their IMF as a function (to my understanding) of the highest multiplicity order AND the maximum ZAMS mass.

Hi @AmedeoRom. I wasn't aware of BOSSA, which I see was published shortly before I submitted Dyad to JOSS. Yes, it seems that they have implemented the Moe and Di Stefano (2017) distributions using the same formulae. For example in `https://github.com/LucasMDeSa/BOSSA/blob/main/src/bossa/zams.py` I see classes `EccentricityDistribution` and `MassRatioDistribution` each with a method `prob` that does the same thing as the method `pdf` in my classes `stats.eccentricity.moe2017` and `stats.mass_ratio.moe2017`. For what it's worth I will say that I implement these using Scipy's native `rv_continuous` so there the additional functionality of the other `rv_continuous` methods ([here](https://docs.scipy.org/doc/scipy/reference/generated/scipy.stats.rv_continuous.html)).
