### 2025-10-02 #9011 @AmeryGration
> Adding to the comment of [@stegmaja](https://github.com/stegmaja), I have a few points while I am checking the code:
>
>     * I second his statement on Cartesian coordinates, and I would also add that it would be great also to choose different units as an added condition. For instance, many of us use Solar units for everything, even orbital separation. It would be of great impact if there were options to change units on the fly, constants included. Not saying something like every variable must have options to covert units, but at least choosing big classes such as "everything in solar units" or "everything in cgs"
>
>     * An added very simple functionality that would be of great interest for the massive star/objects community would be also tracking the evolution of the orbit past one revolution, even more within the context of orbital energy loss, which can either be arbitrary or modeled. The whole gravitational wave (GW) community would be extremely interested if the code had the opportunity to calculate the change in eccentricity and separation due to GW emission. If this could be attached also to the GW strain and frequency, this code would have the potential of being consistently used by several groups.
>
>     * I would dare saying that the case of N>2 particles like MSE (https://arxiv.org/abs/2011.04513)  would be another extremely interesting part to add, but I am no expert on how to implement the Kozai-Lidov mechanism. It would be a major feature not only in the context of multiple stellar systems (and then one would need to think about the IMF and initial orbital configuration of the other objects in the populations), but also in terms of planetology studies, since it could be used to study systems of an arbitrary number of planets around an arbitrary number of stars

Hi @AmedeoRom. I wasn't aware of BOSSA, which I see was published shortly before I submitted Dyad to JOSS. And, yes, it seems that they have implemented the Moe and Di Stefano (2017) distributions using the same formulae. For example in [BOSSA/src/bossa/zams.py](https://github.com/LucasMDeSa/BOSSA/blob/main/src/bossa/zams.py) I see classes `EccentricityDistribution` and `MassRatioDistribution` each with a method `prob` that does the same thing as the method `pdf` in my classes `stats.eccentricity.moe2017` and `stats.mass_ratio.moe2017`. For what it's worth I will say that I implement these using Scipy's native `rv_continuous` class so there is also the additional functionality of that class's other methods ([here](https://docs.scipy.org/doc/scipy/reference/generated/scipy.stats.rv_continuous.html)).

It seems that the greatest difference is in the sampling mechanism. With BOSSA you may use an out-of-module sampler in conjunction with their `prob` methods or (as far I can tell) draw from a precomputed sample population ([here](https://github.com/LucasMDeSa/BOSSA/blob/main/notebooks/63_Orbital_Sampling_Eval.ipynb)). In Dyad you can use the class method `rvs` (e.g. `stats.mass_ratio.moe2017.pdf`). This allows you to sample from each random variable separately. You can build a complete population in this way, though a convenience function that does this for you might be useful.
