### 2025-10-15 #9011 @AmeryGration
HI @stegmaja. I'm sorry! I hadn't realized that this had actually stopped you from installing. I've updated my `setup.py` and added a `requirements.txt`. The issue is with your version of Scipy and I have specified the required version.

I've also updated the installation instructions as you suggested.
