### 2025-10-15 d1070dc
Installation instructions in documentation updated
