+++ docs/source/getting_started/installation.rst
@@ -7,8 +7,7 @@ Installation
 You may install Dyad from the source distribution, available at
 the Dyad repository
 (`https://github.com/AmeryGration/dyad`__). Make a local copy of
-the repository and navigate to its top level (the directory
-``dyad``). Then run the following command.
+the repository and navigate to its top level. Then run the following command.
 
 .. sourcecode:: bash
 
