### 2025-10-15 #9011 @warrickball
As a general remark on Python packages, whenever a git repo contains a Python package, [you can install it directly](https://pip.pypa.io/en/stable/topics/vcs-support/) (without having to clone it yourself first) with e.g. `pip install git+https://github.com/AmeryGration/dyad.git`. This just worked for me in a fresh virtual environment in Python 3.13.7, though my system version of SciPy already satisfies the new version requirement.
