### 2025-11-03 b864499
RV mass.salpeter1955 now a truncated distributuion. Circularization period for {log_period,period}.moe2017 now a constant value. Tutorial on moe2017 distributions added. Error in log-period corrected.
