+++ docs/build/html/_sources/api_reference/generated/dyad.stats.period.rst.txt
@@ -1,0 +1,14 @@
+﻿dyad.stats.period
+=================
+
+.. automodule:: dyad.stats.period
+
+   
+   .. rubric:: Classes
+
+   .. autosummary::
+   
+      duquennoy1991_gen
+      moe2017_gen
+      trunclognorm_gen
+   
