+++ docs/build/html/_sources/generated/generated/generated/dyad.stats.eccentricity.duquennoy1991.rst.txt
@@ -1,0 +1,6 @@
+﻿dyad.stats.eccentricity.duquennoy1991
+=====================================
+
+.. currentmodule:: dyad.stats.eccentricity
+
+.. autodata:: duquennoy1991
