+++ docs/build/html/_sources/getting_started/installation.rst.txt
@@ -1,0 +1,18 @@
+.. _installation:
+
+************
+Installation
+************
+
+You may install Dyad from the source distribution, available at
+the Dyad repository
+(`https://github.com/AmeryGration/dyad`__). Make a local copy of
+the repository and navigate to its top level. Then run the following command.
+
+.. sourcecode:: bash
+
+   $ python -m pip install .
+   
+__ https://github.com/AmeryGration/dyad
+
+Dyad requires Python version 3.10 or greater.
