+++ docs/build/html/_sources/generated/generated/dyad.TwoBody.rst.txt
@@ -1,0 +1,22 @@
+﻿dyad.TwoBody
+============
+
+.. currentmodule:: dyad
+
+.. autoclass:: TwoBody
+
+   
+   .. automethod:: __init__
+
+   
+   .. rubric:: Methods
+
+   .. autosummary::
+   
+      ~TwoBody.__init__
+   
+   
+
+   
+   
+   
