+++ docs/build/html/_sources/generated/generated/dyad.stats.mass.splitpowerlaw.rst.txt
@@ -1,0 +1,6 @@
+﻿dyad.stats.mass.splitpowerlaw
+=============================
+
+.. currentmodule:: dyad.stats.mass
+
+.. autodata:: splitpowerlaw

[…]
