+++ docs/build/html/_sources/tutorials/tutorial_3.rst.txt
@@ -47,7 +47,7 @@ To allow for broadcasting we must increase the dimension of ``m_1`` using the no
 .. doctest:: python
 
    >>> import dyad.stats as stats
-   >>> log_p = np.linspace(-1., 9.)
+   >>> log_p = np.linspace(0.2, 8.)
    >>> f = stats.log_period.moe2017(m_1[:,None]).pdf(log_p)
 
 Now let us plot these values.
@@ -224,7 +224,7 @@ Let us now synthesize a population of binary systems.
 We will use the primary-constrained pairing method to synthesize the primary- and secondary-star masses [K09]_.
 According to this method we synthesize the primary-star star mass assuming that it is distributed according to the initial mass function and then synthesize the secondary-star mass using the conditional distribution of secondary-star mass given primary-star mass.
 
-First, speciy a sample size.
+First, specify a sample size.
 
 .. doctest:: python
 
