### 2025-11-04 157df42
Random variable stats.mass.kroupa2002 now in truncated form
