### 2025-11-17 cbb0e7a
Tests for random variables added
