+++ tests/unit_tests/data_test_mass.py
@@ -1,0 +1,97 @@
+"""Data to be used when testing Dyad's stats.mass module
+
+This module is not to be run on its own. It is imported by
+`test_stats_mass.py`.
+
+When checking for value errors each list contains arguments that
+should cause such an error. If a value error is raised then the test
+is passed.
+
+Each element in a list is itself a list that contains two elements:
+1. the arguments for a function being tested, and
+2. the desired return value of that function for these arguments.
+
+For example to test the function `dyad.func(*args)` we use the
+following data.
+
+.. code-block:: python
+   func = [
+       [args, val],
+       [args, val],
+       ...
+       [args, val],
+   ]
+
+If the function `dyad.stats.mass.{attribute,method}(*args)` returns
+the value `val` then the test is passed.
+
+"""
+salpeter1955_pdf = [
+    [[0.08, 0.08, 150.0], 16.875643737236214],
+    [[150.0, 0.08, 150.0], 3.4332652598078764e-07],
+    [[97.38967978113911, 0.08, 150.0], 9.473629886889312e-07],
+    [[[0.08, 10.0, 50.0, 100.0, 150.0], 0.08, 150.0], [16.875643737236214, 0.0001993066630108997, 4.5388131832896714e-06, 8.902701617333015e-07, 3.4332652598078764e-07]],
+]
+salpeter1955_cdf = [
+    [[0.08, 0.08, 150.0], 0.0],
+    [[150.0, 0.08, 150.0], 1.0],
+    [[97.38967978113911, 0.08, 150.0], 0.999969804148735],
+    [[[0.08, 10.0, 50.0, 100.0, 150.0], 0.08, 150.0], [0.0, 0.9985618017398431, 0.999870043199802, 0.9999722014538694, 1.0]],
+]
+salpeter1955_ppf = [
+    [[0.0, 0.08, 150.0], 0.08],
+    [[1.0, 0.08, 150.0], 150.0],
+    [[0.999969804148735, 0.08, 150.0], 97.38967978118919],
+    [[[0.0, 0.9985618017398431, 0.999870043199802, 0.9999722014538694, 1.0], 0.08, 150.0], [0.08, 10.000000000000265, 50.00000000000647, 99.999999999936, 150.0]],
+]
+salpeter1955_rvs = [
+    [[0.08, 150.0, 0.0, 1.0, 1, 93438701], [0.10180963931891547]],
+    [[0.08, 150.0, 0.0, 1.0, 2, 93438701], [0.10180963931891547, 0.14331112352366235]],
+    [[0.08, 150.0, 0.0, 1.0, (2, 3), 93438701], [[0.10180963931891547, 0.14331112352366235, 0.14428697751151806], [0.08919215550721653, 0.2213921758922736, 0.18503214231025114]]],
+]
+kroupa2002_pdf = [
+    [[0.08, 0.08, 150.0], 6.7444570546415665],
+    [[150.0, 0.08, 150.0], 1.2500405037958306e-06],
+    [[97.38967978113911, 0.08, 150.0], 3.3756252280887457e-06],
+    [[[0.08, 10.0, 50.0, 100.0, 150.0], 0.08, 150.0], [6.7444570546415665, 0.0006337733613780491, 1.5642385010405695e-05, 3.1763911773572216e-06, 1.2500405037958306e-06]],
+]
[…]
