### 2025-11-17 87816f5
Missing .gitignore restored
