### 2025-11-17 fee9b95
Missing .gitignore restored
