+++ docs/source/conf.py
@@ -1,534 +1,0 @@
-import math
-import os
-import re
-import sys
-import warnings
-import matplotlib
-import matplotlib.pyplot as plt
-import scipy
-import scipy._lib.uarray as ua
-import numpydoc.docscrape as np_docscrape
-
-from os.path import relpath, dirname
-from docutils import nodes
-from docutils.parsers.rst import Directive
-from intersphinx_registry import get_intersphinx_mapping
-from numpydoc.docscrape_sphinx import SphinxDocString
-from sphinx.util import inspect
-from scipy._lib._util import _rng_html_rewrite
-# Workaround for sphinx-doc/sphinx#6573
-# ua._Function should not be treated as an attribute
-from scipy.stats._distn_infrastructure import rv_generic
-from scipy.stats._multivariate import multi_rv_generic
-
-old_isdesc = inspect.isdescriptor
-inspect.isdescriptor = (lambda obj: old_isdesc(obj)
-                        and not isinstance(obj, ua._Function))
-
-sys.path.insert(0, os.path.abspath(os.path.dirname(__file__)))
-
-########################################################################
-# General configuration
-########################################################################
-
-extensions = [
-    "sphinx.ext.viewcode",
-    "sphinx.ext.autodoc",
-    "sphinx.ext.doctest",
-    "sphinx.ext.autosummary",
-    "sphinx.ext.coverage",
-    "sphinx.ext.mathjax",
-    "sphinx.ext.intersphinx",
-    "numpydoc",
-    "sphinx_copybutton",
-    "sphinx_design",
-    # "scipyoptdoc",
-    # "doi_role",
-    "matplotlib.sphinxext.plot_directive",
-    # "myst_nb",
-    # "jupyterlite_sphinx",
-]
-
-html_logo = "dyad_logo_white.png"
-
-# Do some matplotlib config in case users have a matplotlibrc that will break
-# things
-matplotlib.use("agg")
-plt.ioff()
-
-# Add any paths that contain templates here, relative to this directory.
-templates_path = ["_templates"]
-
-# The suffix of source filenames.
-source_suffix = ".rst"
-
-numfig = True
-
-# The main toctree document.
-master_doc = "index"
-
-# General substitutions.
-sys.path.insert(0, os.path.abspath("../../dyad"))
-
-project = "Dyad"
-copyright = "2024, Amery Gration"
-author = "Amery Gration"
-version = "0.0.0"
-release = "0.0.0"
-
-if os.environ.get("CIRCLE_JOB", False) and \
-        os.environ.get("CIRCLE_BRANCH", "") != "main":
-    version = os.environ["CIRCLE_BRANCH"]
-    release = version
-
-print(f"{project} (VERSION {version})")
-
-# There are two options for replacing |today|: either, you set today to some
-# non-false value, then it is used:
-#today = ""
-# Else, today_fmt is used as the format for a strftime call.
-today_fmt = "%B %d, %Y"
-
-# List of documents that shouldn"t be included in the build.
-#unused_docs = []
-
-# The reST default role (used for this markup: `text`) to use for all documents.
-default_role = "autolink"
-
-# List of directories, relative to source directories, that shouldn"t be searched
-# for source files.
-exclude_dirs = []
-exclude_patterns = [  # glob-style
-    "**.ipynb",
-]
-
-# If true, "()" will be appended to :func: etc. cross-reference text.
-add_function_parentheses = False
-
-# If true, the current module name will be prepended to all description
-# unit titles (such as .. function::).
-add_module_names = True
-
-# If true, sectionauthor and moduleauthor directives will be shown in the
-# output. They are ignored by default.
-show_authors = False
-
[…]
