+++ docs/source/api_reference/constants.rst
@@ -1,2 +1,0 @@
-.. automodule:: dyad.constants
-   :members:
