+++ docs/source/generated/generated/dyad.semimajor_axis_from_period.rst
@@ -1,6 +1,0 @@
-﻿dyad.semimajor\_axis\_from\_period
-==================================
-
-.. currentmodule:: dyad
-
-.. autofunction:: semimajor_axis_from_period
