+++ docs/source/generated/generated/dyad.stats.mass_ratio.moe2017.rst
@@ -1,6 +1,0 @@
-﻿dyad.stats.mass\_ratio.moe2017
-==============================
-
-.. currentmodule:: dyad.stats.mass_ratio
-
-.. autodata:: moe2017
