+++ docs/source/api_reference/generated/dyad.stats.secondary_mass.rst
@@ -1,12 +1,0 @@
-﻿dyad.stats.secondary\_mass
-==========================
-
-.. automodule:: dyad.stats.secondary_mass
-
-   
-   .. rubric:: Classes
-
-   .. autosummary::
-   
-      moe2017_gen
-   

[…]
