### 2025-11-17 6ba8cbf
Missing .gitignore restored
