+++ dyad/stats/_tests/_period_test.py
@@ -1,10 +1,0 @@
-#!/usr/bin/env python
-
-import numpy as np
-import matplotlib as mpl
-import matplotlib.pyplot as plt
-
-from distributions import *
-
-mpl.style.use("sm")
-
