+++ dyad/stats/_tests/_mass_ratio_extended_test.py
@@ -1,662 +1,0 @@
-#!/usr/bin/env python3
-
-"""Plot mass-ratio distributions (Moe & Stefano, 2017)"""
-
-import time
-import numpy as np
-import matplotlib as mpl
-import matplotlib.pyplot as plt
-import plot
-
-from dyad.stats import mass_ratio
-
-mpl.style.use("sm")
-
-N_HIST = 100_000
-
-# Test utility functions: F_twin, c_twin
-primary_mass_boundary = (0.8, 1.2, 3.5, 6., 40., 0.08, 150.)
-log10_period_boundary = (
-    0.2, 1., 1.3, 2., 2.5, 3.4, 3.5, 4., 4.5, 5.5, 6., 6.5, 8.
-)
-
-n = 50
-primary_mass = np.array([1., 3.5, 7., 12.5, 25., 0.1, 100.])
-log10_period = np.hstack(
-    [
-        np.linspace(0.2 + 1.e-6, 1., n),
-        np.linspace(1., 1.3, n)[1:],
-        np.linspace(1.3, 2., n)[1:],
-        np.linspace(2., 2.5, n)[1:],
-        np.linspace(2.5, 3.4, n)[1:],
-        np.linspace(3.4, 3.5, n)[1:],
-        np.linspace(3.5, 4., n)[1:],
-        np.linspace(4., 4.5, n)[1:],
-        np.linspace(4.5, 5.5, n)[1:],
-        np.linspace(5.5, 6., n)[1:],
-        np.linspace(6., 6.5, n)[1:],
-        np.linspace(6.5, 8., n)[1:],
-    ]
-)
-
-log10_twin_excess_period = mass_ratio._moe2017_log10_excess_twin_period(
-    primary_mass
-)
-
-F_twin = mass_ratio._moe2017_twin_excess_fraction(
-    log10_period, primary_mass.reshape([-1, 1])
-)
-F_twin_closed_dots_x = np.array(
-    [0.2, 0.2, 0.2, 0.2, 0.2, 0.2, 0.2, 8., 8., 8., 8., 8., 8., 8.]
-)
-F_twin_closed_dots_x += 1.e-6
-F_twin_closed_dots_y = mass_ratio._moe2017_twin_excess_fraction(
-    F_twin_closed_dots_x, np.tile(primary_mass, 2)
-)
-
-delta = mass_ratio._moe2017_delta(log10_period, primary_mass.reshape([-1, 1]))
-c_twin = mass_ratio._moe2017_twin_excess_constant(
-    delta, log10_period, primary_mass.reshape([-1, 1])
-)
-c_twin_closed_dots_x = np.array(
-    [0.2, 0.2, 0.2, 0.2, 0.2, 0.2, 0.2, 8., 8., 8., 8., 8., 8., 8.]
-)
-c_twin_closed_dots_x += 1.e-6
-c_twin_closed_dots_y = mass_ratio._moe2017_twin_excess_constant(
-    np.hstack([np.tile(delta[0, 0], 7), np.tile(delta[0, -1], 7)]),
-    c_twin_closed_dots_x,
-    np.tile(primary_mass, 2)
-)
-
-fig, ax = plot.array(2, 1, sharex=True)
-ax[0][0].plot(log10_period, F_twin[0], color="red", ls="solid",
-           label=r"$M_{{1}} = {}$".format(primary_mass[0]))
-ax[0][0].plot(log10_period, F_twin[1], color="orange", ls="solid",
-           label=r"$M_{{1}} = {}$".format(primary_mass[1]))
-ax[0][0].plot(log10_period, F_twin[2], color="green", ls="solid",
-           label=r"$M_{{1}} = {}$".format(primary_mass[2]))
[…]
