### 2025-11-17 0acf2cd
Missing .gitignore restored
