+++ requirements.txt
@@ -1,9 +1,1 @@
-numpy>=1.21.6
-scipy>=1.11
-matplotlib
-numpydoc
-intersphinx_registry
-sphinx_copybutton
-sphinx_design
-sphinx-rtd-theme
-parameterized
+-e .
