+++ environment.yml
@@ -1,0 +1,13 @@
+# This file may be used to create an environment using Conda (see	https://docs.conda.io/projects/conda/en/latest/user-guide/tasks/manage-environments.html)
+# To use it run the following from the command line.
+# $ conda env create -f environment.yml
+# $ conda activate dyad
+#
+name: dyad
+channels:
+  - conda-forge
+dependencies:
+  - python>=3.10
+  - pip
+  - pip:
+    - -r requirements.txt
