### 2025-11-20 a6e27c7
Tutorial figures now use matplotlib defaults
