+++ docs/source/figures/tutorial_3.py
@@ -1,0 +1,143 @@
+#!/usr/bin/env python3
+
+"""
+Code for tutorial_3.rst.
+
+"""
+
+# import plot
+
+import numpy as np
+m_1 = np.array([1., 3.5, 7., 12, 28.])
+
+import dyad.stats as stats
+log_p = np.linspace(-1., 9., 500)
+f = stats.log_period.moe2017(m_1[:,None]).pdf(log_p)
+
+import matplotlib.pyplot as plt
+label = [
+    r"$M_{1} = 1\mathrm{M}_{\odot}$",
+    r"$M_{1} = 3.5\mathrm{M}_{\odot}$",
+    r"$M_{1} = 7\mathrm{M}_{\odot}$",
+    r"$M_{1} = 12\mathrm{M}_{\odot}$",
+    r"$M_{1} = 28\mathrm{M}_{\odot}$"
+]
+color = ["red", "orange", "green", "blue", "magenta"]
+# fig, ax = plot.plot()
+fig, ax = plt.subplots()
+for (f_i, label_i, color_i) in zip(f, label, color):
+    ax.plot(log_p, f_i, label=label_i, color=color_i)
+ax.legend(frameon=False)
+ax.set_ylim(0., 0.3)
+ax.set_xlabel(r"$\log_{10}(p/\mathrm{d})$")
+ax.set_ylabel(r"$f_{\log_{10}(P)}$")
+fig.savefig("./moe2017_pdf_logperiod.pdf", dpi=300)
+fig.savefig("./moe2017_pdf_logperiod.jpg", dpi=300)
+plt.show()
+
+p = np.logspace(-1., 9.)
+f = stats.period.moe2017(m_1[:,None]).pdf(p)
+
+q = np.linspace(0., 1.1, 500)
+f = stats.mass_ratio.moe2017(0.2, m_1[:,None]).pdf(q)
+
+# fig, ax = plot.plot()
+fig, ax = plt.subplots()
+for (f_i, label_i, color_i) in zip(f, label, color):
+    ax.plot(q, f_i, label=label_i, color=color_i)
+ax.legend(frameon=False)
+ax.set_xlabel(r"$q$")
+ax.set_ylabel(r"$f_Q$")
+fig.savefig("./moe2017_pdf_mass_ratio_short_period.pdf", dpi=300)
+fig.savefig("./moe2017_pdf_mass_ratio_short_period.jpg", dpi=300)
+plt.show()
+
+f = stats.mass_ratio.moe2017(8., m_1[:,None]).pdf(q)
+
+# fig, ax = plot.plot()
+fig, ax = plt.subplots()
+for (f_i, label_i, color_i) in zip(f, label, color):
+    ax.plot(q, f_i, label=label_i, color=color_i)
+ax.legend(frameon=False)
+ax.set_xlabel(r"$q$")
+ax.set_ylabel(r"$f_Q$")
+fig.savefig("./moe2017_pdf_mass_ratio_long_period.pdf", dpi=300)
+fig.savefig("./moe2017_pdf_mass_ratio_long_period.jpg", dpi=300)
+plt.show()
+
+e = np.linspace(-0.1, 1.1, 500)
+f = stats.eccentricity.moe2017(1., m_1[:,None]).pdf(e)
+
+# fig, ax = plot.plot()
+fig, ax = plt.subplots()
+for (f_i, label_i, color_i) in zip(f, label, color):
+    ax.plot(e, f_i, label=label_i, color=color_i)
+ax.legend(frameon=False)
+ax.set_ylim(0., 5.)
+ax.set_xlabel(r"$e$")
+ax.set_ylabel(r"$f_E$")
+fig.savefig("./moe2017_pdf_eccentricity_short_period.pdf", dpi=300)
[…]
