+++ docs/source/generated/generated/generated/dyad.stats.semimajor_axis.opik1924.rst
@@ -1,4 +1,4 @@
-﻿dyad.stats.semimajor\_axis.opik1924
+dyad.stats.semimajor\_axis.opik1924
 ===================================
 
 .. currentmodule:: dyad.stats.semimajor_axis
