+++ docs/source/api_reference/generated/dyad.stats.semimajor_axis.rst
@@ -4,3 +4,9 @@
 .. automodule:: dyad.stats.semimajor_axis
 
    
+   .. rubric:: Classes
+
+   .. autosummary::
+   
+      opik1924_gen
+   
