+++ docs/source/figures/tutorial_3.py
@@ -114,28 +114,35 @@ r_2 = binary.secondary.radius(theta)
 bins = np.logspace(-3., 6., 25)
 edge_r1, count_r1 = np.histogram(r_1, bins=bins)
 edge_r2, count_r2 = np.histogram(r_2, bins=bins)
+
 # fig, ax = plot.plot()
 fig, ax = plt.subplots()
-ax.stairs(edge_r1, count_r1, label="primary")
-ax.stairs(edge_r2, count_r2, label="primary")
+# ax.stairs(edge_r1, count_r1, label="primary")
+# ax.stairs(edge_r2, count_r2, label="primary")
+ax.hist(np.log10(r_1), bins="auto", alpha=0.2)
+ax.hist(np.log10(r_2), bins="auto", alpha=0.2)
 ax.legend(frameon=False)
-ax.set_xscale("log")
+# ax.set_xscale("log")
 ax.set_xlabel(r"$r/\mathrm{AU}$")
 ax.set_ylabel(r"$\nu$")
 fig.savefig("./moe2017_sample_radius.pdf", dpi=300)
 fig.savefig("./moe2017_sample_radius.jpg", dpi=300)
 plt.show()
+
 v_1 = binary.primary.speed(theta)
 v_2 = binary.secondary.speed(theta)
 bins = np.logspace(-3., 6., 25)
 edge_v1, count_v1 = np.histogram(r_1, bins=bins)
 edge_v2, count_v2 = np.histogram(r_2, bins=bins)
+
 # fig, ax = plot.plot()
 fig, ax = plt.subplots()
-ax.stairs(edge_v1, count_v1, label="primary")
-ax.stairs(edge_v2, count_v2, label="primary")
+# ax.stairs(edge_v1, count_v1, label="primary")
+# ax.stairs(edge_v2, count_v2, label="primary")
+ax.hist(np.log10(v_1), bins="auto", alpha=0.2)
+ax.hist(np.log10(v_2), bins="auto", alpha=0.2)
 ax.legend(frameon=False)
-ax.set_xscale("log")
+# ax.set_xscale("log")
 ax.set_xlabel(r"$v/\mathrm{km}~\mathrm{s}^{-1}$")
 ax.set_ylabel(r"$\nu$")
 fig.savefig("./moe2017_sample_speed.pdf", dpi=300)
