+++ docs/source/figures/tutorial_2.py
@@ -12,7 +12,7 @@ pdf = rv.pdf(x)
 sample = rv.rvs(size=10_000)
 
 fig, ax = plt.subplots()
-ax.hist(sample, bins=25, density=True, histtype="step")
+ax.hist(sample, bins='auto', density=True, histtype='stepfilled', alpha=0.2)
 ax.set_xticks([0., 0.5*np.pi, np.pi], [r"$0$", r"$\pi/2$", r"$\pi$"])
 ax.set_xlabel(r"$i$")
 ax.set_ylabel(r"$f_{I}$")
@@ -28,7 +28,7 @@ pdf = rv.pdf(x)
 sample = rv.rvs(size=10_000)
 
 fig, ax = plt.subplots()
-ax.hist(sample, bins=25, density=True, histtype="step")
+ax.hist(sample, bins='auto', density=True, histtype='stepfilled', alpha=0.2)
 ax.plot(x, pdf)
 ax.set_xticks([0., np.pi, 2.*np.pi], [r"$0$", r"$\pi$", r"$2\pi$"])
 ax.set_xlabel(r"$\theta$")
