+++ docs/source/tutorials/tutorial_2.rst
@@ -29,7 +29,7 @@ Now compute the PDF on the interval :math:`[0, \pi]`.
    >>> x = np.linspace(0., np.pi)
    >>> pdf = rv.pdf(x)
 
-And synthesize :math:`10~000` realizations.
+And generate a sample of size :math:`10\,000`.
 
 .. doctest:: python
 
@@ -41,7 +41,7 @@ Now plot our results.
 
    >>> import matplotlib.pyplot as plt
    >>> fig, ax = plt.subplots()
-   >>> ax.hist(sample, bins=25, density=True, histtype="step")
+   >>> ax.hist(sample, bins="auto", density=True, histtype="stepfilled", alpha=0.2)
    >>> ax.set_xticks([0., 0.5*np.pi, np.pi], [r"$0$", r"$\pi/2$", r"$\pi$"])
    >>> ax.set_xlabel(r"$i$")
    Text(0.5, 0, '$i$')
@@ -57,7 +57,7 @@ Now plot our results.
 
    The probability density function for inclination.
 
-Some of Dyad's random variables are dependent on other random variables. We must use their shape parameters to fully specify them. For example, the true anomaly of a body moving on an elliptical orbit in a gravitational central potential depends on that orbit's eccentricity, and so is a conditional random variable, :math:`\Theta|E = e`. Again, we may synthesize a sample and compute the PDF. Suppose that :math:`e = 0.5`.
+Some of Dyad's random variables are dependent on other random variables. We must use their shape parameters to fully specify them. For example, the true anomaly of a body moving on an elliptical orbit in a gravitational central potential depends on that orbit's eccentricity, so we are interested in the conditional distribution of :math:`\Theta` given :math:`E`. Again, we may synthesize a sample and compute the PDF. Suppose that :math:`e = 0.5`.
 
 .. doctest:: python
 
@@ -66,12 +66,12 @@ Some of Dyad's random variables are dependent on other random variables. We must
    >>> pdf = rv.pdf(x)
    >>> sample = rv.rvs(size=10_000)
 
-Plot the results.
+Then plot the results.
 
 .. doctest:: python
 
    >>> fig, ax = plt.subplots()
-   >>> ax.hist(sample, bins=25, density=True, histtype="step")
+   >>> ax.hist(sample, bins="auto", density=True, histtype="stepfilled", alpha=0.2)
    >>> ax.plot(x, pdf)
    >>> ax.set_xticks([0., np.pi, 2.*np.pi], [r"$0$", r"$\pi$", r"$2\pi$"])
    >>> ax.set_xlabel(r"$\theta$")
@@ -118,14 +118,29 @@ Its methods are available in the same way as before.
 A complete population
 =====================
 
-Let us synthesize the complete orbital properties of a population of binary stars: mass, mass ratio, and orbital elements. We will use the distributions of Duquennoy and Mayor [DM91]_. Assume that the primary stars of our populations have masses of :math:`0.8~\mathrm{M}_{\odot}` and sample the mass ratio and the period.
+Let us synthesize the complete orbital properties of a population of binary stars, namely the mass, mass ratio, and orbital elements.
+These have been determined for binary systems with sun-like primary stars in the solar neighbourhood by Duquennoy and Mayor [DM91]_.
[…]
