+++ docs/source/getting_started/what_is_dyad.rst
@@ -30,6 +30,13 @@ The basics
    :start-after: `\mathrm{M}_{\odot}`.
    :end-before: Classes
 
+Units
+=====
+
+Dyad requires mass to be specified in units of solar mass, :math:`\mathrm{M}_{\odot}`, distance in units of astronomical unit, :math:`\mathrm{AU}`, and time in units of days, :math:`\mathrm{d}`.
+In this system the gravitational constant is :math:`\mathrm{G} = 2.959 \times 10^{-4} \mathrm{AU}^{3}\mathrm{M}_{\odot}\mathrm{d}^{-2}`.
+However, speed is in units of kilometers per second, :math:`\mathrm{km}~\mathrm{s}^{-1}`.
+
 References
 ==========
 

[…]
