### 2025-11-21 #9011 @AmeryGration
Hi @AmedeoRom. Thank you so much for your comments. I found them very helpful. You certainly weren't nit picking.

>     * I agree with [@stegmaja](https://github.com/stegmaja) that a YAML file is recommended. It saved a lot of complications from my side

I was a little confused by this issue at first, because I'm not a user of Conda. Having done some digging, I believe this is a file `environment.yml` that Conda uses to specify a virtual environment. Am I right? Assuming that I am, I've added a file accordingly.

>     * In the code description, one can see what the orbit.state array contains only at the very end, when binary populations are involved, despite the tutorial already goes through it in the Keplerian systems. I would recommend adding even just a line introducing what is actually displayed

Looking at the documentation with fresh eyes I see exactly what you mean. I've modified Tutorial 2 accordingly and think it is much better for your suggestion. I've also added a note on units at the end of `What is Dyad?'

>     * I would advise to revise how the code snapshots are put in their respective cells. Firstly, clicking on the proper "copy to clipboard" button copies also the ">>>" part, which brings a syntax error. Also, by simply doing instead ctrl-C + ctrl-V of everything, instead, sometimes there are some terminal outputs within the code cells that of course bring errors if everything is ctrl-Ved . Of course this just a minor issue that only slightly slows the tutorial down. I only wanted to highlight it directly

I hadn't picked up on this. I've modified my Sphinx `conf.py` and the code cells should now behave as you suggest.

>     * Typo in the synthetic population section: "These probability DISTRIBITIONS [...]"

Changed.

>     * I leave attached the plot for the probability density function for true anomaly by following to the letter the tutorial code. Despite the data is still the same from the website, the plot produced is considerably different. I would recommend to either specify what plot style has been used, or to conform to the default pyplot version in the example plots so that users will not face any contradiction with plotting

I think you are right and I have now changed all plots to use matplotlib's defaults.
