### 2025-11-21 362e11a
Circularization period for eccentricity.duquennoy1991 changed to 11.6 d from 11 d
