### 2025-11-21 836b9bf
Workflow template modified to allow testing with GitHub CI
