### 2025-11-21 485a628
Workflow template modified to allow testing with GitHub CI
