### 2025-11-21 049a270
Workflow template modified to allow testing with GitHub CI
