### 2025-11-21 aa4ec46
Workflow template modified to allow testing with GitHub CI
