### 2025-11-21 bed5a72
Workflow template modified to allow testing with GitHub CI
