### 2025-11-21 5a7f360
Workflow template modified to allow testing with GitHub CI
