### 2025-11-21 7ebbf3f
Workflow template modified to allow testing with GitHub CI
