### 2025-11-21 d58f008
Workflow template modified to allow testing with GitHub CI
