+++ setup.py
@@ -56,6 +56,7 @@ setup(
         "sphinx_design",
         "sphinx-rtd-theme",
         "parameterized",
+        "unittest",
     ],
     include_package_data=True,
     python_requires=">=3.10",
