+++ .github/workflows/python-app.yml
@@ -29,5 +29,5 @@ jobs:
         pip install -r requirements.txt
     - name: Test with unittest
       run: |
-        python -m pip install unittest
+        pip install unittest
         python -m unittest discover -s ./tests/unittests/
