### 2025-11-23 2c52853
Clean up spurious dependency on unittest
