+++ requirements.txt
@@ -5,7 +5,6 @@ charset-normalizer==3.4.4
 contourpy==1.3.2
 cycler==0.12.1
 docutils==0.21.2
--e git+https://github.com/AmeryGration/dyad.git@176ae1471a17e57fcad635bfe0c3793154a6b9e4#egg=dyad
 fonttools==4.60.1
 idna==3.11
 imagesize==1.4.1
