### 2025-11-23 a0eff10
Clean up spurious dependency on unittest
