+++ .github/workflows/python-app.yml
@@ -29,5 +29,4 @@ jobs:
         pip install -r requirements.txt
     - name: Test with unittest
       run: |
-        pip install unittest
         python -m unittest discover -s ./tests/unittests/
