### 2025-11-23 ff0245c
CI for unit tests
