### 2025-11-24 b11bd7f
Tutorial doctests revised and working
