+++ docs/source/tutorials/tutorial_2.rst
@@ -16,6 +16,11 @@ generates random variates (i.e. which generates a sample).
 
 For example, consider the random variable for inclination, :math:`I`, where :math:`\cos(I) \sim U(-1, 1)`. This is implemented by Dyad using the class :class:`dyad.stats.inclination`. Instantiate this class as follows.
 
+.. testsetup::
+
+   import numpy as np
+   np.random.seed(0)
+
 .. doctest:: python
 
    >>> import dyad
@@ -42,12 +47,27 @@ Now plot our results.
    >>> import matplotlib.pyplot as plt
    >>> fig, ax = plt.subplots()
    >>> ax.hist(sample, bins="auto", density=True, histtype="stepfilled", alpha=0.2)
+   (array([0.04519806, 0.08320552, 0.12943081, 0.18798285, 0.24242597,
+	  0.22701754, 0.32768595, 0.36055726, 0.36774787, 0.39445581,
+	  0.42938159, 0.46430736, 0.43965387, 0.49615145, 0.49615145,
+	  0.53415891, 0.49306976, 0.49512422, 0.47868856, 0.43451773,
+	  0.44376279, 0.4406811 , 0.41089147, 0.34206715, 0.32357703,
+	  0.29995077, 0.27427006, 0.21160911, 0.18490116, 0.11607684,
+	  0.06882432, 0.0287624 ]), array([0.01702368, 0.11437298, 0.21172229, 0.3090716 , 0.40642091,
+	  0.50377021, 0.60111952, 0.69846883, 0.79581814, 0.89316744,
+	  0.99051675, 1.08786606, 1.18521536, 1.28256467, 1.37991398,
+	  1.47726329, 1.57461259, 1.6719619 , 1.76931121, 1.86666052,
+	  1.96400982, 2.06135913, 2.15870844, 2.25605775, 2.35340705,
+	  2.45075636, 2.54810567, 2.64545498, 2.74280428, 2.84015359,
+	  2.9375029 , 3.03485221, 3.13220151]), [<matplotlib.patches.Polygon object at 0x...>])
    >>> ax.set_xticks([0., 0.5*np.pi, np.pi], [r"$0$", r"$\pi/2$", r"$\pi$"])
+   [<matplotlib.axis.XTick object at 0x...>, <matplotlib.axis.XTick object at 0x...>, <matplotlib.axis.XTick object at 0x...>]
    >>> ax.set_xlabel(r"$i$")
    Text(0.5, 0, '$i$')
    >>> ax.set_ylabel(r"$f_{I}$")
    Text(0, 0.5, '$f_{I}$')
    >>> ax.plot(x, pdf)
+   [<matplotlib.lines.Line2D object at 0x...>]
    >>> plt.show()
 
 .. _inclination:
@@ -72,8 +92,32 @@ Then plot the results.
 
    >>> fig, ax = plt.subplots()
    >>> ax.hist(sample, bins="auto", density=True, histtype="stepfilled", alpha=0.2)
+   (array([0.04892705, 0.05045602, 0.04969153, 0.05580741, 0.06192329,
+	  0.06574572, 0.06192329, 0.07491954, 0.06956814, 0.07950645,
+	  0.08103542, 0.11620173, 0.11314379, 0.12690453, 0.13990077,
+	  0.17353812, 0.19953061, 0.21329134, 0.2614539 , 0.31879028,
+	  0.34478278, 0.36771733, 0.40441261, 0.39829673, 0.41358644,
[…]
