### 2025-11-24 795f494
Tutorial on ZAMS stars revised
