### 2025-11-24 689b24c
Tutorial on ZAMS stars revised
