### 2025-11-24 0434d24
CI tests now include doctests for tutorials
