+++ docs/source/build/doctest/output.txt
@@ -1,4 +1,4 @@
-Results of doctest builder run on 2025-11-24 14:02:07
+Results of doctest builder run on 2025-11-24 14:15:17
 =====================================================
 
 Document: tutorials/tutorial_3
