+++ .github/workflows/python-app.yml
@@ -30,3 +30,5 @@ jobs:
     - name: Test with unittest
       run: |
         python -m unittest discover -s ./tests/unit_tests/
+	cd docs/source/
+	sphinx-build -b doctest . build/doctest ./tutorials/tutorial_{1,2,3}.rst 
