### 2025-11-24 d88e9d8
CI tests now include doctests for tutorials
