### 2025-11-24 1884f9b
CI doctests removed
