### 2025-11-25 92214a1
Dyad now uses the astronomical unit system. All code and documentation updated
