+++ tests/unit_tests/data_test_core.py
@@ -195,13 +195,13 @@ radius = [
 ]
 
 speed = [
-    [-0.5*np.pi, 38.179929773012695],
-    [ 0.       , 52.40890006011136 ],
-    [ 0.5*np.pi, 38.179929773012695],
-    [     np.pi, 12.98927516967835 ],
-    [ 1.5*np.pi, 38.179929773012695],
-    [ 2.0*np.pi, 52.40890006011136 ],
-    [ 2.5*np.pi, 38.179929773012695],
+    [-0.5*np.pi, 0.022050754579278225],
+    [ 0.       , 0.030268672568704025],
+    [ 0.5*np.pi, 0.022050754579278225],
+    [     np.pi, 0.007501934147918388],
+    [ 1.5*np.pi, 0.02205075457927822 ],
+    [ 2.0*np.pi, 0.030268672568704025],
+    [ 2.5*np.pi, 0.02205075457927823 ],
 ]
 
 position = [
@@ -215,13 +215,13 @@ position = [
 ]
 
 velocity = [
-    [-0.5*np.pi, [  8.976256597882106, 31.825371778374805, 19.0861092454535 ]],
-    [ 0.       , [-35.82907143316502 , 25.42714318605686 , 28.57325384242511]],
-    [ 0.5*np.pi, [-35.92527774596863 ,-12.70021469093256 ,  2.40541122293335]],
-    [     np.pi, [  8.880050285078491, -6.301986098614625, -7.08173337403831]],
-    [ 1.5*np.pi, [  8.976256597882113, 31.825371778374805, 19.08610924545345]],
-    [ 2.0*np.pi, [-35.829071433165005, 25.427143186056874, 28.57325384242511]],
-    [ 2.5*np.pi, [-35.925277745968614,-12.70021469093257 ,  2.40541122293334]],
+    [-0.5*np.pi, [ 0.00518422,  0.01838069,  0.01102315]],
+    [ 0.       , [-0.02069302,  0.0146854 ,  0.01650243]],
+    [ 0.5*np.pi, [-0.02074858, -0.00733499,  0.00138924]],
+    [     np.pi, [ 0.00512866, -0.0036397 , -0.00409004]],
+    [ 1.5*np.pi, [ 0.00518422,  0.01838069,  0.01102315]],
+    [ 2.0*np.pi, [-0.02069302,  0.0146854 ,  0.01650243]],
+    [ 2.5*np.pi, [-0.02074858, -0.00733499,  0.00138924]]
 ]
 
 state = [
