+++ docs/source/build/doctest/output.txt
@@ -1,4 +1,4 @@
-Results of doctest builder run on 2025-11-24 14:15:17
+Results of doctest builder run on 2025-11-25 14:51:24
 =====================================================
 
 Document: tutorials/tutorial_3
