+++ docs/source/generated/dyad.rst
@@ -12,5 +12,4 @@ Subpackages
 .. toctree::
    :maxdepth: 1
 
-   dyad.constants
    dyad.stats
