### 2025-11-25 b2a7bef
Dyad now uses the astronomical unit system. All code and documentation updated
