+++ docs/source/figures/tutorial_3.py
@@ -143,7 +143,7 @@ ax.hist(np.log10(v_1), bins="auto", alpha=0.2)
 ax.hist(np.log10(v_2), bins="auto", alpha=0.2)
 ax.legend(frameon=False)
 # ax.set_xscale("log")
-ax.set_xlabel(r"$\log_{10}(v/\mathrm{km}~\mathrm{s}^{-1})$")
+ax.set_xlabel(r"$\log_{10}(v/\mathrm{AU}~\mathrm{d}^{-1})$")
 ax.set_ylabel(r"$\nu$")
 fig.savefig("./moe2017_sample_speed.pdf", dpi=300)
 fig.savefig("./moe2017_sample_speed.jpg", dpi=300)
