### 2025-11-25 a56e6b7
Typos in tutorials corrected
