### 2025-11-25 c40b3fb
Typos in tutorials corrected
