### 2025-11-25 c777cf2
Typos in tutorials corrected
