### 2025-11-26 b1947a6
Conversion of Delaunay elements to/from orbital elements in _core.py
