+++ dyad/_core.py
@@ -18,6 +18,10 @@ __all__ = [
     "secondary_semimajor_axis_from_semimajor_axis",
     "primary_semimajor_axis_from_secondary_semimajor_axis",
     "secondary_semimajor_axis_from_primary_semimajor_axis",
+    "delaunay_elements_from_orbital_elements",
+    "orbital_elements_from_delaunay_elements",
+    "modified_delaunay_elements_from_orbital_elements",
+    "orbital_elements_from_modified_delaunay_elements",
     "Orbit",
     "TwoBody",
 ]
@@ -637,6 +641,404 @@ def secondary_semimajor_axis_from_primary_semimajor_axis(a, q):
     q = _check_semimajor_axis(q)
     res = a/q
     res = res[()]
+
+    return res
+
+def delaunay_elements_from_orbital_elements(a, e, Omega, i, omega, theta, m):
+    """Return the Delaunay elements given the orbital elements
+
+    Consider a body moving on an elliptical orbit in a gravitational
+    central potential generated by a central mass of :math:`m`. The
+    Delaunay elements are
+
+    .. math::
+
+       J_{1} = \sqrt{\mathrm{G}ma(1 - e^{2})}\cos(i)
+       J_{2} = \sqrt{\mathrm{G}ma(1 - e^{2})}
+       J_{3} = \sqrt{\mathrm{G}ma}
+       \Theta_{1} = \Omega
+       \Theta_{2} = \omega
+       \Theta_{3} = \mu(\theta)
+
+    where :math:`a` is the semimajor axis, :math:`e` is the
+    eccentricity, :math:`\Omega` is the longitude of the ascending
+    node, :math:`i` is the inclindation, :math:`\omega` is the
+    argument of pericentre, and :math:`\mu(\theta)` is the mean anomaly.
+
+    Parameters
+    ----------
+
+    a : array-like
+
+        Semimajor axis
+
+    e : array-like
+
+        Eccentricity
+
+    Omega : array-like
+    
+        Longitude of the ascending node
+
+    i : array-like
+
+        Incination
+
+    omega : array-like
+
+        Argument of pericentre
+
+    theta : array-like
+
+        True anomaly
+
+    m : array-like
+
+        Central mass
+
+    Returns
+    -------
+
+    res : list
+
+        Delaunay elements ``(J_1, J_2, J_3, Theta_1, Theta_2, Theta_3)``
+
+    Examples
+    --------
+
+    Scalar parameters.
+    
+    >>> dyad.delaunay_elements_from_orbital_elements(1., 0., 0., 0., 0.,
+    ...     0., 1.)
+    (0.017202098944262438, 0.017202098944262438, 0.017202098944262438, 0.0, 0.0, 0.0)
+
+    Array-like parameters defining multiple orbits.
+
+    >>> a, e, Omega, i, omega, theta, m  = [1., 1.], [0., 0.], [0., 0.],
+    ...     [0., 0.], [0., 0.], [0., 0.], [1., 1.]
+    >>> dyad.delaunay_elements_from_orbital_elements(a, e, Omega, i,
+    ...     omega, theta, m)
+
+    """
+    a = np.asarray(a)[()]
+    e = np.asarray(e)[()]
+    theta = np.asarray(theta)[()]
+    Omega = np.asarray(Omega)[()]
+    i = np.asarray(i)[()]
+    omega = np.asarray(omega)[()]
+    m = np.asarray(m)[()]
+    
+    a = _check_semimajor_axis(a)
+    e = _check_eccentricity(e)
+    Omega = _check_angle(Omega)
[…]
