+++ tests/unit_tests/data_test_core.py
@@ -234,72 +234,39 @@ state = [
     [ 2.5*np.pi, position[6][1] + velocity[6][1]],
 ]
     
-# potential = [
-#     [0.09225864169478659, -0.000570718801363782],
-# ]
-    
-# kinetic_energy = [
-#     [0.09225864169478659, 4108996849.478827],
-# ]
-
 delaunay_elements_from_orbital_elements = [
     # [
     #     [a, e, Omega, i, omega, theta, m],
     #     [J_1, J_2, J_3, Theta_1, Theta_2, Theta_3]
     # ]
     [
-        [1., 0., 0., 0., 0., 0., 1.],
-        [ 0.017202098944262, 0.017202098944262, 0.017202098944262,
-          0.               , 0.               , 0.              ]
+        [1., 0.5, 0., 1., 0., 0., 1.],
+        [0.00804912911741,  0.014897454684145, 0.017202098944262,
+         0.              ,  0.               , 0.               ],        
     ],
     [
-        [1., 0., 2.*np.pi, 0., 0., 0., 1.],
-        [ 0.017202098944262, 0.017202098944262, 0.017202098944262,
-          6.283185307179586, 0.               , 0.              ]
+        [1., 0.5, 2.*np.pi, 1., 0., 0., 1.],
+        [0.00804912911741 , 0.014897454684145, 0.017202098944262,
+         6.283185307179586, 0.               , 0.               ],        
     ],
     [
-        [1., 0., 0., np.pi, 0., 0., 1.],
-        [-0.017202098944262, 0.017202098944262, 0.017202098944262,
-         0.               , 0.               , 0.              ]
+        [1., 0.5, 0., 1., 2.*np.pi, 0., 1.],
+        [0.00804912911741 , 0.014897454684145, 0.017202098944262,
+         0.               , 6.283185307179586, 0.               ]
     ],
     [
-        [1., 0., 0., 0., 2.*np.pi, 0., 1.],
-        [ 0.017202098944262, 0.017202098944262, 0.017202098944262,
-          0.               , 6.283185307179586, 0.              ]
+        [1., 0.5, 0., 1., 0., 2.*np.pi, 1.],
+        [0.00804912911741 , 0.014897454684145, 0.017202098944262,
+         0.               , 0.               , 6.283185307179586]
     ],
     [
-        [1., 0., 0., 0., 0., 2.*np.pi, 1.],
-        [ 0.017202098944262, 0.017202098944262, 0.017202098944262,
-          0.               , 0.               , 6.283185307179586]
+        [1., 0.5, 2.*np.pi, 1., 2.*np.pi, 0., 1.],
+        [0.00804912911741 , 0.014897454684145, 0.017202098944262,
+         6.283185307179586, 6.283185307179586, 0.               ]
     ],
     [
-        [1., 0., 2.*np.pi, np.pi, 0., 0., 1.],
-        [-0.017202098944262, 0.017202098944262, 0.017202098944262,
-         6.283185307179586, 0.               , 0.              ]
-    ],
-    [
-        [1., 0., 2.*np.pi, 0., 2.*np.pi, 0., 1.],
-        [ 0.017202098944262, 0.017202098944262, 0.017202098944262,
-          6.283185307179586, 6.283185307179586, 0.              ]
-    ],
-    [
-        [1., 0., 2.*np.pi, 0., 0., 2.*np.pi, 1.],
-        [ 0.017202098944262, 0.017202098944262, 0.017202098944262,
-          6.283185307179586, 0.               , 6.283185307179586]
-    ],
-    [
[…]
