+++ dyad/_core.py
@@ -645,7 +645,7 @@ def secondary_semimajor_axis_from_primary_semimajor_axis(a, q):
     return res
 
 def delaunay_elements_from_orbital_elements(a, e, Omega, i, omega, theta, m):
-    """Return the Delaunay elements given the orbital elements
+    r"""Return the Delaunay elements given the orbital elements
 
     Consider a body moving on an elliptical orbit in a gravitational
     central potential generated by a central mass of :math:`m`. The
@@ -653,17 +653,19 @@ def delaunay_elements_from_orbital_elements(a, e, Omega, i, omega, theta, m):
 
     .. math::
 
-       J_{1} = \sqrt{\mathrm{G}ma(1 - e^{2})}\cos(i)
-       J_{2} = \sqrt{\mathrm{G}ma(1 - e^{2})}
-       J_{3} = \sqrt{\mathrm{G}ma}
-       \Theta_{1} = \Omega
-       \Theta_{2} = \omega
-       \Theta_{3} = \mu(\theta)
+       J_{1} &= \sqrt{\mathrm{G}ma(1 - e^{2})}\cos(i)\\
+       J_{2} &= \sqrt{\mathrm{G}ma(1 - e^{2})}\\
+       J_{3} &= \sqrt{\mathrm{G}ma}\\
+       \Theta_{1} &= \Omega\\
+       \Theta_{2} &= \omega\\
+       \Theta_{3} &= \mu(\theta)
 
-    where :math:`a` is the semimajor axis, :math:`e` is the
-    eccentricity, :math:`\Omega` is the longitude of the ascending
-    node, :math:`i` is the inclindation, :math:`\omega` is the
-    argument of pericentre, and :math:`\mu(\theta)` is the mean anomaly.
+    where :math:`a/\text{AU} \in (0, \infty)` is the semimajor axis,
+    :math:`e = (0, 1)` is the eccentricity, :math:`\Omega \in [0,
+    2\pi)` is the longitude of the ascending node, :math:`i \in (0,
+    \pi)` is the inclination, :math:`\omega \in [0, 2\pi)` is the
+    argument of pericentre, and :math:`\mu(\theta) \in [0, 2\pi)` is
+    the mean anomaly.
 
     Parameters
     ----------
@@ -677,7 +679,7 @@ def delaunay_elements_from_orbital_elements(a, e, Omega, i, omega, theta, m):
         Eccentricity
 
     Omega : array-like
-    
+
         Longitude of the ascending node
 
     i : array-like
@@ -703,14 +705,20 @@ def delaunay_elements_from_orbital_elements(a, e, Omega, i, omega, theta, m):
 
         Delaunay elements ``(J_1, J_2, J_3, Theta_1, Theta_2, Theta_3)``
 
+    Warnings
+    --------
+
+    Note that :math:`e \neq 0` and :math:`i \neq 0`.
+
     Examples
     --------
 
     Scalar parameters.
-    
+
     >>> dyad.delaunay_elements_from_orbital_elements(1., 0., 0., 0., 0.,
     ...     0., 1.)
-    (0.017202098944262438, 0.017202098944262438, 0.017202098944262438, 0.0, 0.0, 0.0)
+    array([0.017202098944262, 0.017202098944262, 0.017202098944262,
+           0.               , 0.               , 0.               ])
 
     Array-like parameters defining multiple orbits.
 
@@ -718,6 +726,10 @@ def delaunay_elements_from_orbital_elements(a, e, Omega, i, omega, theta, m):
     ...     [0., 0.], [0., 0.], [0., 0.], [1., 1.]
[…]
