### 2025-11-27 a76c05c
Various attributes and methods added to Orbit class
