+++ dyad/_core.py
@@ -758,7 +758,7 @@ def delaunay_elements_from_orbital_elements(a, e, Omega, i, omega, theta, m):
     Theta_1 = Omega
     Theta_2 = omega
     Theta_3 = dyad.mean_anomaly_from_true_anomaly(theta, e)
-    res = np.array([J_1, J_2, J_3, Theta_1, Theta_2, Theta_3]).T
+    res = J_1, J_2, J_3, Theta_1, Theta_2, Theta_3
 
     return res
 
@@ -877,7 +877,7 @@ def orbital_elements_from_delaunay_elements(
     Omega = Theta_1
     i = np.arccos(J_1/J_2)
     omega = Theta_2
-    res = np.array([a, e, Omega, i, omega, theta]).T
+    res = a, e, Omega, i, omega, theta
 
     return res
 
@@ -991,9 +991,7 @@ def modified_delaunay_elements_from_orbital_elements(
         + omega
         + dyad.mean_anomaly_from_true_anomaly(theta, e)
     )
-    res = np.array(
-        [J_pi, J_Omega, J_lambda, Theta_pi, Theta_Omega, Theta_lambda]
-    ).T
+    res = J_pi, J_Omega, J_lambda, Theta_pi, Theta_Omega, Theta_lambda
 
     return res
 
@@ -1115,7 +1113,7 @@ def orbital_elements_from_modified_delaunay_elements(
     Omega = -Theta_Omega
     i = np.arccos(1 - J_Omega/(J_lambda - J_pi))
     omega = -Theta_pi + Theta_Omega
-    res = np.array([a, e, Omega, i, omega, theta]).T
+    res = a, e, Omega, i, omega, theta
 
     return res
 
@@ -1194,6 +1192,37 @@ class Orbit:
         self._argument_of_pericentre = omega
 
     @property
+    def orbital_elements(self):
+        """Get the orbital elements of the orbit"""
+        res = (
+            self.semimajor_axis,
+            self.eccentricity,
+            self.longitude_of_ascending_node,
+            self.inclination,
+            self.argument_of_pericentre
+        )
+
+        return res
+
+    @property
+    def delaunay_elements(self):
+        """Get the Delaunay elements of the orbit"""
+        res = delaunay_elements_from_orbital_elements(
+            *self.orbital_elements, theta=0., m=self.mass
+        )[:-1]
+
+        return res
+
+    @property
+    def modified_delaunay_elements(self):
+        """Get the modified Delaunay elements of the orbit"""
+        res = modified_delaunay_elements_from_orbital_elements(
+            *self.orbital_elements, theta=0., m=self.mass
+        )[:-1]
+
+        return res
+
+    @property
     def mass(self):
         """Get the central mass"""
         return self._mass
@@ -1257,6 +1286,123 @@ class Orbit:
                 /(_constants.GRAV_CONST*self.mass)
             )
         )
+
+    @property
+    def eccentricity_vector(self):
+        """Get the eccentricity vector of the orbit"""
+        res = (
+            np.cross(self._velocity(0.), self.angular_momentum)
+            /(_constants.GRAV_CONST*self.mass)
+            - self._position(0.)/self.radius(0.)
+        )
+
+        return res
+
+    @property
+    def lrl_vector(self):
+        """Get the Laplace-Runge-Lenz vector of the orbit"""
+        res = (
+            np.cross(self._velocity(0.), self.angular_momentum)
+            - (
+                _constants.GRAV_CONST*self.mass*self._position(0.)
+                /self.radius(0.)
+            )
+        )
+
+        return res
+
+    @property
+    def energy(self):
+        """Get the body's specific total energy"""
+        res = self.potential(0.) + self.kinetic_energy(0.)
+
[…]
