+++ docs/source/generated/generated/dyad.Orbit.rst
@@ -15,7 +15,9 @@
    
       ~Orbit.__init__
       ~Orbit.eccentric_anomaly
+      ~Orbit.kinetic_energy
       ~Orbit.mean_anomaly
+      ~Orbit.potential
       ~Orbit.radius
       ~Orbit.speed
       ~Orbit.state
@@ -28,13 +30,21 @@
 
    .. autosummary::
    
+      ~Orbit.angular_momentum
+      ~Orbit.angular_momentum_magnitude
       ~Orbit.apoapsis
       ~Orbit.area
       ~Orbit.argument_of_pericentre
+      ~Orbit.delaunay_elements
       ~Orbit.eccentricity
+      ~Orbit.eccentricity_vector
+      ~Orbit.energy
       ~Orbit.inclination
       ~Orbit.longitude_of_ascending_node
+      ~Orbit.lrl_vector
       ~Orbit.mass
+      ~Orbit.modified_delaunay_elements
+      ~Orbit.orbital_elements
       ~Orbit.periapsis
       ~Orbit.period
       ~Orbit.semilatus_rectum
