### 2025-11-28 105d849
Various attributes and methods added to classes Orbit and TwoBody. Period random variable removed from stats module
