+++ dyad/_core.py
@@ -1184,48 +1184,73 @@ class Orbit:
         i = _check_angle(i)
         omega = _check_angle(omega)
 
-        self._mass = m
+        self._central_mass = m
         self._semimajor_axis = a
         self._eccentricity = e
         self._longitude_of_ascending_node = Omega
         self._inclination = i
         self._argument_of_pericentre = omega
 
-    @property
-    def orbital_elements(self):
+    # @property
+    def orbital_elements(self, theta=None):
         """Get the orbital elements of the orbit"""
-        res = (
-            self.semimajor_axis,
-            self.eccentricity,
-            self.longitude_of_ascending_node,
-            self.inclination,
-            self.argument_of_pericentre
-        )
+        if theta == None:
+            res = (
+                self.semimajor_axis,
+                self.eccentricity,
+                self.longitude_of_ascending_node,
+                self.inclination,
+                self.argument_of_pericentre,
+            )
+        else:
+            res = (
+                self.semimajor_axis,
+                self.eccentricity,
+                self.longitude_of_ascending_node,
+                self.inclination,
+                self.argument_of_pericentre,
+                theta
+            )
+        res = np.array(list(np.broadcast(*res))).squeeze()
 
         return res
 
-    @property
-    def delaunay_elements(self):
+    # @property
+    def delaunay_elements(self, theta=None):
         """Get the Delaunay elements of the orbit"""
-        res = delaunay_elements_from_orbital_elements(
-            *self.orbital_elements, theta=0., m=self.mass
-        )[:-1]
+        if theta:
+            res = delaunay_elements_from_orbital_elements(
+                *self.orbital_elements(theta=theta).T, m=self.central_mass
+            )
+        else:
+            res = delaunay_elements_from_orbital_elements(
+                *self.orbital_elements().T, theta=0., m=self.central_mass
+            )
+            res = res[:-1]
+        res = np.array(list(np.broadcast(*res))).squeeze()
 
         return res
 
-    @property
-    def modified_delaunay_elements(self):
+    # @property
+    def modified_delaunay_elements(self, theta=None):
         """Get the modified Delaunay elements of the orbit"""
-        res = modified_delaunay_elements_from_orbital_elements(
-            *self.orbital_elements, theta=0., m=self.mass
-        )[:-1]
+        if theta:
+            res = modified_delaunay_elements_from_orbital_elements(
+                *self.orbital_elements(theta=theta).T, m=self.central_mass
+            )
+        else:
+            res = modified_delaunay_elements_from_orbital_elements(
+                *self.orbital_elements().T, theta=0., m=self.central_mass
+            )
+            res = res[:-1]
+        res = np.array(list(np.broadcast(*res))).squeeze()
 
         return res
 
     @property
-    def mass(self):
+    def central_mass(self):
         """Get the central mass"""
-        return self._mass
+        return self._central_mass
 
     @property
     def semimajor_axis(self):
@@ -1283,7 +1308,7 @@ class Orbit:
         return (
             2.*np.pi*np.sqrt(
                 self.semimajor_axis**3.
-                /(_constants.GRAV_CONST*self.mass)
+                /(_constants.GRAV_CONST*self.central_mass)
             )
         )
 
@@ -1292,7 +1317,7 @@ class Orbit:
         """Get the eccentricity vector of the orbit"""
[…]
