+++ dyad/stats/__init__.py
@@ -41,7 +41,6 @@ of the condition.
    log_period
    mass
    mass_ratio
-   period
    secondary_mass
    semimajor_axis
 
@@ -71,7 +70,7 @@ from . import eccentricity
 from . import mass
 from . import mass_ratio
 from . import log_period
-from . import period
+# from . import period
 from . import secondary_mass
 from . import semimajor_axis
 
