+++ dyad/stats/period.py
@@ -1,329 +1,0 @@
-"""
-=================================
-Period (:mod:`dyad.stats.period`)
-=================================
-
-.. currentmodule:: dyad.stats.period
-
-This module contains probability distributions for the orbital periods
-of a population of binary stars. In its documentation the random
-variable is denoted :math:`P` and a realization of that random
-variable is denoted :math:`p`.
-
-Probability distributions
-=========================
-
-.. autosummary::
-   :toctree: generated/
-
-   trunclognorm
-   duquennoy1991
-   moe2017
-
-"""
-
-__all__ = [
-    "trunclognorm",
-    "duquennoy1991",
-    "moe2017",
-]
-
-import numpy as np
-import scipy as sp
-import dyad
-
-from importlib.resources import files
-from scipy._lib._util import _lazywhere
-from scipy.interpolate import RegularGridInterpolator
-from scipy.interpolate import LinearNDInterpolator
-from dyad.stats import mass_ratio
-from . import _distn_infrastructure
-
-def _lognorm_logpdf(x, s):
-    # Replica of ~scipy.stats._continuous_distns.cd._lognorm_logpdf~,
-    # which is called by ~scipy.stats.lognorm~.
-    return _lazywhere(
-        x != 0,
-        (x, s),
-        lambda x, s: -np.log(x)**2./(2.*s**2.) - np.log(s*x*np.sqrt(2*np.pi)),
-        -np.inf
-    )
-
-
-class trunclognorm_gen(_distn_infrastructure.rv_continuous):
-    r"""A truncated lognormal random variable
-
-    %(before_notes)s
-
-    Notes
-    -----
-    The probability density function for `trunclognorm` is:
-
-    .. math::
-       f_{P}(p, a, b, s) =
-       \dfrac{\exp\left(-\dfrac{\log(p)^{2}}{2s^{2}}\right)}
-       {\sqrt{2\pi}sx\left(\operatorname{erf}\left(\dfrac{\log(b)}{s}\right)
-       - \operatorname{erf}\left(\dfrac{\log(a)}{s}\right)\right)}
-
-    where :math:`p \in (a, b)`, :math:`a, b \in (0, \infty)`, and
-    :math:`s \in (0, \infty)`.
-
-    The probability density function `trunclognorm` takes ``a``,
-    ``b``, and ``s`` as shape parameters for $a$, $b$, and $s$.
-
-    Notice that the truncation values, $a$ and $b$, are defined in
-    standardized form:
-
-    .. math::
-       a = (u_{\mathrm{l}} - \mathrm{loc})/\mathrm{scale},
-       b = (u_{\mathrm{r}} - \mathrm{loc})/\mathrm{scale}
-
-    where :math:`u_{\mathrm{l}}` and :math:`u_{\text{r}}` are the
-    specific left and right truncation values, respectively. In other
-    words, the support of the distribution becomes
-    :math:`(a*\mathrm{scale} + \mathrm{loc}) < x <= (b*\mathrm{scale}
-    + \mathrm{loc})` when :math:`\mathrm{loc}` and/or
-    :math:`\mathrm{scale}` are provided.
-
-    Suppose a normally distributed random variable ``X`` has mean
-    ``mu`` and standard deviation ``sigma``. Then ``Y = exp(X)`` is
-    lognormally distributed with ``s = sigma`` and ``scale =
-    exp(mu)``. To change the base of the lognormal distribution from
-    ``e`` to base ``b`` multiply ``mu`` and ``sigma`` by ``ln(b)``.
-    
-    %(after_notes)s
-
-    References
[…]
