+++ docs/source/generated/generated/dyad.Orbit.rst
@@ -14,9 +14,12 @@
    .. autosummary::
    
       ~Orbit.__init__
+      ~Orbit.delaunay_elements
       ~Orbit.eccentric_anomaly
       ~Orbit.kinetic_energy
       ~Orbit.mean_anomaly
+      ~Orbit.modified_delaunay_elements
+      ~Orbit.orbital_elements
       ~Orbit.potential
       ~Orbit.radius
       ~Orbit.speed
@@ -35,16 +38,13 @@
       ~Orbit.apoapsis
       ~Orbit.area
       ~Orbit.argument_of_pericentre
-      ~Orbit.delaunay_elements
+      ~Orbit.central_mass
       ~Orbit.eccentricity
       ~Orbit.eccentricity_vector
       ~Orbit.energy
       ~Orbit.inclination
       ~Orbit.longitude_of_ascending_node
       ~Orbit.lrl_vector
-      ~Orbit.mass
-      ~Orbit.modified_delaunay_elements
-      ~Orbit.orbital_elements
       ~Orbit.periapsis
       ~Orbit.period
       ~Orbit.semilatus_rectum

[…]
