+++ docs/source/tmp.txt
@@ -1,132 +1,0 @@
-Running Sphinx v8.1.3
-Dyad (VERSION 0.0.0)
-loading translations [en]... done
-Converting `source_suffix = '.rst'` to `source_suffix = {'.rst': 'restructuredtext'}`.
-loading pickled environment... The configuration has changed (1 option: 'html_static_path')
-done
-[autosummary] generating autosummary for: api_reference/generated/dyad.stats.argument_of_pericentre.rst, api_reference/generated/dyad.stats.eccentricity.rst, api_reference/generated/dyad.stats.inclination.rst, api_reference/generated/dyad.stats.log_period.rst, api_reference/generated/dyad.stats.longitude_of_ascending_node.rst, api_reference/generated/dyad.stats.mass.rst, api_reference/generated/dyad.stats.mass_ratio.rst, api_reference/generated/dyad.stats.period.rst, api_reference/generated/dyad.stats.secondary_mass.rst, api_reference/generated/dyad.stats.semimajor_axis.rst, ..., generated/generated/generated/dyad.stats.secondary_mass.moe2017.rst, generated/generated/generated/dyad.stats.semimajor_axis.opik1924.rst, generated/modules.rst, getting_started/installation.rst, getting_started/what_is_dyad.rst, index.rst, tutorials/_tutorial_4dev.rst, tutorials/tutorial_1.rst, tutorials/tutorial_2.rst, tutorials/tutorial_3.rst
-building [mo]: targets for 0 po files that are specified
-writing output... 
-building [doctest]: 1 source files given on command line
-updating environment: 0 added, 0 changed, 0 removed
-reading sources... 
-looking for now-outdated files... none found
-preparing documents... done
-copying assets... 
-copying assets: done
-running tests...
-
-Document: tutorials/tutorial_1
-------------------------------
-**********************************************************************
-File "tutorials/tutorial_1.rst", line 137, in python
-Failed example:
-    orbit.modified_delaunay_elements()
-Expected:
-    (2.398913957225284e-06, 0.0, 0.017202098944262438, -0.0, -0.0)
-Got:
-    array([[ 2.39891396e-06,  7.90666244e-03,  1.72020989e-02,
-            -2.00000000e+00, -1.00000000e+00]])
-**********************************************************************
-File "tutorials/tutorial_1.rst", line 144, in python
-Failed example:
-    orbit.modifed_delaunay_elements(theta=1.)
-Exception raised:
-    Traceback (most recent call last):
-      File "/home/clws00198-ag0082/.pyenv/versions/3.10.12/lib/python3.10/doctest.py", line 1350, in __run
-        exec(compile(example.source, filename, "single",
-      File "<doctest python[0]>", line 1, in <module>
-        orbit.modifed_delaunay_elements(theta=1.)
-    AttributeError: 'Orbit' object has no attribute 'modifed_delaunay_elements'
-**********************************************************************
-File "tutorials/tutorial_1.rst", line 174, in python
-Failed example:
-    binary.state(1.)
-Expected:
-    array([ 7.14066735e+00,  1.11209305e+01,  0.00000000e+00, -4.96132768e-03,
-            6.24861298e-03,  0.00000000e+00])
-Got:
-    array([ 7.14066735e+00,  1.11209305e+01,  0.00000000e+00, -4.96110162e-03,
-            6.24832826e-03,  0.00000000e+00])
-**********************************************************************
-File "tutorials/tutorial_1.rst", line 191, in python
-Failed example:
-    binary.primary.state(1.)
-Expected:
-    array([ 3.26507954e+00,  5.08506010e+00,  0.00000000e+00, -2.26928956e-03,
[…]
