### 2025-11-28 a4e5253
Example scripts added
