+++ examples/README.md
@@ -1,0 +1,8 @@
+This directory contains example scripts that use Dyad to synthesize a population of binary stars.
+
+Run them as follows.
+
+```sh
+$ python duquennoy1991.py
+$ python duquennoy1991.py
+```
