### 2025-11-28 ba15799
Example scripts added
