+++ examples/README.md
@@ -3,6 +3,6 @@ This directory contains example scripts that use Dyad to synthesize a population
 Run them as follows.
 
 ```sh
-$ python duquennoy1991.py
-$ python duquennoy1991.py
+$ python duquennoy1991_population.py
+$ python moe2017_population.py
 ```
