+++ .github/workflows/python-app.yml
@@ -29,4 +29,4 @@ jobs:
         pip install -r requirements.txt
     - name: Test with unittest
       run: |
-        python -m unittest discover -s ./tests/unit_tests/
+        python -m unittest discover -s ./tests/
