### 2025-12-18 8e69878
Installation instructions updated
