### 2025-12-18 #9011 @AmeryGration
Hi @stegmaja. Thanks for getting back to me. I've updated the installation instructions to include instructions on running the test suite. I'll get onto your comments about the paper itself as soon as possible.
