+++ requirements.txt
@@ -2,28 +2,18 @@ alabaster==1.0.0
 babel==2.17.0
 certifi==2025.11.12
 charset-normalizer==3.4.4
-contourpy==1.3.2
-cycler==0.12.1
 docutils==0.21.2
-fonttools==4.60.1
 idna==3.11
 imagesize==1.4.1
 intersphinx_registry==0.2511.7
 Jinja2==3.1.6
-kiwisolver==1.4.9
 MarkupSafe==3.0.3
-matplotlib==3.10.7
-numpy==1.26.4
 numpydoc==1.9.0
 packaging==25.0
 parameterized==0.9.0
-pillow==12.0.0
 Pygments==2.19.2
 pyparsing==3.2.5
-python-dateutil==2.9.0.post0
 requests==2.32.5
-scipy==1.11.0
-six==1.17.0
 snowballstemmer==3.0.1
 Sphinx==8.1.3
 sphinx-copybutton==0.5.2
