+++ requirements.txt
@@ -15,6 +15,7 @@ Pygments==2.19.2
 pyparsing==3.2.5
 requests==2.32.5
 snowballstemmer==3.0.1
+scipy==1.11.0
 Sphinx==8.1.3
 sphinx-copybutton==0.5.2
 sphinx-rtd-theme==3.0.2
