### 2026-01-06 7eaa063
Add random variable
