+++ tests/test_stats_mass_ratio.py
@@ -4,7 +4,6 @@
 Test functions for stats.mass_ratio module
 
 """
-
 import unittest
 import dyad.stats.mass_ratio as mass_ratio
 import data_test_mass_ratio
@@ -17,11 +16,23 @@ duquennoy1991_data = [
     data_test_mass_ratio.duquennoy1991_ppf,
     data_test_mass_ratio.duquennoy1991_rvs
 ]
-moe2017_data = [
-    data_test_mass_ratio.moe2017_pdf,
-    data_test_mass_ratio.moe2017_cdf,
-    data_test_mass_ratio.moe2017_ppf,
-    data_test_mass_ratio.moe2017_rvs
+moe2017_data_a = [
+    data_test_mass_ratio.moe2017_pdf_a,
+    data_test_mass_ratio.moe2017_cdf_a,
+    data_test_mass_ratio.moe2017_ppf_a,
+    data_test_mass_ratio.moe2017_rvs_a
+]
+moe2017_data_b = [
+    data_test_mass_ratio.moe2017_pdf_b,
+    data_test_mass_ratio.moe2017_cdf_b,
+    data_test_mass_ratio.moe2017_ppf_b,
+    data_test_mass_ratio.moe2017_rvs_b
+]
+uniform_data = [
+    data_test_mass_ratio.uniform_pdf,
+    data_test_mass_ratio.uniform_cdf,
+    data_test_mass_ratio.uniform_ppf,
+    data_test_mass_ratio.uniform_rvs
 ]
 
 
@@ -30,7 +41,15 @@ class TestDuquennoy1991(
     pass
 
 
-class TestMoe2017(test_factory(mass_ratio.moe2017, moe2017_data)):
+class TestMoe2017A(test_factory(mass_ratio.moe2017, moe2017_data_a)):
+    pass
+
+
+class TestMoe2017B(test_factory(mass_ratio.moe2017, moe2017_data_b)):
+    pass
+
+
+class TestUniform(test_factory(mass_ratio.uniform, uniform_data)):
     pass
 
 
