+++ dyad/stats/_distr_params.py
@@ -13,11 +13,12 @@ distcont = [
     ["log_period.moe2017", (1.,)],
     ["mass.splitpowerlaw", ((0.5, 0.1, 60., -1.3, -2.3),)],
     ["mass_ratio.moe2017", ((2.56, 1.),)],
-    ["period.trunclognorm", ((5.30, 1.58e-07, 15800000., 63100.),)],
-    ["period.duquennoy1991", (1.,)],
-    ["period.moe2017", (1.,)],
+    ["mass_ratio.uniform", ((100., 0.08, 0.1),)],
+    # ["period.trunclognorm", ((5.30, 1.58e-07, 15800000., 63100.),)],
+    # ["period.duquennoy1991", (1.,)],
+    # ["period.moe2017", (1.,)],
     # ["semimajor_axis.opik1924", ((15., 10.),)],
-    ["secondary_mass.moe2017", (60.,)]
+    ["secondary_mass.moe2017", (60.,)],
 ]
 
 distdiscrete = []
