### 2026-01-06 da6649b
Ensure moe2017 mass bounds are correct throughout
