### 2026-01-06 2b443f2
Ensure moe2017 mass bounds are correct throughout
