+++ tests/data_test_mass_ratio.py
@@ -1,120 +1,0 @@
-"""Data to be used when testing Dyad's stats.mass_ratio module
-
-This module is not to be run on its own. It is imported by
-`test_stats_mass_ratio.py`.
-
-When checking for value errors each list contains arguments that
-should cause such an error. If a value error is raised then the test
-is passed.
-
-Each element in a list is itself a list that contains two elements:
-1. the arguments for a function being tested, and
-2. the desired return value of that function for these arguments.
-
-For example to test the function `dyad.func(*args)` we use the
-following data.
-
-.. code-block:: python
-   func = [
-       [args, val],
-       [args, val],
-       ...
-       [args, val],
-   ]
-
-If the function `dyad.stats.mass_ratio.{attribute,method}(*args)` returns
-the value `val` then the test is passed.
-
-"""
-duquennoy1991_pdf = [
-    [[0.0], 1.1547652495302683],
-    [[1.0], 0.24989482022633835],
-    [[0.6184501482777122], 0.8747104796769749],
-    [[[0.0, 0.25, 0.5, 0.75, 1.0]], [1.1547652495302683, 1.3400495621605901, 1.0911233193188794, 0.6233796707374122, 0.24989482022633835]],
-]
-duquennoy1991_cdf = [
-    [[0.0], 0.0],
-    [[1.0], 0.9528595879197985],
-    [[0.6184501482777122], 0.7492824890475676],
-    [[[0.0, 0.25, 0.5, 0.75, 1.0]], [0.0, 0.320629755657039, 0.6325566968647302, 0.8476880875446968, 0.9528595879197985]],
-]
-duquennoy1991_ppf = [
-    [[0.0], 0.0],
-    [[0.9528595879197985], 1.000000000000001],
-    [[0.7492824890475676], 0.6184501482777125],
-    [[[0.0, 0.320629755657039, 0.6325566968647302, 0.8476880875446968, 0.9528595879197985]], [0.0, 0.2500000000000002, 0.5000000000000003, 0.7500000000000007, 1.000000000000001]],
-]
-duquennoy1991_rvs = [
-    [[0.0, 1.0, 1, 93438701], [0.21807489417693618]],
-    [[0.0, 1.0, 2, 93438701], [0.21807489417693618, 0.4237701636423711]],
-    [[0.0, 1.0, (2, 3), 93438701], [[0.21807489417693618, 0.4237701636423711, 0.42721858592271766], [0.11122988787393193, 0.6158263298786724, 0.5427489420368812]]],
-]
-moe2017_pdf_a = [
-    [[0.0, 0.2, 1.0], 0.0],
-    [[1.0, 0.2, 1.0], 5.428294278549463],
-    [[0.6184501482777122, 0.2, 1.0], 0.7885656454535611],
-    [[[0.0, 0.25, 0.5, 0.75, 1.0], 0.2, 1.0], [0.0, 1.0719516585177544, 0.877011321029882, 0.7160767450558069, 5.428294278549463]],
-]
-moe2017_cdf_a = [
-    [[0.0, 0.2, 1.0], 0.0],
-    [[1.0, 0.2, 1.0], 1.0],
[…]
