### 2026-01-06 6a87b2f
Ensure moe2017 mass bounds are correct throughout
