+++ dyad/stats/mass.py
@@ -1,0 +1,325 @@
+"""
+=============================
+Mass (:mod:`dyad.stats.mass`)
+=============================
+
+.. currentmodule:: dyad.stats.mass
+
+This module contains probability distributions for the masses of a
+population of stars. In its documentation the random variable is
+denoted :math:`M` and a realization of that random variable is denoted
+:math:`m`.
+
+Probability distributions
+=========================
+
+.. autosummary::
+   :toctree: generated/
+
+   splitpowerlaw
+   kroupa2002
+   salpeter1955
+
+"""
+
+__all__ = [
+    "splitpowerlaw",
+    "salpeter1955",
+    # "millerscalo1979",
+    "kroupa2002",
+    # "chabrier2003",
+    # "maschberger2013"
+]
+
+import numpy as np
+import scipy as sp
+
+from . import _distn_infrastructure
+
+
+class splitpowerlaw_gen(_distn_infrastructure.rv_continuous):
+    r"""The two-piece split-power-law mass random variable
+
+    %(before_notes)s
+
+    Notes
+    -----
+    The probability density function for `splitpower` is:
+
+    .. math::
+       f_{M}(m)
+       =
+       A
+       \begin{cases}
+       m^{-\alpha}&\text{ if $m \in [a, b)$,}\\
+       Bm^{-\beta}&\text{ if $m \in [b, c]$}
+       \end{cases}
+
+    for
+
+    .. math::
+       A = \dfrac{1}{1 - \alpha}(b^{1 - \alpha} - a^{1 - \alpha}) + \dfrac{B}{1 - \beta}(c^{1 - \beta} - b^{1 - \beta}),
+
+    and
+
+    .. math::
+       B = b^{\beta - \alpha}.
+
+    and :math:`m \in [a, c]`.
+    
+    %(after_notes)s
+
+    %(example)s
+
+    """
+    def _argcheck(self, s, a, b, c, d):
+        return (0. < a) & (a < b) & (a < s) & (s < b) & (c < 0.) & (d < 0.)
+
+    def _shape_info(self):
+        is_ = _ShapeInfo("s", False, (0., np.inf), (True, False))
+        ia = _ShapeInfo("a", False, (0., np.inf), (True, False))
+        ib = _ShapeInfo("b", False, (0., np.inf), (True, False))
+        ic = _ShapeInfo("c", False, (-np.inf, np.inf), (False, False))
+        id_ = _ShapeInfo("d", False, (-np.inf, np.inf), (False, False))
+
+        return [is_, ia, ib, ic, id_]
+
+    def _get_support(self, s, a, b, c, d):
+        return a, b
+
+    def _pdf(self, x, s, a, b, c, d):
+        def f1(x):
+            return x**c
+
+        def f2(x):
+            return s**(c - d)*x**d
+
+        A = (
+            (s**(c + 1.) - a**(c + 1.))/(c + 1.)
+            + s**(c - d)*(b**(d + 1.) - s**(d + 1.))/(d + 1.)
+        )
+
+        return np.where(x < s, f1(x), f2(x))/A
+
+    def _cdf(self, x, s, a, b, c, d):
+        def f1(x):
+            return (x**(c + 1.) - a**(c + 1.))/(c + 1.)
+
+        def f2(x):
+            return f1(s) + s**(c - d)*(x**(d + 1.) - s**(d + 1.))/(d + 1.)
+
+        A = (
+            (s**(c + 1.) - a**(c + 1.))/(c + 1.)
+            + s**(c - d)*(b**(d + 1.) - s**(d + 1.))/(d + 1.)
+        )
+
[…]
