### 2026-01-06 e48a363
Ensure moe2017 mass bounds are correct throughout
