+++ dyad/stats/__init__.py
@@ -41,7 +41,6 @@ of the condition.
    log_period
    mass
    mass_ratio
-   secondary_mass
    semimajor_axis
 
 Warnings/errors used in :mod:`dyad.stats`
@@ -71,7 +70,7 @@ from . import mass
 from . import mass_ratio
 from . import log_period
 # from . import period
-from . import secondary_mass
+# from . import secondary_mass
 from . import semimajor_axis
 
 
