+++ docs/source/api_reference/generated/generated/dyad.stats.secondary_mass.moe2017.rst
@@ -1,6 +1,0 @@
-dyad.stats.secondary\_mass.moe2017
-==================================
-
-.. currentmodule:: dyad.stats.secondary_mass
-
-.. autodata:: moe2017
