+++ dyad/stats/data/_period/plot.py
@@ -22,8 +22,8 @@ ax.set_xscale("log")
 ax.set_yscale("log")
 ax.set_xlabel(r"$\log(p)$")
 ax.set_ylabel(r"$m_{1}/\mathrm{M}_{\odot}$")
-# plt.savefig("./Figures/pdf_log10_period.pdf")
-# plt.savefig("./Figures/pdf_log10_period.jpg")
+plt.savefig("./Figures/pdf_log10_period.pdf")
+plt.savefig("./Figures/pdf_log10_period.jpg")
 plt.show()
 
 fig, ax = plt.subplots()
@@ -37,6 +37,6 @@ ax.set_xscale("log")
 ax.set_yscale("log")
 ax.set_xlabel(r"$\log(p)$")
 ax.set_ylabel(r"$m_{1}/\mathrm{M}_{\odot}$")
-# plt.savefig("./Figures/cdf_log10_period.pdf")
-# plt.savefig("./Figures/cdf_log10_period.jpg")
+plt.savefig("./Figures/cdf_log10_period.pdf")
+plt.savefig("./Figures/cdf_log10_period.jpg")
 plt.show()
