+++ docs/source/api_reference/generated/generated/dyad.stats.mass_ratio.uniform.rst
@@ -1,0 +1,6 @@
+﻿dyad.stats.mass\_ratio.uniform
+==============================
+
+.. currentmodule:: dyad.stats.mass_ratio
+
+.. autodata:: uniform
