### 2026-01-07 82d76a3
Check documentation for completeness and consistency
