+++ docs/source/conf.py
@@ -9,6 +9,7 @@ import scipy
 import scipy._lib.uarray as ua
 import numpydoc.docscrape as np_docscrape
 import doctest
+import sphinx_rtd_theme
 
 from os.path import relpath, dirname
 from docutils import nodes
@@ -33,21 +34,17 @@ sys.path.insert(0, os.path.abspath(os.path.dirname(__file__)))
 ########################################################################
 
 extensions = [
-    "sphinx.ext.viewcode",
     "sphinx.ext.autodoc",
     "sphinx.ext.doctest",
     "sphinx.ext.autosummary",
-    "sphinx.ext.coverage",
+    "sphinx.ext.viewcode",
     "sphinx.ext.mathjax",
     "sphinx.ext.intersphinx",
-    "numpydoc",
+    "sphinx.ext.coverage",
     "sphinx_copybutton",
     "sphinx_design",
-    # "scipyoptdoc",
-    # "doi_role",
+    "numpydoc",
     "matplotlib.sphinxext.plot_directive",
-    # "myst_nb",
-    # "jupyterlite_sphinx",
 ]
 
 html_logo = "dyad_logo_white.png"
@@ -72,7 +69,7 @@ master_doc = "index"
 sys.path.insert(0, os.path.abspath("../../dyad"))
 
 project = "Dyad"
-copyright = "2024, Amery Gration"
+copyright = "2026, Amery Gration"
 author = "Amery Gration"
 version = "0.0.0"
 release = "0.0.0"
@@ -84,24 +81,10 @@ if os.environ.get("CIRCLE_JOB", False) and \
 
 print(f"{project} (VERSION {version})")
 
-# There are two options for replacing |today|: either, you set today to some
-# non-false value, then it is used:
-#today = ""
-# Else, today_fmt is used as the format for a strftime call.
 today_fmt = "%B %d, %Y"
-
-# List of documents that shouldn"t be included in the build.
-#unused_docs = []
 
 # The reST default role (used for this markup: `text`) to use for all documents.
 default_role = "autolink"
-
-# List of directories, relative to source directories, that shouldn"t be searched
-# for source files.
-exclude_dirs = []
-exclude_patterns = [  # glob-style
-    "**.ipynb",
-]
 
 # If true, "()" will be appended to :func: etc. cross-reference text.
 add_function_parentheses = False
@@ -117,157 +100,14 @@ show_authors = False
 # The name of the Pygments (syntax highlighting) style to use.
 pygments_style = "sphinx"
 
-# # Ensure all our internal links work
-# nitpicky = True
-# nitpick_ignore = [
-#     # This ignores errors for classes (OptimizeResults, sparse.dok_matrix)
-#     # which inherit methods from `dict`. missing references to builtins get
-#     # ignored by default (see https://github.com/sphinx-doc/sphinx/pull/7254),
-#     # but that fix doesn"t work for inherited methods.
-#     ("py:class", "a shallow copy of D"),
-#     ("py:class", "a set-like object providing a view on D's keys"),
-#     ("py:class", "a set-like object providing a view on D's items"),
-#     ("py:class", "an object providing a view on D's values"),
-#     ("py:class", "None.  Remove all items from D."),
-#     ("py:class", "(k, v), remove and return some (key, value) pair as a"),
-#     ("py:class", "None.  Update D from dict/iterable E and F."),
-#     ("py:class", "v, remove specified key and return the corresponding value."),
-# ]
+doctest_optionflags = [doctest.ELLIPSIS]
 
-# # be strict about warnings in our examples, we should write clean code
-# # (exceptions permitted for pedagogical purposes below)
-# warnings.resetwarnings()
-# warnings.filterwarnings("error")
-# # allow these and show them
[…]
