+++ docs/source/tutorials/tutorial_1.rst
@@ -4,7 +4,7 @@
 Representing a Keplerian system
 *******************************
 
-To a good approximation the Earth--Sun system is Keplerian. The Earth moves on an elliptical orbit with semimajor axis :math:`a = 1~\text{AU}` and ellipticity :math:`e = 0.0167` in the gravitational central potential generated by the Sun, which has mass :math:`m = \text{M}_{\odot}`. We may represent this system in Dyad using the class :class:`dyad.Orbit`.
+To a good approximation the Earth--Sun system is Keplerian. The Earth moves on an elliptical orbit with semimajor axis :math:`a = 1~\text{AU}` and ellipticity :math:`e = 0.0167` in the gravitational central potential generated by the Sun, which has mass :math:`m = 1~\text{M}_{\odot}`. We may represent this system in Dyad using the class :class:`dyad.Orbit`.
 
 .. doctest:: python
 
@@ -21,7 +21,13 @@ The state (i.e. position and velocity) of the body is specified by its true anom
    array([ 0.53532139,  0.83371367,  0.        , -0.01447709,  0.00958295,
            0.        ])
 
-This is given in the format :math:`(x/\mathrm{AU}, y/\mathrm{AU}, z/\mathrm{AU}, v_{x}/\mathrm{AU}~\mathrm{d}^{-1}, v_{y}/\mathrm{AU}~\mathrm{d}^{-1}, v_{z}/\mathrm{AU}~\mathrm{d}^{-1})`. We can also compute the radius and speed of the body, which are given in units of :math:`\mathrm{AU}` and :math:`\mathrm{AU}~\mathrm{d}^{-1}`.
+This is given in the format :math:`(x, y, z, v_{x}, v_{y}, v_{z})`
+where :math:`x`, :math:`y`, and :math:`z` are in units of
+:math:`\mathrm{AU}` and :math:`v_{x}`, :math:`v_{y}`, and
+:math:`v_{z}` are in units of :math:`\mathrm{AU}~\mathrm{d}^{-1}`. We
+can also compute the orbital radius and speed of the body, which are
+also in units of :math:`\mathrm{AU}` and
+:math:`\mathrm{AU}~\mathrm{d}^{-1}`.
 
 .. doctest:: python
 
@@ -42,7 +48,7 @@ attributes that represent constants of the orbit. These include the arguments wi
    >>> orbit.eccentricity
    0.0167
 
-They also include more interesting constants, such as the period, periapsis and apoapsis. A full list is available in the API documentation for :class:`dyad.Orbit`.
+They also include more interesting constants, such as the period (in :math:`\mathrm{d}`), periapsis (in :math:`\mathrm{AU}`) and apoapsis (in :math:`\mathrm{AU}`). A full list is available in the API documentation for :class:`dyad.Orbit`.
 
 .. doctest:: python
 
@@ -53,9 +59,7 @@ They also include more interesting constants, such as the period, periapsis and 
    >>> orbit.apoapsis
    1.0167
 
-These are given in units of :math:`\mathrm{d}` and :math:`\mathrm{AU}`.
-
-If the perifocal frame differs from the observer's frame we may specify its orientation using the Euler angles :math:`\Omega`, :math:`i`, and :math:`\omega` (longitude of the ascending node, inclination, and argument of pericentre). Suppose that :math:`\Omega = 1`, :math:`i = 1`, and :math:`\omega = 1`.
+If the perifocal frame differs from the observer's frame we can specify its orientation using the Euler angles :math:`\Omega`, :math:`i`, and :math:`\omega` (longitude of the ascending node, inclination, and argument of pericentre). Suppose that :math:`\Omega = 1`, :math:`i = 1`, and :math:`\omega = 1`.
 
 .. doctest:: python
 
[…]

[…]
