### 2026-01-07 05dd9c6
Check documentation for completeness and consistency
