+++ dyad/stats/semimajor_axis.py
@@ -30,7 +30,44 @@ import scipy as sp
 from . import _distn_infrastructure
 
 
-class opik1924_gen(sp.stats._continuous_distns.reciprocal_gen):
+# class opik1924_gen(sp.stats._continuous_distns.reciprocal_gen):
+#     r"""The semimajor-axis random variable of Öpik (1924)
+
+#     %(before_notes)s
+
+#     Notes
+#     -----
+#     The probability density function for `opik1924` is:
+
+#     .. math::
+#        f_{A}(a; b, c) = \dfrac{1}{a\log_{10}(c/b)}
+
+#     for :math:`a \in [b, c]`, :math:`b, c \in (0, \infty)`, and
+#     :math:`b < c`. The probability density function `opik1924` is
+#     identical to `scipy.stats.reciprocal`.
+
+#     `opik1924` takes ``b`` and ``c`` as shape parameters.
+
+#     %(after_notes)s
+
+#     References
+#     ----------
+#     Öpik, E. 1924. \'Statistical studies of double stars: on the
+#     distribution of relative luminosities and distances of double
+#     stars in the Harvard Revised Photometry North of
+#     Declination---31°\'. *Publications of the Tartu Astrofizica
+#     Observatory* 25 (January):1.
+    
+#     %(example)s
+
+#     """
+#     def __init__(self, *args, **kwargs):
+#         super().__init__(*args, **kwargs)
+
+        
+# opik1924 = opik1924_gen(name="semimajor_axis.opik1924")
+
+class opik1924_gen(_distn_infrastructure.rv_continuous):
     r"""The semimajor-axis random variable of Öpik (1924)
 
     %(before_notes)s
@@ -40,7 +77,7 @@ class opik1924_gen(sp.stats._continuous_distns.reciprocal_gen):
     The probability density function for `opik1924` is:
 
     .. math::
-       f_{A}(a; b, c) = \dfrac{1}{a\log_{10}(c/b)}
+       f_{A}(a; b, c) = \dfrac{1}{a\log(c/b)}
 
     for :math:`a \in [b, c]`, :math:`b, c \in (0, \infty)`, and
     :math:`b < c`. The probability density function `opik1924` is
@@ -61,8 +98,29 @@ class opik1924_gen(sp.stats._continuous_distns.reciprocal_gen):
     %(example)s
 
     """
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
+    def _argcheck(self, b, c):
+        return (b > 0) & (c > b)
 
-        
+    def _shape_info(self):
+        ib = _ShapeInfo("b", False, (0, np.inf), (False, False))
+        ic = _ShapeInfo("c", False, (0, np.inf), (False, False))
+        return [ib, ic]
+
+    def _get_support(self, b, c):
+        return b, c
+
+    def _pdf(self, x, b, c):
+        return np.exp(self._logpdf(x, b, c))
+
+    def _logpdf(self, x, b, c):
+        return -np.log(x) - np.log(np.log(c) - np.log(b))
+
+    def _cdf(self, x, b, c):
+        return (np.log(x)-np.log(b)) / (np.log(c) - np.log(b))
+
+    def _ppf(self, q, b, c):
+        return np.exp(np.log(b) + q*(np.log(c) - np.log(b)))
+
+
 opik1924 = opik1924_gen(name="semimajor_axis.opik1924")
+opik1924._support = ("b", "c")
