### 2026-01-07 b6f420b
Check documentation for completeness and consistency
