### 2026-01-07 cc066fe
Check documentation for completeness and consistency
