+++ tests/data_test_eccentricity.py
@@ -127,49 +127,3 @@ moe2017_rvs = [
     [[8.0, 1.0, 0.0, 1.0, 2, 93438701], [0.4273712300660013, 0.6682649957508963]],
     [[8.0, 1.0, 0.0, 1.0, (2, 3), 93438701], [[0.4273712300660013, 0.6682649957508963, 0.6716402178446679], [0.2667561780405715, 0.8239676036573242, 0.7723680206617772]]],
 ]
-uniform_pdf = [
-    [[0.0], 1.0],
-    [[1.0], 1.0],
-    [[0.7276105655530771], 1.0],
-    [[[0.0, 0.25, 0.5, 0.75, 0.99]], [1.0, 1.0, 1.0, 1.0, 1.0]],
-]
-uniform_cdf = [
-    [[0.0], 0.0],
-    [[1.0], 1.0],
-    [[0.7276105655530771], 0.7276105655530771],
-    [[[0.0, 0.25, 0.5, 0.75, 0.99]], [0.0, 0.25, 0.5, 0.75, 0.99]],
-]
-uniform_ppf = [
-    [[0.0], 0.0],
-    [[1.0], 1.0],
-    [[0.7276105655530771], 0.7276105655530771],
-    [[[0.0, 0.25, 0.5, 0.75, 0.99]], [0.0, 0.25, 0.5, 0.75, 0.99]],
-]
-uniform_rvs = [
-    [[0.0, 1.0, 1, 93438701], [0.27781228549600867]],
-    [[0.0, 1.0, 2, 93438701], [0.27781228549600867, 0.5448298357085596]],
-    [[0.0, 1.0, (2, 3), 93438701], [[0.27781228549600867, 0.5448298357085596, 0.5489811591923758], [0.1365685339802215, 0.7469807794194006, 0.6776312233429724]]],
-]
-powerlaw_pdf = [
-    [[0.0, 2.0], 0.0],
-    [[1.0, 2.0], 2.0],
-    [[0.7276105655530771, 2.0], 1.4552211311061543],
-    [[[0.0, 0.25, 0.5, 0.75, 0.99], 2.0], [0.0, 0.5, 1.0, 1.5, 1.98]],
-]
-powerlaw_cdf = [
-    [[0.0, 2.0], 0.0],
-    [[1.0, 2.0], 1.0],
-    [[0.7276105655530771, 2.0], 0.5294171351044688],
-    [[[0.0, 0.25, 0.5, 0.75, 0.99], 2.0], [0.0, 0.0625, 0.25, 0.5625, 0.9801000000000001]],
-]
-powerlaw_ppf = [
-    [[0.0, 2.0], 0.0],
-    [[1.0, 2.0], 1.0],
-    [[0.5294171351044688, 2.0], 0.7276105655530771],
-    [[[0.0, 0.0625, 0.25, 0.5625, 0.9801000000000001], 2.0], [0.0, 0.25, 0.5, 0.75, 0.9900000000000001]],
-]
-powerlaw_rvs = [
-    [[2.0, 0.0, 1.0, 1, 93438701], [0.5270790125740245]],
-    [[2.0, 0.0, 1.0, 2, 93438701], [0.5270790125740245, 0.7381258942135546]],
-    [[2.0, 0.0, 1.0, (2, 3), 93438701], [[0.5270790125740245, 0.7381258942135546, 0.7409326279712453], [0.36955180148420536, 0.8642804981135468, 0.8231835903022924]]],
-]
