### 2026-01-07 1334e64
Check documentation for completeness and consistency
