+++ dyad/_core.py
@@ -88,18 +88,18 @@ def semimajor_axis_from_period(p, m_1, m_2):
 
     m_1 : array-like
 
-        Mass of the primary body, :math:`m_{1}`.
+        Mass of the primary body, :math:`m_{1}`
 
     m_2 : array-like
 
-        Mass of the secondary body, :math:`m_{2}`.
+        Mass of the secondary body, :math:`m_{2}`
 
     Returns
     -------
 
     res : ndarray
 
-        Semimajor axis.
+        Semimajor axis
 
     Examples
     --------
@@ -135,22 +135,22 @@ def period_from_semimajor_axis(a, m_1, m_2):
 
     a : array-like
 
-        Total semimajor axis, :math:`a = a_{1} + a_{2}`.
+        Total semimajor axis, :math:`a = a_{1} + a_{2}`
 
     m_1 : array-like
 
-        Mass of the primary body, :math:`m_{1}`.
+        Mass of the primary body, :math:`m_{1}`
 
     m_2 : array-like
 
-        Mass of the secondary body, :math:`m_{2}`.
+        Mass of the secondary body, :math:`m_{2}`
 
     Returns
     -------
 
     res : ndarray
 
-        Semimajor axis.
+        Semimajor axis
 
     Examples
     --------
@@ -183,11 +183,11 @@ def mean_anomaly_from_eccentric_anomaly(eta, e):
 
     eta : array-like
 
-        Eccentric anomaly.
+        Eccentric anomaly
 
     e : array-like
 
-        Eccentricity.
+        Eccentricity
 
     Returns
     -------
@@ -665,7 +665,8 @@ def delaunay_elements_from_orbital_elements(a, e, Omega, i, omega, theta, m):
     2\pi)` is the longitude of the ascending node, :math:`i \in (0,
     \pi)` is the inclination, :math:`\omega \in [0, 2\pi)` is the
     argument of pericentre, and :math:`\mu(\theta) \in [0, 2\pi)` is
-    the mean anomaly.
+    the mean anomaly corresponding to the true anomaly, :math:`\theta
+    \in [0, 2\pi)`.
 
     Parameters
     ----------
@@ -703,7 +704,8 @@ def delaunay_elements_from_orbital_elements(a, e, Omega, i, omega, theta, m):
 
     res : list
 
-        Delaunay elements ``(J_1, J_2, J_3, Theta_1, Theta_2, Theta_3)``
+        Delaunay elements, :math:`(J_{1}, J_{2}, J_{3}, \Theta_{1},
+        \Theta_{2}, \Theta_{3})`.
 
     Warnings
     --------
@@ -780,13 +782,19 @@ def orbital_elements_from_delaunay_elements(
        \omega &= \Theta_{2}
 
     where
-    :math:`J_{1}/(\mathrm{AU}^{2}~\mathrm{d}^{-1}) \in (-\infty, \infty)`,
-    :math:`J_{2}/(\mathrm{AU}^{2}~\mathrm{d}^{-1}) \in (0, \infty)`,
-    :math:`J_{3}/(\mathrm{AU}^{2}~\mathrm{d}^{-1}) \in (0, \infty)`,
+    :math:`J_{1}/(\mathrm{AU}^{2}~\mathrm{d}^{-1}) \in (-\infty, \infty)`
+    is the first Delaunay action,
+    :math:`J_{2}/(\mathrm{AU}^{2}~\mathrm{d}^{-1}) \in (0, \infty)`
+    is the second Delaunay action,
+    :math:`J_{3}/(\mathrm{AU}^{2}~\mathrm{d}^{-1}) \in (0, \infty)`
+    is the third Delaunay action,
[…]
