### 2026-01-07 5fa27fb
Check documentation for completeness and consistency
