+++ docs/source/api_reference/index.rst
@@ -11,11 +11,3 @@ The following packages and subpackages are public.
 
 * :mod:`dyad`
 * :mod:`dyad.stats`
-
-.. toctree::
-   :maxdepth: 1
-   :hidden:
-   :titlesonly:
-
-   dyad.stats <stats>
-
