### 2026-01-07 3d3cbc6
Check documentation for completeness and consistency
