+++ docs/source/generated/generated/dyad.stats.true_anomaly.rst
@@ -1,6 +1,0 @@
-﻿dyad.stats.true\_anomaly
-========================
-
-.. currentmodule:: dyad.stats
-
-.. autodata:: true_anomaly
