+++ docs/source/api_reference/generated/dyad.semimajor_axis_from_period.rst
@@ -1,0 +1,6 @@
+﻿dyad.semimajor\_axis\_from\_period
+==================================
+
+.. currentmodule:: dyad
+
+.. autofunction:: semimajor_axis_from_period
