+++ docs/source/generated/generated/dyad.stats.mass_ratio.uniform.rst
@@ -1,6 +1,0 @@
-﻿dyad.stats.mass\_ratio.uniform
-==============================
-
-.. currentmodule:: dyad.stats.mass_ratio
-
-.. autodata:: uniform
