+++ docs/source/generated/generated/dyad.delaunay_elements_from_orbital_elements.rst
@@ -1,6 +1,0 @@
-﻿dyad.delaunay\_elements\_from\_orbital\_elements
-================================================
-
-.. currentmodule:: dyad
-
-.. autofunction:: delaunay_elements_from_orbital_elements
