+++ docs/source/api_reference/generated/dyad.modified_delaunay_elements_from_orbital_elements.rst
@@ -1,0 +1,6 @@
+﻿dyad.modified\_delaunay\_elements\_from\_orbital\_elements
+==========================================================
+
+.. currentmodule:: dyad
+
+.. autofunction:: modified_delaunay_elements_from_orbital_elements
