+++ docs/source/generated/generated/dyad.orbital_elements_from_delaunay_elements.rst
@@ -1,6 +1,0 @@
-﻿dyad.orbital\_elements\_from\_delaunay\_elements
-================================================
-
-.. currentmodule:: dyad
-
-.. autofunction:: orbital_elements_from_delaunay_elements
