+++ docs/source/api_reference/generated/generated/dyad.stats.mass_ratio.uniform.rst
@@ -1,4 +1,4 @@
-﻿dyad.stats.mass\_ratio.uniform
+dyad.stats.mass\_ratio.uniform
 ==============================
 
 .. currentmodule:: dyad.stats.mass_ratio
