+++ docs/source/index.rst
@@ -54,12 +54,3 @@ Dyad was written and is maintained by Amery Gration. It is free
 software made available under the GNU General Public License. If you
 make use of it then please cite `my paper`_ in *Journal of open source
 software*.
-
-Indices and tables
-------------------
-
-* :ref:`genindex`
-* :ref:`modindex`
-* :ref:`search`
-
-.. _my paper: xxx
