+++ docs/source/api_reference/generated/dyad.TwoBody.rst
@@ -1,0 +1,56 @@
+﻿dyad.TwoBody
+============
+
+.. currentmodule:: dyad
+
+.. autoclass:: TwoBody
+
+   
+   .. automethod:: __init__
+
+   
+   .. rubric:: Methods
+
+   .. autosummary::
+   
+      ~TwoBody.__init__
+      ~TwoBody.delaunay_elements
+      ~TwoBody.eccentric_anomaly
+      ~TwoBody.kinetic_energy
+      ~TwoBody.mean_anomaly
+      ~TwoBody.modified_delaunay_elements
+      ~TwoBody.orbital_elements
+      ~TwoBody.potential
+      ~TwoBody.radius
+      ~TwoBody.speed
+      ~TwoBody.state
+   
+   
+
+   
+   
+   .. rubric:: Attributes
+
+   .. autosummary::
+   
+      ~TwoBody.angular_momentum
+      ~TwoBody.angular_momentum_magnitude
+      ~TwoBody.apoapsis
+      ~TwoBody.area
+      ~TwoBody.argument_of_pericentre
+      ~TwoBody.central_mass
+      ~TwoBody.eccentricity
+      ~TwoBody.eccentricity_vector
+      ~TwoBody.energy
+      ~TwoBody.inclination
+      ~TwoBody.longitude_of_ascending_node
+      ~TwoBody.lrl_vector
+      ~TwoBody.periapsis
+      ~TwoBody.period
+      ~TwoBody.primary
+      ~TwoBody.secondary
+      ~TwoBody.semilatus_rectum
+      ~TwoBody.semimajor_axis
+      ~TwoBody.semiminor_axis
+   
+   
