+++ docs/source/api_reference/generated/dyad.period_from_semimajor_axis.rst
@@ -1,0 +1,6 @@
+﻿dyad.period\_from\_semimajor\_axis
+==================================
+
+.. currentmodule:: dyad
+
+.. autofunction:: period_from_semimajor_axis
