+++ docs/source/generated/generated/dyad.eccentric_anomaly_from_true_anomaly.rst
@@ -1,6 +1,0 @@
-﻿dyad.eccentric\_anomaly\_from\_true\_anomaly
-============================================
-
-.. currentmodule:: dyad
-
-.. autofunction:: eccentric_anomaly_from_true_anomaly
