+++ docs/source/generated/generated/dyad.stats.semimajor_axis.rst
@@ -1,12 +1,0 @@
-﻿dyad.stats.semimajor\_axis
-==========================
-
-.. automodule:: dyad.stats.semimajor_axis
-
-   
-   .. rubric:: Classes
-
-   .. autosummary::
-   
-      opik1924_gen
-   
