### 2026-01-08 c05733f
Update installation instructions
