### 2026-01-08 fe2e091
Check documentation for completeness and consistency
