+++ dyad/__init__.py
@@ -54,7 +54,7 @@ and :math:`\omega` (being the *longitude of the ascending node*, the
 The geometry of a binary system's component orbits is shown in Figure
 :numref:`twobodysystem`. Both bodies move on elliptical orbits with
 primary foci coincident with the system's centre of mass (COM)
-frame. The more-massive (resp. less-massive) body moves as if in a
+frame. The primary (resp. secondary) body moves as if in a
 central potential generated by a mass of :math:`m_{2}^{3}/(m_{1} +
 m_{2})^{2}` (resp. :math:`m_{1}^{3}/(m_{1} + m_{2})^{2}`). Its orbit
 is described by its semimajor axis, :math:`a_{1}`
@@ -63,7 +63,7 @@ is described by its semimajor axis, :math:`a_{1}`
 its ascending node, :math:`\Omega_{1}` (resp. :math:`\Omega_{2}`),
 inclination, :math:`i_{1}` (resp. :math:`i_{2}`), and argument of
 pericentre, :math:`\omega_{1}` (resp. :math:`\omega_{2}`). The phase
-of the more-massive (resp. less-massive) body is described by its true
+of the primary (resp. secondary) body is described by its true
 anomaly, :math:`\theta_{1}` (resp. :math:`\theta_{2})`. It is the case
 that :math:`a_{2} = a_{1}/q`, :math:`e_{2} = e_{1}`. :math:`\Omega_{2}
 = \Omega_{1}`, :math:`i_{2} = i_{1}`, :math:`\omega_{2} = \omega_{1} +
