### 2026-01-08 dc1fd01
Check documentation for completeness and consistency
