### 2026-01-14 0cdc550
Path handling for examples output updated
