+++ tests/data_test_mass.py
@@ -49,25 +49,25 @@ salpeter1955_rvs = [
     [[0.08, 150.0, 0.0, 1.0, 2, 93438701], [0.10180963931891547, 0.14331112352366235]],
     [[0.08, 150.0, 0.0, 1.0, (2, 3), 93438701], [[0.10180963931891547, 0.14331112352366235, 0.14428697751151806], [0.08919215550721653, 0.2213921758922736, 0.18503214231025114]]],
 ]
-kroupa2002_pdf = [
+kroupa2001_pdf = [
     [[0.08, 0.08, 150.0], 6.7444570546415665],
     [[150.0, 0.08, 150.0], 1.2500405037958306e-06],
     [[97.38967978113911, 0.08, 150.0], 3.3756252280887457e-06],
     [[[0.08, 10.0, 50.0, 100.0, 150.0], 0.08, 150.0], [6.7444570546415665, 0.0006337733613780491, 1.5642385010405695e-05, 3.1763911773572216e-06, 1.2500405037958306e-06]],
 ]
-kroupa2002_cdf = [
+kroupa2001_cdf = [
     [[0.08, 0.08, 150.0], 0.0],
     [[150.0, 0.08, 150.0], 1.0],
     [[97.38967978113911, 0.08, 150.0], 0.9998913500119574],
     [[[0.08, 10.0, 50.0, 100.0, 150.0], 0.08, 150.0], [0.0, 0.9952690557398375, 0.9995426052500377, 0.999899897659872, 1.0]],
 ]
-kroupa2002_ppf = [
+kroupa2001_ppf = [
     [[0.0, 0.08, 150.0], 0.08],
     [[1.0, 0.08, 150.0], 150.0],
     [[0.9998913500119574, 0.08, 150.0], 97.38967978113419],
     [[[0.0, 0.9952690557398375, 0.9995426052500377, 0.999899897659872, 1.0], 0.08, 150.0], [0.08, 9.999999999999845, 49.999999999995524, 99.99999999998343, 150.0]],
 ]
-kroupa2002_rvs = [
+kroupa2001_rvs = [
     [[0.08, 150.0, 0.0, 1.0, 1, 93438701], [0.13995490339681924]],
     [[0.08, 150.0, 0.0, 1.0, 2, 93438701], [0.13995490339681924, 0.26638279550229754]],
     [[0.08, 150.0, 0.0, 1.0, (2, 3), 93438701], [[0.13995490339681924, 0.26638279550229754, 0.26934424018350556], [0.10409095145919046, 0.47869059810439013, 0.3868995651368416]]],
