### 2026-01-16 a278f9d
Second draft of paper
