+++ paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'Dyad: a binary-star dynamics and statistics package for Python'
+title: 'Dyad: a binary-star dynamics and statistics library for Python'
 tags:
   - Python
   - astronomy
@@ -12,57 +12,103 @@ authors:
     affiliation: 1
 affiliations:
   - name: Astrophysics Research Group, University of Surrey, Guildford, GU2 7XH, United Kingdom
-    index: 1
+	index: 1
     ror: 00ks66431
-date: 3 March 2025
+date: 16 January 2026
 bibliography: paper.bib
 ---
 
 # Summary
 
-Dyad is a pure-Python two-body dynamics and binary-star statistics
-package for astrophysicists. It allows the user to compute the
-kinematic properties of a bound gravitational two-body system given
-that system's component masses and orbital elements. It also allows
-the user to synthesize a population of binary stars with component
-masses and orbital elements that follow a given
-distribution. Specifically, Dyad allows the user to synthesize
-primary- and secondary-star masses in a manner consistent with given
-distributions of stellar masses and mass ratios. It does so by
-implementing the method of @gration2025b. Accordingly, Dyad provides a
-library of distributions for stellar mass, mass-ratio, and the orbital
-elements. This library includes (but is not limited to) the
-distributions for (1) the initial stellar mass published by
-@chabrier2003, @kroupa2001, and @salpeter1955, and (2) the
-mass-ratios and orbital elements of binary stars in the Solar
-neighbourhood published by @duquennoy1991 and @moe2017.
+Dyad is a Python library for studying the dynamics of binary stars
+considered as gravitational two-body systems. By convention,
+astrophysicists designate the brighter of the two components of a binary
+star as a reference body, which they call the \`primary star\', and the
+dimmer of the two components as a subject body, which they call the
+\`secondary star\'. In a frame centred on the primary star the secondary
+star then moves on an elliptical orbit with one focus located at the
+origin. This orbit can be specified by its orbital elements, namely the
+semimajor axis (which specifies the size of the ellipse), eccentricity
+(which specifies the shape of the ellipse), and true anomaly (which
+specifies the secondary star\'s location on the ellipse) along with the
+longitude of its ascending node, its inclination, and its argument of
+pericentre (which together specify the orientation of the ellipse).
+
+The dynamics of the system are completely determined by the two stars\'
+masses and the secondary star\'s orbital elements. In a population of
+binary stars these eight parameters vary from member to member and can
+each be treated as a random variable having some probability
+distribution. Dyad provides a class, `dyad.TwoBody`, and a module,
+`dyad.stats`, for dealing with such a population of binary stars. The
+`dyad.TwoBody` class represents a gravitational two-body system while
+the `dyad.stats` module provides a suite of classes representing the
+probability distributions of stellar mass, mass-ratio, and orbital
+elements. Dyad implements these probability distributions in the same
+way that SciPy [@virtanen2020] implements its probability distributions
+(see, for example,
+[`scipy.stats.norm`](https://docs.scipy.org/doc/scipy/reference/generated/scipy.stats.norm.html),
+[`scipy.stats.lognorm`](https://docs.scipy.org/doc/scipy/reference/generated/scipy.stats.lognorm.html),
+or
+[`scipy.stats.expon`](https://docs.scipy.org/doc/scipy/reference/generated/scipy.stats.expon.html)).
+
+You can initialize `dyad.TwoBody` by specifying the primary and
+secondary bodies\' masses together with the secondary body\'s orbital
+elements and then use that class\'s methods to compute the two bodies\'
+angular momenta, total energies, and eccentricity vectors as well as
+positions, velocities, kinetic energies, and potential energies either
[…]
