+++ docs/build/html/_sources/api_reference/generated/dyad.orbital_elements_from_modified_delaunay_elements.rst.txt
@@ -1,0 +1,6 @@
+﻿dyad.orbital\_elements\_from\_modified\_delaunay\_elements
+==========================================================
+
+.. currentmodule:: dyad
+
+.. autofunction:: orbital_elements_from_modified_delaunay_elements
