+++ dyad.egg-info/requires.txt
@@ -1,0 +1,9 @@
+numpy>=1.21.6
+scipy>=1.11
+matplotlib
+numpydoc
+intersphinx_registry
+sphinx_copybutton
+sphinx_design
+sphinx-rtd-theme
+parameterized
