+++ docs/build/html/_sources/api_reference/index.rst.txt
@@ -1,0 +1,13 @@
+.. _dyad-api:
+
+********
+Dyad API
+********
+
+API definition
+==============
+
+The following packages and subpackages are public.
+
+* :mod:`dyad`
+* :mod:`dyad.stats`

[…]
