### 2026-01-16 7dce93b
Second draft of paper
