+++ paper.bib
@@ -5,8 +5,8 @@
   journal = {The Astrophysical Journal Supplement Series},
   volume = {230},
   number = {2},
-  pages = {15},
-  doi = {10.3847/1538-4365/aa6fb6,}
+  pages = {15}
+  doi = {10.3847/1538-4365/aa6fb6},
 }
 
 @article{chabrier2003,
@@ -54,28 +54,23 @@
 @article{arroyo-polonio2023,
   author = {{Arroyo-Polonio}, J. M. and Battaglia, G. and Thomas, G. F. and Irwin, M. J. and McConnachie, A. W. and Tolstoy, E.},
   year = {2023},
-  month = sep,
+  month = sep
   volume = {677},
-  pages = {A95},
+  pages = {A95}
   journal = {Astronomy \& Astrophysics},
   doi = {10.1051/0004-6361/202346843},
 }
 
-@article{gration2025a,
-  author = {Gration, A. and Hendricks, D. and Das, P. and Heber, D. and Izzard, R.},
+@article{gration2025,
+  author = {Gration, A., Hendricks, D., Das, P., Heber, D., Izzard, R.},
   year = {2025},
   month = sep,
   journal = {arXiv e-prints},
   doi = {10.48550/arXiv.2509.14316},
 }
 
-@article{gration2025b,
-  author = {Gration, A. and Izzard, R. G. and Das, P.},
-  pubstate = {inpreparation},
-}
-
 @article{hendriks2023,
-  author = {Hendriks, D. D. and Izzard, R. G.},
+  author = {Hendriks, D. D., and Izzard, R. G.},
   journal = {Journal of Open Source Software},
   year = {2023},
   publisher = {The Open Journal},
@@ -135,12 +130,13 @@
   doi = {10.1017/9781108553070.006},
 }
 
-@article{duquennoy1991,
-  author = {Duquennoy, A. and Mayor, M.},
-  year = {1991},
-  month = aug,
-  journal = {Astronomy and Astrophysics},
-  volume = {248},
-  pages = {485--524},
-  doi = {10.3847/1538-4365/aa6fb6}
+@article{virtanen2020,
+  title = {{SciPy} 1.0: Fundamental Algorithms for Scientific Computing in {Python}},
+  author = {Virtanen, Pauli and Gommers, Ralf and Oliphant, Travis E. and Haberland, Matt and Reddy, Tyler and Cournapeau, David and Burovski, Evgeni and Peterson, Pearu and Weckesser, Warren and Bright, Jonathan and family=Walt, given=St\'efan J., prefix=van der, useprefix=true and Brett, Matthew and Wilson, Joshua and Millman, K. Jarrod and Mayorov, Nikolay and Nelson, Andrew R. J. and Jones, Eric and Kern, Robert and Larson, Eric and Carey, C. J. and Polat, \.Ilhan and Feng, Yu and Moore, Eric W. and VanderPlas, Jake and Laxalde, Denis and Perktold, Josef and Cimrman, Robert and Henriksen, Ian and Quintero, E. A. and Harris, Charles R. and Archibald, Anne M. and Ribeiro, Ant\^onio H. and Pedregosa, Fabian and family=Mulbregt, given=Paul, prefix=van, useprefix=true},
+  year = {2020},
+  journaltitle = {Nature Methods},
+  volume = {17},
+  number = {3},
+  pages = {261--272},
+  publisher = {Nature Publishing Group},
 }
