### 2026-01-16 b046934
Second draft of paper
