+++ docs/build/html/_sources/tutorials/tutorial_2.rst.txt
@@ -1,272 +1,0 @@
-.. _tutorial_2:
-
-*****************************************
-Synthesizing a population of binary stars
-*****************************************
-
-Dyad has a subpackage, :mod:`dyad.stats`, that contains probability
-distributions for the mass ratios and orbital elements of a population
-of binary stars. These probability distributions are implemented in
-the same way as Scipy's continous random variables (see
-:class:`scipy.stats.rv_continuous` and, for an example,
-:class:`scipy.stats.loguniform`). As a result, they come equipped with
-a large number of useful methods, in particular ``pdf``, which
-computes the probability density function (PDF), and ``rvs``, which
-generates random variates (i.e. which generates a sample).
-
-For example, consider the random variable for inclination, :math:`I`, which has PDF given by
-
-.. math::
-   
-   f_{I}(i) =
-   \begin{cases}
-   \frac{1}{2}\sin(i) &\text{ if $i \in [0, \pi)$}\\
-   0 &\text{ otherwise.}
-   \end{cases}
-
-This random variable is implemented by Dyad using the class :class:`dyad.stats.inclination`. Instantiate the class as follows.
-
-.. testsetup::
-
-   import numpy as np
-   np.random.seed(0)
-
-.. doctest:: python
-
-   >>> import dyad
-   >>> rv = dyad.stats.inclination
-
-Now compute the PDF on the interval :math:`[0, \pi)`.
-
-.. doctest:: python
-
-   >>> import numpy as np
-   >>> x = np.linspace(0., np.pi)
-   >>> pdf = rv.pdf(x)
-
-And generate a sample of size :math:`10\,000`.
-
-.. doctest:: python
-
-   >>> sample = rv.rvs(size=10_000)
-
-Now plot our results.
-
-.. doctest:: python
-
-   >>> import matplotlib.pyplot as plt
-   >>> fig, ax = plt.subplots()
-   >>> ax.hist(sample, bins="auto", density=True, histtype="stepfilled", alpha=0.2)
-   (array([0.04519806, 0.08320552, 0.12943081, 0.18798285, 0.24242597,
-	  0.22701754, 0.32768595, 0.36055726, 0.36774787, 0.39445581,
-	  0.42938159, 0.46430736, 0.43965387, 0.49615145, 0.49615145,
-	  0.53415891, 0.49306976, 0.49512422, 0.47868856, 0.43451773,
-	  0.44376279, 0.4406811 , 0.41089147, 0.34206715, 0.32357703,
-	  0.29995077, 0.27427006, 0.21160911, 0.18490116, 0.11607684,
-	  0.06882432, 0.0287624 ]), array([0.01702368, 0.11437298, 0.21172229, 0.3090716 , 0.40642091,
-	  0.50377021, 0.60111952, 0.69846883, 0.79581814, 0.89316744,
-	  0.99051675, 1.08786606, 1.18521536, 1.28256467, 1.37991398,
-	  1.47726329, 1.57461259, 1.6719619 , 1.76931121, 1.86666052,
-	  1.96400982, 2.06135913, 2.15870844, 2.25605775, 2.35340705,
-	  2.45075636, 2.54810567, 2.64545498, 2.74280428, 2.84015359,
-	  2.9375029 , 3.03485221, 3.13220151]), [<matplotlib.patches.Polygon object at 0x...>])
-   >>> ax.set_xticks([0., 0.5*np.pi, np.pi], [r"$0$", r"$\pi/2$", r"$\pi$"])
[…]
