+++ docs/build/html/_sources/api_reference/generated/dyad.Orbit.rst.txt
@@ -1,54 +1,0 @@
-﻿dyad.Orbit
-==========
-
-.. currentmodule:: dyad
-
-.. autoclass:: Orbit
-
-   
-   .. automethod:: __init__
-
-   
-   .. rubric:: Methods
-
-   .. autosummary::
-   
-      ~Orbit.__init__
-      ~Orbit.delaunay_elements
-      ~Orbit.eccentric_anomaly
-      ~Orbit.kinetic_energy
-      ~Orbit.mean_anomaly
-      ~Orbit.modified_delaunay_elements
-      ~Orbit.orbital_elements
-      ~Orbit.potential
-      ~Orbit.radius
-      ~Orbit.speed
-      ~Orbit.state
-   
-   
-
-   
-   
-   .. rubric:: Attributes
-
-   .. autosummary::
-   
-      ~Orbit.angular_momentum
-      ~Orbit.angular_momentum_magnitude
-      ~Orbit.apoapsis
-      ~Orbit.area
-      ~Orbit.argument_of_pericentre
-      ~Orbit.central_mass
-      ~Orbit.eccentricity
-      ~Orbit.eccentricity_vector
-      ~Orbit.energy
-      ~Orbit.inclination
-      ~Orbit.longitude_of_ascending_node
-      ~Orbit.lrl_vector
-      ~Orbit.periapsis
-      ~Orbit.period
-      ~Orbit.semilatus_rectum
-      ~Orbit.semimajor_axis
-      ~Orbit.semiminor_axis
-   
-   
