+++ docs/build/html/_sources/tutorials/tutorial_1.rst.txt
@@ -1,230 +1,0 @@
-.. _keplerian_system:
-
-*******************************
-Representing a Keplerian system
-*******************************
-
-To a good approximation the Earth--Sun system is Keplerian. The Earth moves on an elliptical orbit with semimajor axis :math:`a = 1~\text{AU}` and ellipticity :math:`e = 0.0167` in the gravitational central potential generated by the Sun, which has mass :math:`m = 1~\text{M}_{\odot}`. We may represent this system in Dyad using the class :class:`dyad.Orbit`.
-
-.. doctest:: python
-
-   >>> import dyad
-   >>> m, a, e = 1., 1., 0.0167
-   >>> orbit = dyad.Orbit(m, a, e)
-
-The state (i.e. position and velocity) of the body is specified by its true anomaly, :math:`\theta`, which we can compute using the method :meth:`dyad.Orbit.state`. Suppose that :math:`\theta = 1`.
-
-.. doctest:: python
-
-   >>> theta = 1.
-   >>> orbit.state(theta)
-   array([ 0.53532139,  0.83371367,  0.        , -0.01447709,  0.00958295,
-           0.        ])
-
-This is given in the format :math:`(x, y, z, v_{x}, v_{y}, v_{z})`
-where :math:`x`, :math:`y`, and :math:`z` are in units of
-:math:`\mathrm{AU}` and :math:`v_{x}`, :math:`v_{y}`, and
-:math:`v_{z}` are in units of :math:`\mathrm{AU}~\mathrm{d}^{-1}`. We
-can also compute the orbital radius and speed of the body, which are
-also in units of :math:`\mathrm{AU}` and
-:math:`\mathrm{AU}~\mathrm{d}^{-1}`.
-
-.. doctest:: python
-
-   >>> orbit.radius(theta)
-   0.9907812427855317
-   >>> orbit.speed(theta)
-   0.017361418667972937
-
-An instance of the class :class:`dyad.Orbit` also has a number of
-attributes that represent constants of the orbit. These include the arguments with which :class:`dyad.Orbit` was called.
-
-.. doctest:: python
-
-   >>> orbit.central_mass
-   1.0
-   >>> orbit.semimajor_axis
-   1.0
-   >>> orbit.eccentricity
-   0.0167
-
-They also include more interesting constants, such as the period (in :math:`\mathrm{d}`), periapsis (in :math:`\mathrm{AU}`) and apoapsis (in :math:`\mathrm{AU}`). A full list is available in the API documentation for :class:`dyad.Orbit`.
-
-.. doctest:: python
-
-   >>> orbit.period
-   365.25689844815537
-   >>> orbit.periapsis
-   0.9833
-   >>> orbit.apoapsis
-   1.0167
-
-If the perifocal frame differs from the observer's frame we can specify its orientation using the Euler angles :math:`\Omega`, :math:`i`, and :math:`\omega` (longitude of the ascending node, inclination, and argument of pericentre). Suppose that :math:`\Omega = 1`, :math:`i = 1`, and :math:`\omega = 1`.
-
-.. doctest:: python
-
-   >>> Omega, i, omega = 1., 1., 1.
-   >>> orbit = dyad.Orbit(m, a, e, Omega, i, omega)
-
-All methods and attributes are available as before.
-
-We may not always wish to work with true anomaly but instead prefer mean anomaly or eccentric anomaly.
-Dyad allows us to convert between these quantities.
-
-.. doctest:: python
-
-   >>> mu = dyad.mean_anomaly_from_true_anomaly(theta, e)
-   >>> dyad.true_anomaly_from_mean_anomaly(mu, e)
-   1.0
[…]
