+++ docs/build/html/_sources/api_reference/generated/dyad.stats.mass.rst.txt
@@ -1,14 +1,0 @@
-﻿dyad.stats.mass
-===============
-
-.. automodule:: dyad.stats.mass
-
-   
-   .. rubric:: Classes
-
-   .. autosummary::
-   
-      kroupa2001_gen
-      salpeter1955_gen
-      splitpowerlaw_gen
-   

[…]
