### 2026-01-16 c7c2f1e
Second draft of paper
