+++ paper.bib
@@ -5,7 +5,7 @@
   journal = {The Astrophysical Journal Supplement Series},
   volume = {230},
   number = {2},
-  pages = {15}
+  pages = {15},
   doi = {10.3847/1538-4365/aa6fb6},
 }
 
