### 2026-01-16 2424776
Second draft of paper
