+++ paper.bib
@@ -56,7 +56,7 @@
   year = {2023},
   month = sep,
   volume = {677},
-  pages = {A95}
+  pages = {A95},
   journal = {Astronomy \& Astrophysics},
   doi = {10.1051/0004-6361/202346843},
 }
@@ -88,13 +88,6 @@
   volume = {896},
   pages = {152},
   doi = {10.3847/1538-4357/ab910b},
-}
-
-@misc{ginsburg2021,
-  author = {Ginsburg, Adam},
-  year = {2021},
-  month = {feb},
-  url = {https://github.com/keflavich/imf}
 }
 
 @article{riley2022,
