### 2026-01-16 bf95e3c
Second draft of paper
