+++ paper.bib
@@ -62,11 +62,14 @@
 }
 
 @article{gration2025,
-  author = {Gration, A., Hendricks, D., Das, P., Heber, D., Izzard, R.},
+  title = {Stellar Velocity Distributions in Binary-Rich Ultrafaint Dwarf Galaxies},
+  author = {Gration, A. and Hendriks, D. D. and Das, P. and Heber, D. and Izzard, R. G.},
   year = {2025},
-  month = sep,
-  journal = {arXiv e-prints},
-  doi = {10.48550/arXiv.2509.14316},
+  journaltitle = {Monthly Notices of the Royal Astronomical Society},
+  volume = {543},
+  pages = {1120--1132},
+  publisher = {OUP},
+  doi = {10.1093/mnras/staf1481},
 }
 
 @article{hendriks2023,
@@ -132,4 +135,5 @@
   number = {3},
   pages = {261--272},
   publisher = {Nature Publishing Group},
+  doi = {10.1038/s41592-019-0686-2},
 }
