### 2026-01-16 11abb16
Second draft of paper
