+++ paper.bib
@@ -137,3 +137,11 @@
   publisher = {Nature Publishing Group},
   doi = {10.1038/s41592-019-0686-2},
 }
+
+@article{duquennoy1991,
+  author = {Duquennoy, A. and Mayor, M.},
+  year = {1991},
+  journaltitle = {Astronomy and Astrophysics},
+  volume = {248},
+  pages = {485},
+}
