+++ paper.bib
@@ -145,3 +145,11 @@
   volume = {248},
   pages = {485},
 }
+
+@article{izzard2023,
+  title = {Binary\_c: {{Stellar}} Population Synthesis Software Framework},
+  author = {Izzard, Robert},
+  year = {2023},
+  journaltitle = {Astrophysics Source Code Library},
+  pages = {ascl:2307.035},
+}
