### 2026-01-16 #9011 @AmeryGration
Hi @stegmaja. I have finally got round to making those changes.

> Regarding the Moe & di Stefano implementation (https://dyad-a-two-body-dynamics-and-binary-star-statistics-package.readthedocs.io/en/latest/tutorials/tutorial_3.html) I see the example:
>
> ```
> e = np.zeros(n_binary)
> idx = (log_p > 0.9375)
> e[idx] = stats.eccentricity.moe2017(log_p[idx], m_1[idx]).rvs()
> ```
>
> Wouldn't it be more sensible to have this log_p dependency included into `stats.eccentricity.moe2017`, so that one could simply write
>
> ```
> e = stats.eccentricity.moe2017(log_p, m_1).rvs()
> ```
>
> and the function takes care of the log_p threshold below which binaries are circularised?
>
> Otherwise, I think people will incorrectly sample MdS with your code, if they haven't super carefully read your documentation.

I agree with you that the implementation is cumbersome but this is a limitation of the `scipy.stats.rv_continuous` class rather than my implementation of the subclass `stats.eccentricity.moe2017`. As I have written it `stats.eccentricity.moe2017` does in fact check the validity of the shape parameters passed to it. But the `rvs` method of `rv_continuous` does not behave as you might expect it to. Passing an invalid shape parameter causes that method to fail rather than to return `nan`. Consider the following example, using shape parameters `log10_period=(0.1, 1.)` and `primary_mass=1.`. Here, the first value of log-period, 0.1, is smaller than the circularization log-period of 0.9375 while the second value, 1., is larger.
```
>>> from dyad.stats.eccentricity import moe2017
>>> moe2017(log10_period=[0.1, 1.], primary_mass=1.).rvs()
...
ValueError: Domain error in arguments. The `scale` parameter must be positive for all distributions, and many distributions have restrictions on shape parameters. Please see the `scipy.stats.eccentricity.moe2017` documentation for details.
```
This seems to be a quirk of `rv_continuous` because its method `pdf` does indeed work as you would expect.
```
>>> moe2017(log10_period=[0.1, 1.], primary_mass=1.).pdf(0.5)
array([       nan, 0.37862403])
```
So, regrettably, I think the implementation has to stay as it is. The problem for the user will not be incorrect sampling, though, because the sampler will just fail if an incorrect log-period is passed to it. In that case the error message does at least refer the user to the API documentation.

> On the same link, I think it could be slightly confusing to talk about stellar radii. Perhaps distance/separation is more appropriate.

Yes, my use of 'stellar radii' is very confusing. I've change it to 'orbital radii'.
