+++ dyad/stats/primary_mass/random.py
@@ -1,0 +1,252 @@
+__all__ = [
+    "kroupa2001",
+    "salpeter1955",
+    "splitpowerlaw",
+]
+
+import numpy as np
+import scipy as sp
+
+from dyad.stats import mass
+from .. import _distn_infrastructure
+
+
+class kroupa2001_gen(_distn_infrastructure.rv_continuous):
+    r"""The primary-star mass random variable for random pairing
+
+    %(before_notes)s
+
+    Notes
+    -----
+    xxx
+    
+    %(after_notes)s
+
+    See also
+    --------
+
+    References
+    ----------
+
+    %(example)s
+
+    """
+    def _argcheck(self, a, b):
+        res = (0. < a) & (a < 0.5) & (0.5 < b)
+
+        return res
+
+    def _get_support(self, a, b):
+        res = a, b
+
+        return res
+    
+    def _pdf(self, x, a, b):
+        res = 2.*mass.kroupa2001(a, b).pdf(x)*mass.kroupa2001(a, b).cdf(x)
+        
+        return res
+
+    def _cdf(self, x, a, b):
+        res = mass.kroupa2001(a, b).cdf(x)**2.
+                
+        return res
+
+    def _ppf(self, q, a, b):
+        res = mass.kroupa2001(a, b).ppf(np.sqrt(q))
+        
+        return res
+
+
+kroupa2001 = kroupa2001_gen("primary_mass.random.kroupa2001")
+
+
+class salpeter1955_gen(_distn_infrastructure.rv_continuous):
+    r"""The primary-star mass random variable for random pairing
+
+    %(before_notes)s
+
+    Notes
+    -----
+    xxx
+    
+    %(after_notes)s
+
+    See also
+    --------
+
+    References
+    ----------
+
+    %(example)s
+
+    """
+    def _argcheck(self, a, b):
+        res = (0. < a) & (a < b)
+
+        return res
+
+    def _get_support(self, a, b):
+        res = a, b
+
+        return res
+    
+    def _pdf(self, x, a, b):
+        res = 2.*mass.salpeter1955(a, b).pdf(x)*mass.salpeter1955(a, b).cdf(x)
+        
+        return res
+
+    def _cdf(self, x, a, b):
+        res = mass.salpeter1955(a, b).cdf(x)**2.
+                
+        return res
+
+    def _ppf(self, q, a, b):
+        res = mass.salpeter1955(a, b).ppf(np.sqrt(q))
+        
+        return res
+
+
+salpeter1955 = salpeter1955_gen("primary_mass.random.salpeter1955")
+
+
+class splitpowerlaw_gen(_distn_infrastructure.rv_continuous):
+    r"""The primary-star mass random variable for random pairing
+
+    %(before_notes)s
+
+    Notes
+    -----
+    xxx
+    
+    %(after_notes)s
+
+    See also
+    --------
+
+    References
+    ----------
+
+    %(example)s
+
+    """
+    def _argcheck(self, s, a, b, c, d):
+        res = (0. < a) & (a < b) & (a < s) & (s < b) & (c < 0.) & (d < 0.)
+
+        return res
+
+    def _get_support(self, s, a, b, c, d):
+        res = a, b
+
+        return res
+    
+    def _pdf(self, x, s, a, b, c, d):
+        res = (
+            2.
+            *mass.splitpowerlaw(s, a, b, c, d).pdf(x)
+            *mass.splitpowerlaw(s, a, b, c, d).cdf(x)
+        )
+        
+        return res
+
[…]
