+++ dyad/stats/mass_ratio.py
@@ -1,0 +1,1196 @@
+"""
+=========================================
+Mass ratio (:mod:`dyad.stats.mass_ratio`)
+=========================================
+
+.. currentmodule:: dyad.stats.mass_ratio
+
+This module contains probability distributions for the mass ratios of
+the components of a population of binary stars. In its documentation
+the random variable is denoted :math:`Q` and a realization of that
+random variable is denoted :math:`q`.
+
+Probability distributions
+=========================
+
+.. autosummary::
+   :toctree: generated/
+
+   duquennoy1991
+   moe2017
+   uniform
+
+"""
+
+__all__ = [
+    "duquennoy1991",
+    "moe2017",
+    "uniform"
+]
+
+import numpy as np
+import scipy as sp
+
+from scipy._lib._util import _lazyselect
+from scipy._lib._util import _lazywhere
+from . import _distn_infrastructure
+
+_truncnorm = sp.stats.truncnorm
+
+
+class duquennoy1991_gen(_distn_infrastructure.rv_continuous):
+    r"""The mass-ratio random variable of Duquennoy and Mayor (1991)
+
+    %(before_notes)s
+
+    Notes
+    -----
+    The probability density function for `duquennoy1991` is:
+
+    .. math::
+       f_{Q}(q) = \dfrac{1}{1 - \Phi(0; \mu, \sigma^{2})}\phi(q; \mu,
+       \sigma^{2})
+
+    for mass ratio :math:`q \in (0, \infty)` where :math:`\phi(\cdot;
+    \mu, \sigma^{2})` and :math:`\Phi(\cdot; \mu, \sigma^{2})` are the
+    probability density function and the cumulative distribution
+    function for a Gaussian random variable with mean :math:`\mu` and
+    variance :math:`\sigma^{2}` and where :math:`\mu = 0.23` and
+    :math:`\sigma = 0.42`.
+    
+    %(after_notes)s
+
+    References
+    ----------
+    Duquennoy, A., and M. Mayor. 1991. \'Multiplicity among
+    solar-type stars in the solar neighbourhood---II. Distribution of
+    the orbital elements in an unbiased Sample\'. *Astronomy and
+    Astrophysics* 248 (August): 485.
+
+    %(example)s
+
+    """
+    def _pdf(self, x):
+        res = _duquennoy1991.pdf(x)
+
+        return res
+
+    def _cdf(self, x):
+        res = _duquennoy1991.cdf(x)
+
+        return res
+
+    def _ppf(self, x):
+        res = _duquennoy1991.ppf(x)
+
+        return res
+
+
+_duquennoy1991_loc = 0.23
+_duquennoy1991_scale = 0.42
+_duquennoy1991_a = (0. - _duquennoy1991_loc)/_duquennoy1991_scale
+_duquennoy1991_b = (np.inf - _duquennoy1991_loc)/_duquennoy1991_scale
+_duquennoy1991 = _truncnorm(
+    a=_duquennoy1991_a, b=_duquennoy1991_b, loc=_duquennoy1991_loc,
+    scale=_duquennoy1991_scale
+)
+duquennoy1991 = duquennoy1991_gen(
+    a=0., b=np.inf, name="mass_ratio.duquennoy1991"
+)
+
+
+class moe2017_gen(_distn_infrastructure.rv_continuous):
+    r"""The mass-ratio random variable of Moe and Stefano (2017)
+
+    %(before_notes)s
+
+    Notes
+    -----
+    The probability density function for ``moe2017`` is:
+
+    .. math::
+       f_{Q|P, M_{1}}(q|p, m_{1})
+       =
+       A_{Q}(p, m_{1})
+       \begin{cases}
[…]
