### 2026-02-03 5c9d58c
Update docstrings
