+++ dyad/stats/mass_ratio/random.py
@@ -13,7 +13,7 @@ __all__ = [
 
 
 class kroupa2001_gen(_distn_infrastructure.rv_continuous):
-    r"""The random-pairing mass-ratio random variable (Kroupa mass function)
+    r"""A random-pairing mass-ratio random variable
 
     %(before_notes)s
 
@@ -23,22 +23,251 @@ class kroupa2001_gen(_distn_infrastructure.rv_continuous):
 
     .. math::
 
-       f_{Q|M_{1}}(q|m_{1}) = 
-       = m_{1}\dfrac{f_{M}(qm_{1})}{F_{M}(m_{1})}
+       f_{Q|M_{1}}(q|m_{1}) = m_{1}\dfrac{f_{M}(qm_{1})}{F_{M}(m_{1})}
 
     for mass ratio :math:`q \in (0, 1]` and primary_mass :math:`m_{1}
-    \in [a, b]` where :math:`f_{M}` and :math:`F_{M}` are the
-    probability density function and the cumulative distribution
-    function for the mass random variable of Kroupa
-    (`dyad.stats.mass.kroupa2001`), which iself has support on the
-    interval :math:`[a, b]`.
+    \in [m_{\min}, m_{\max}]` where :math:`f_{M}` and :math:`F_{M}`
+    are the probability density function and the cumulative
+    distribution function for the mass random variable of Kroupa
+    (2001). The probability density function :math:`f_{M}` is given by
 
+    .. math::
+       f_{M}(m)
+       =
+       A_{M}
+       \begin{cases}
+       m^{-1.3} &\text{ if $m \in [a, 0.5)$}\\
+       0.5m^{-2.3} &\text{ if $m \in [0.5, b]$}\\
+       \end{cases}
+
+    for
+
+    .. math::
+       A_{M} := \dfrac{0.5(0.5^{-1.3} - b^{-1.3})}{1.3} +
+       \dfrac{a^{-0.3} - 0.5^{-0.3}}{0.3}
+
+    and :math:`m \in [m_{\min}, m_{\max}]` where :math:`m_{\min} <
+    0.5` and :math:`m_{\max} > 0.5`. The cumulative distribution
+    function is given by
+
+    .. math::
+       F_{M}(m) = \int_{m_{\min}}^{m}f_{M}(u}\mathrm{d}\,u.
+  
     `random.kroupa2001` takes ``m_{1}`` as a shape parameter for
-    :math:`m_{1}`, the primary mass, ``a`` as a shape parameter for
-    :math:`a`, the minimum allowed mass, and ``b`` as a shape
-    parameter for :math:`b`, the maximum allowed mass.
+    :math:`m_{1}`, the primary mass, ``m_min`` as a shape parameter
+    for :math:`m_{\min}`, the minimum allowed mass, and ``m_max``
+    as a shape parameter for :math:`m_{\max}`, the maximum allowed mass.
 
     %(after_notes)s
+
+    See also
+    --------
+    `dyad.stats.mass.kroupa2001`
+
+    References
+    ----------
+    Malkov, O., and H. Zinnecker. 2001. \'Binary stars and the
+    fundamental initial mass function\'. *Monthly Notices of the Royal
+    Astronomical Society* 321 (1): 149–54.
+
+    Kroupa, P. 2002. \'The initial mass function and its variation
+    (review)\'. *ASP conference series* 285 (January): 86.
+
+    %(example)s
+
+    """
+    def _argcheck(self, m_1, m_min, m_max):
+        res = (0. < m_min) & (m_min < 0.5) & (0.5 < m_max)
+
+        return res
+
+    def _get_support(self, m_1, m_min, m_max):
+        res = m_min/m_1, 1.
+
[…]
