### 2026-02-04 7edaa24
Add random-pairing primary-mass random variables
