### 2026-02-04 aa67bc7
Remove superfluous data
