### 2026-02-04 84ee911
Remove superfluous data
