### 2026-02-04 #9011 @warrickball
Great! Thanks @stegmaja and @AmedeoRom for your reviews.

@AmeryGration, we'll now move on to the post-review steps, for which I'll generate a checklist in the moment for me to tick off. The main thing for you to do will be to

* create a versioned release and let me know the version number, and
* archive it (on [Zenodo](https://zenodo.org/), [figshare](https://figshare.com/), or another similar service) and let me know the DOI.

Please check that the metadata (title, authors, licence) of the Zenodo archive matches the JOSS paper. If the Zenodo archive is generated automatically (e.g. using the Zenodo-GitHub integration: [GitHub docs](https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content), [Zenodo docs](https://help.zenodo.org/docs/github/)), these can be mismatched for a number of reasons.

Changing the archive metadata usually doesn't require updating the files and we aren't worried about the paper, so you won't need a new archive after I've done my additional editorial checks.
