### 2026-02-04 #9011 @warrickball
I'll have a closer look at the paper with my track editor hat on later, but there are two things you should fix now.

* Please add titles to all the articles. We do render them in our citation style (unlike most astro journals).
* The "Kingdom" of "United Kingdom" has gone missing from your affiliation.
