### 2026-02-09 78874e6
Add titles to bibliography
