+++ paper.bib
@@ -1,5 +1,6 @@
 @article{moe2017,
   author = {Moe, M. and Di Stefano, R.},
+  title = {Mind your {{Ps}} and {{Qs}}: the interrelation between period ({{P}}) and mass-ratio ({{Q}}) distributions of binary Stars}
   year = {2017},
   month = jun,
   journal = {The Astrophysical Journal Supplement Series},
@@ -11,6 +12,7 @@
 
 @article{chabrier2003,
   author = {Chabrier, G.},
+  title = {Galactic stellar and substellar initial mass function},
   year = {2003},
   month = jul,
   journal = {Publications of the Astronomical Society of the Pacific},
@@ -22,6 +24,7 @@
 
 @article{kroupa2001,
   author = {Kroupa, P.},
+  title = {On the Variation of the initial mass function},
   year = {2001},
   month = apr,
   journal = {Monthly Notices of the Royal Astronomical Society},
@@ -33,6 +36,7 @@
 
 @article{salpeter1955,
   author = {Salpeter, E. E.},
+  title = {The luminosity function and stellar Evolution},
   year = {1955},
   month = jan,
   journal = {The Astrophysical Journal},
@@ -43,6 +47,7 @@
 
 @article{minor2010,
   author = {Minor, Quinn E. and Martinez, G. and Bullock, J. and Kaplinghat, M. and Trainor, R.},
+  title = {Correcting velocity dispersions of dwarf spheroidal galaxies for binary orbital motion},
   year = {2010},
   month = sep,
   journal = {The Astrophysical Journal},
@@ -53,6 +58,7 @@
 
 @article{arroyo-polonio2023,
   author = {{Arroyo-Polonio}, J. M. and Battaglia, G. and Thomas, G. F. and Irwin, M. J. and McConnachie, A. W. and Tolstoy, E.},
+  title = {Binary star population of the {{Sculptor}} dwarf galaxy},
   year = {2023},
   month = sep,
   volume = {677},
@@ -74,6 +80,7 @@
 
 @article{hendriks2023,
   author = {Hendriks, D. D., and Izzard, R. G.},
+  title = {binary\_c-Python: a {{Python-based}} stellar population synthesis tool and interface to binary\_c},
   journal = {Journal of Open Source Software},
   year = {2023},
   publisher = {The Open Journal},
@@ -85,6 +92,7 @@
 
 @article{rastello2020,
   author = {Rastello, Sara and Carraro, Giovanni and {Capuzzo-Dolcetta}, Roberto},
+  title = {Effect of binarity in star cluster dynamical mass determination},
   year = {2020},
   month = jun,
   journal = {The Astrophysical Journal},
@@ -95,6 +103,7 @@
 
 @article{riley2022,
   author = {Riley, Jeff and Agrawal, Poojan and Barrett, Jim W. and Boyett, Kristan N. K. and Broekgaarden, Floor S. and Chattopadhyay, Debatri and Gaebel, Sebastian M. and Gittins, Fabian and Hirai, Ryosuke and Howitt, George and Justham, Stephen and Khandelwal, Lokesh and Kummer, Floris and Lau, Mike Y. M. and Mandel, Ilya and {de Mink}, Selma E. and Neijssel, Coenraad and Riley, Tim and {van Son}, Lieke and Stevenson, Simon and {Vigna-G{\'o}mez}, Alejandro and Vinciguerra, Serena and Wagg, Tom and Willcox, Reinhold and {Team Compas}},
+  title = {Rapid stellar and binary population synthesis with {{COMPAS}}},
   year = {2022},
   month = feb,
   journal = {The Astrophysical Journal Supplement Series},
@@ -105,6 +114,7 @@
 
[…]
