+++ dyad.egg-info/SOURCES.txt
@@ -1,0 +1,31 @@
+COPYING
+MANIFEST.in
+README.md
+setup.py
+dyad/__init__.py
+dyad/_constants.py
+dyad/_core.py
+dyad.egg-info/PKG-INFO
+dyad.egg-info/SOURCES.txt
+dyad.egg-info/dependency_links.txt
+dyad.egg-info/requires.txt
+dyad.egg-info/top_level.txt
+dyad/stats/__init__.py
+dyad/stats/_distn_infrastructure.py
+dyad/stats/_distr_params.py
+dyad/stats/eccentricity.py
+dyad/stats/log_period.py
+dyad/stats/mass.py
+dyad/stats/mass_ratio.py
+dyad/stats/secondary_mass.py
+dyad/stats/semimajor_axis.py
+dyad/stats/data/log_period/cumulative_frequency_sample.dat
+dyad/stats/data/log_period/frequency_sample.dat
+dyad/stats/data/log_period/log10_period_sample.dat
+dyad/stats/data/log_period/mass_ratio_sample.dat
+dyad/stats/data/log_period/period_sample.dat
+dyad/stats/data/log_period/primary_mass_sample.dat
+dyad/stats/data/secondary_mass/cumulative_frequency_sample.dat
+dyad/stats/data/secondary_mass/frequency_sample.dat
+dyad/stats/data/secondary_mass/mass_ratio_sample.dat
+dyad/stats/data/secondary_mass/primary_mass_sample.dat
