### 2026-02-09 66adcda
Remove superfluous files and directories
