### 2026-02-09 eb36a70
Add titles to bibliography
