+++ paper.bib
@@ -1,6 +1,6 @@
 @article{moe2017,
   author = {Moe, M. and Di Stefano, R.},
-  title = {Mind your {{Ps}} and {{Qs}}: the interrelation between period ({{P}}) and mass-ratio ({{Q}}) distributions of binary Stars}
+  title = {Mind your {{Ps}} and {{Qs}}: the interrelation between period ({{P}}) and mass-ratio ({{Q}}) distributions of binary Stars},
   year = {2017},
   month = jun,
   journal = {The Astrophysical Journal Supplement Series},
