### 2026-02-09 616ca5c
Add titles to bibliography
