+++ paper.bib
@@ -1,6 +1,6 @@
 @article{moe2017,
   author = {Moe, M. and Di Stefano, R.},
-  title = {Mind your {{Ps}} and {{Qs}}: the interrelation between period ({{P}}) and mass-ratio ({{Q}}) distributions of binary Stars},
+  title = {Mind your {{Ps}} and {{Qs}}: {{the}} interrelation between period ({{P}}) and mass-ratio ({{Q}}) distributions of binary Stars},
   year = {2017},
   month = jun,
   journal = {The Astrophysical Journal Supplement Series},
@@ -80,7 +80,7 @@
 
 @article{hendriks2023,
   author = {Hendriks, D. D., and Izzard, R. G.},
-  title = {binary\_c-Python: a {{Python-based}} stellar population synthesis tool and interface to binary\_c},
+  title = {binary\_c-Python: {{a}} {{Python-based}} stellar population synthesis tool and interface to binary\_c},
   journal = {Journal of Open Source Software},
   year = {2023},
   publisher = {The Open Journal},
@@ -137,7 +137,7 @@
 }
 
 @article{virtanen2020,
-  title = {{SciPy} 1.0: fundamental algorithms for scientific computing in {{Python}}},
+  title = {{SciPy} 1.0: {{fundamental}} algorithms for scientific computing in {{Python}}},
   author = {Virtanen, Pauli and Gommers, Ralf and Oliphant, Travis E. and Haberland, Matt and Reddy, Tyler and Cournapeau, David and Burovski, Evgeni and Peterson, Pearu and Weckesser, Warren and Bright, Jonathan and family=Walt, given=St\'efan J., prefix=van der, useprefix=true and Brett, Matthew and Wilson, Joshua and Millman, K. Jarrod and Mayorov, Nikolay and Nelson, Andrew R. J. and Jones, Eric and Kern, Robert and Larson, Eric and Carey, C. J. and Polat, \.Ilhan and Feng, Yu and Moore, Eric W. and VanderPlas, Jake and Laxalde, Denis and Perktold, Josef and Cimrman, Robert and Henriksen, Ian and Quintero, E. A. and Harris, Charles R. and Archibald, Anne M. and Ribeiro, Ant\^onio H. and Pedregosa, Fabian and family=Mulbregt, given=Paul, prefix=van, useprefix=true},
   year = {2020},
   journaltitle = {Nature Methods},
@@ -150,7 +150,7 @@
 
 @article{duquennoy1991,
   author = {Duquennoy, A. and Mayor, M.},
-  title = {Multiplicity among solar type stars in the solar neighbourhood II. {{Distribution}} of the orbital elements in an unbiased sample},
+  title = {Multiplicity among solar-type stars in the solar neighbourhood II. {{Distribution}} of the orbital elements in an unbiased sample},
   year = {1991},
   journaltitle = {Astronomy and Astrophysics},
   volume = {248},
