### 2026-02-09 #9011 @AmeryGration
I believe I have completed all my assigned tasks. Thank you, @warrickball.

> ## Additional Author Tasks After Review is Complete
>
>     * Double check authors and affiliations (including ORCIDs)

All correct.

>     * Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.

Done. Release Dyad 0.1.0 tagged v0.1.0.

>     * Archive the release on Zenodo/figshare/etc and post the DOI here.

DOI: https://doi.org/10.5281/zenodo.18561678

>     * Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.

Done.

>     * Make sure that the license listed for the archive is the same as the software license.

Done
