+++ setup.py
@@ -19,7 +19,7 @@ long_description = (Path(__file__).parent/"README.md").read_text()
 
 setup(
     name="dyad",
-    version="0.0.0",
+    version="0.1.0",
     description=description, 
     long_description=long_description,
     long_description_content_type="text/markdown",
