### 2026-02-09 #9011 @warrickball
Thanks! The version listed in the sidebar of the Zenodo archive is currently v1. Could update that to v0.1.0 to match the files and release version on GitHub?
