### 2026-02-09 #9011 @warrickball
> I can't find a DOI for [this paper](https://ui.adsabs.harvard.edu/abs/1991A%26A...248..485D/abstract). It has a bibcode if that's any use: 1991A&A...248..485D.

No problem. We only require correct DOIs for objects that have them, and some don't, including A&A articles from before about 2000.
