### 2026-02-10 #9011 @AmeryGration
Thank you for spotting that. Sidebar version number now correct.
