### 2026-02-10 #9011 @AmeryGration
Is it possible to add a funding acknowledgement to my paper? How should I do that?
