### 2026-02-10 #9011 @warrickball
> Is it possible to add a funding acknowledgement to my paper? How should I do that?

In short, you can go ahead and modify the source now and it'll be picked up as we iterate on the text. I'll have a close look at the text in the next few working days and probably propose changes via pull request, so we'll certainly need at least one more iteration on the paper source anyway.
