### 2026-02-11 8c87ca8
Add funding acknowledgement
