### 2026-02-12 #9011 @warrickball
Congratulations @AmeryGration, your paper has been published in JOSS! Thanks once again to @stegmaja and @AmedeoRom for reviewing this submission! JOSS simply wouldn't be possible without its community of volunteers.
