### 2026-02-12 #9011 @AmeryGration
That's great! Thank you so much @warrickball. And thank you @stegmaja and @AmedeoRom for all your work in reviewing Dyad. I really do appreciate your help. The package, documentation, and paper are all the better for it.
