### 2026-02-12 #9011 @AmeryGration
@warrickball, one more thing! Do you have any advice on posting to arXiv, which requires TeX?
