### 2026-02-12 #9011 @AmeryGration
@warrickball, one more thing! I see that I can use the editorialbot to generate a LaTeX document but that document uses the standard article class. Is there a way to JOSS-ify it?
