### 2026-02-14 #9011 @warrickball
> Is there a way to JOSS-ify it?

I asked around the editorial board but, as I suspected, the answer is basically "no". I believe the main issue is that our pipeline uses LuaLaTeX instead of PDFLaTeX. I had a go at seeing if any of the other pipeline outputs are closer or if I could comment out enough LuaLaTeX-specific source to get something but I don't know the pipeline well and despite a bit of tinkering I couldn't see a clear way forward. Sorry!
