### 2026-02-15 #9011 @AmeryGration
Thank you so much for that. No problem! I've also looked on the arXiv itself to see what the convention is. It seems that most people just submit the preprint generated by the editorialbot and give a link to the JOSS paper. So that's what I've done. Thank you again for all your help during the submission and review.
